Function tokio::executor::current_thread::block_on_all

pub fn block_on_all<F>(future: F) -> Result<F::Item, F::Error> where
    F: Future, 

Run the executor bootstrapping the execution with the provided future.

This creates a new CurrentThread executor, spawns the provided future, and blocks the current thread until the provided future and all subsequently spawned futures complete. In other words:

If the provided bootstrap future does not spawn any additional tasks, block_on_all returns once future completes.

If the provided bootstrap future does spawn additional tasks, then block_on_all returns once all spawned futures complete.

See module level documentation for more details.