Struct tokio::executor::current_thread::Entered

pub struct Entered<'a, P: Park + 'a> { /* fields omitted */ }

A CurrentThread instance bound to a supplied execution context.

Methods

impl<'a, P: Park> Entered<'a, P>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn spawn<F>(&mut self, future: F) -> &mut Self where
F: Future<Item = (), Error = ()> + 'static,

Spawn the future on the executor.

This internally queues the future to be executed once run is called.

pub fn block_on<F>(
&mut self,
future: F
) -> Result<F::Item, BlockError<F::Error>> where
F: Future,

Synchronously waits for the provided future to complete.

This function can be used to synchronously block the current thread until the provided future has resolved either successfully or with an error. The result of the future is then returned from this function call.

Note that this function will also execute any spawned futures on the current thread, but will not block until these other spawned futures have completed.

The caller is responsible for ensuring that other spawned futures complete execution.

pub fn run(&mut self) -> Result<(), RunError>

Run the executor to completion, blocking the thread until all spawned futures have completed.

pub fn run_timeout(&mut self, duration: Duration) -> Result<(), RunTimeoutError>

Run the executor to completion, blocking the thread until all spawned futures have completed or duration time has elapsed.

pub fn turn(&mut self, duration: Option<Duration>) -> Result<Turn, TurnError>

Perform a single iteration of the event loop.

This function blocks the current thread even if the executor is idle.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_park(&self) -> &P

Returns a reference to the underlying Park instance.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_park_mut(&mut self) -> &mut P

Returns a mutable reference to the underlying Park instance.