Struct tokio::executor::current_thread::TaskExecutor

pub struct TaskExecutor { /* fields omitted */ }

Executes futures on the current thread.

All futures executed using this executor will be executed on the current thread. As such, run will wait for these futures to complete before returning.

For more details, see the module level documentation.

Methods

impl TaskExecutor

pub fn current() -> TaskExecutor

Returns an executor that executes futures on the current thread.

The user of TaskExecutor must ensure that when a future is submitted, that it is done within the context of a call to run.

For more details, see the module level documentation.

pub fn spawn_local(
&mut self,
future: Box<Future<Item = (), Error = ()>>
) -> Result<(), SpawnError>

Spawn a future onto the current CurrentThread instance.