tokio

模块 fs

异步文件实用工具。

此模块包含用于异步 处理文件系统的实用方法。这包括对文件的读/写以及 对目录的操作。

请注意，大多数操作系统不提供异步文件系统 API。因此，Tokio 在后台将使用普通的阻塞文件操作。 这是通过使用 spawn_blocking 线程池在 后台运行它们来实现的。

tokio::fs 模块只应用于普通文件。尝试将其用于 Linux 上的命名管道等特殊文件可能会导致意外行为， 例如在运行时关闭期间挂起。对于特殊文件， 应使用专门的类型，例如 tokio::net::unix::pipe 或 AsyncFd。

目前，Tokio 始终在所有平台上使用 spawn_blocking， 但未来可能会更改以使用异步文件系统 API（如 io_uring）。

Usage

使用此模块的最简单方法是使用对整个文件进行操作的实用函数：

tokio::fs::read

tokio::fs::read_to_string

tokio::fs::write

两个 read 函数读取整个文件并返回其内容。 write 函数接受文件内容并将这些 内容写入文件。如果存在现有文件，则会覆盖它。

例如，要读取文件：

let contents = tokio::fs::read_to_string("my_file.txt").await?;

println!("File has {} lines.", contents.lines().count());

要覆盖文件：

let contents = "First line.\nSecond line.\nThird line.\n";

tokio::fs::write("my_file.txt", contents.as_bytes()).await?;

Using File

与文件交互的主要类型是 File。它可用于对 给定的文件进行读取和写入。这是通过 AsyncRead 和 AsyncWrite trait 完成的。 当你希望执行比一次性读取或写入整个内容更复杂的操作时， 通常会使用此类型。

注意：在写入 Tokio File 时使用 flush 很重要。 这是因为对 write 的调用将在写入完成之前返回， 而 flush 将等待写入完成。 （即使你不调用 flush，写入也会发生；只是稍后发生。） 这与 std::fs::File 不同，原因是 File 在 后台使用了 spawn_blocking。

例如，要计算文件中的行数而不将整个文件加载到内存中：

use tokio::fs::File;
use tokio::io::AsyncReadExt;

let mut file = File::open("my_file.txt").await?;

let mut chunk = vec![0; 4096];
let mut number_of_lines = 0;
loop {
    let len = file.read(&mut chunk).await?;
    if len == 0 {
        // Length of zero means end of file.
        break;
    }
    for &b in &chunk[..len] {
        if b == b'\n' {
            number_of_lines += 1;
        }
    }
}

println!("File has {} lines.", number_of_lines);

例如，要逐行写入文件：

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::create("my_file.txt").await?;

file.write_all(b"First line.\n").await?;
file.write_all(b"Second line.\n").await?;
file.write_all(b"Third line.\n").await?;

// Remember to call `flush` after writing!
file.flush().await?;

Tuning your file IO

Tokio 的文件操作在后台使用 spawn_blocking， 这会带来严重的性能影响。为了在 Tokio 上获得良好的文件 IO 性能， 建议将你的操作批处理为尽可能少的 spawn_blocking 调用。

通过比较上面的两个读取示例，可以看到这种差异的一个例子。 第一个示例使用 tokio::fs::read， 它在单个 spawn_blocking 调用中读取 整个文件，然后返回它。第二个示例将使用许多 spawn_blocking 调用以块的形式读取文件。这意味着对于大文件， 第二个示例很可能更昂贵。 （当然，对于无法放入内存的超大文件，使用分块读取可能是必要的。）

以下示例将展示一些相关策略：

创建文件时，将数据写入 String 或 Vec<u8>， 然后使用 tokio::fs::write 在 单个 spawn_blocking 调用中写入整个文件。

let mut contents = String::new();

contents.push_str("First line.\n");
contents.push_str("Second line.\n");
contents.push_str("Third line.\n");

tokio::fs::write("my_file.txt", contents.as_bytes()).await?;

使用 BufReader 和 BufWriter 将许多小的读/写操作 缓冲为少数大的读/写操作。 此示例很可能只会执行一次 spawn_blocking 调用。

use tokio::fs::File;
use tokio::io::{AsyncWriteExt, BufWriter};

let mut file = BufWriter::new(File::create("my_file.txt").await?);

file.write_all(b"First line.\n").await?;
file.write_all(b"Second line.\n").await?;
file.write_all(b"Third line.\n").await?;

// Due to the BufWriter, the actual write and spawn_blocking
// call happens when you flush.
file.flush().await?;

在 spawn_blocking 中手动使用 std::fs。

use std::fs::File;
use std::io::{self, Write};
use tokio::task::spawn_blocking;

spawn_blocking(move || {
    let mut file = File::create("my_file.txt")?;

    file.write_all(b"First line.\n")?;
    file.write_all(b"Second line.\n")?;
    file.write_all(b"Third line.\n")?;

    // Unlike Tokio's file, the std::fs file does
    // not need flush.

    io::Result::Ok(())
}).await.unwrap()?;

还应注意 File::set_max_buf_size， 它控制 Tokio 的 File 在单个 spawn_blocking 调用中 将读取或写入的最大字节数。默认为 2 MB，但可能会更改。

结构体

DirBuilder

A builder for creating directories in various manners.

DirEntry

Entries returned by the ReadDir stream.

File

A reference to an open file on the filesystem.

OpenOptions

用于配置文件打开方式的可选项与标志。

ReadDir

Reads the entries in a directory.

函数

canonicalize

返回路径的规范化绝对形式，所有中间组件均被规范化，且符号链接被解析。

copy

Copies the contents of one file to another. This function will also copy the permission bits of the original file to the destination file. This function will overwrite the contents of to.

create_dir

在指定路径创建一个新的空目录。

create_dir_all

Recursively creates a directory and all of its parent components if they are missing.

hard_link

在文件系统上创建一个新的硬链接。

metadata

Given a path, queries the file system to get information about a file, directory, etc.

read

Reads the entire contents of a file into a bytes vector.

read_dir

Returns a stream over the entries within a directory.

read_link

读取符号链接，返回链接指向的文件。

read_to_string

Creates a future which will open a file for reading and read the entire contents into a string and return said string.

remove_dir

Removes an existing, empty directory.

remove_dir_all

删除该路径的目录，包括其全部内容。 Use carefully!

remove_file

从文件系统删除一个文件。

rename

Renames a file or directory to a new name, replacing the original file if to already exists.

set_permissions

Changes the permissions found on a file or a directory.

symlink_dir

Creates a new directory symlink on the filesystem.

symlink_file

Creates a new file symbolic link on the filesystem.

symlink_metadata

Queries the file system metadata for a path.

try_exists

Returns Ok(true) if the path points at an existing entity.

write

Creates a future that will open a file for writing and write the entire contents of contents to it.