Struct tokio::fs::DirBuilder

pub struct DirBuilder { /* fields omitted */ }

A builder for creating directories in various manners.

Additional Unix-specific options are available via importing the DirBuilderExt trait.

This is a specialized version of [std::fs::DirBuilder] for usage on the Tokio runtime.

Implementations

impl DirBuilder

pub fn new() -> Self

Creates a new set of options with default mode/security settings for all platforms and also non-recursive.

This is an async version of std::fs::DirBuilder::new

Examples

use tokio::fs::DirBuilder;

let builder = DirBuilder::new();

pub fn recursive(&mut self, recursive: bool) -> &mut Self

Indicates whether to create directories recursively (including all parent directories). Parents that do not exist are created with the same security and permissions settings.

This option defaults to false.

This is an async version of std::fs::DirBuilder::recursive

Examples

use tokio::fs::DirBuilder;

let mut builder = DirBuilder::new();
builder.recursive(true);

pub async fn create<'_, P: AsRef<Path>>(&'_ self, path: P) -> Result<()>

Creates the specified directory with the configured options.

It is considered an error if the directory already exists unless recursive mode is enabled.

This is an async version of std::fs::DirBuilder::create

Errors

An error will be returned under the following circumstances:

Path already points to an existing file.

Path already points to an existing directory and the mode is non-recursive.

The calling process doesn't have permissions to create the directory or its missing parents.

Other I/O error occurred.

Examples

use tokio::fs::DirBuilder;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    DirBuilder::new()
        .recursive(true)
        .create("/tmp/foo/bar/baz")
        .await?;

    Ok(())
}