Module tokio::future

Asynchronous values.

Traits

Future
A future represents an asynchronous computation.

FutureExt
An extension trait for Future that provides a variety of convenient combinator functions.

Functions

err
Create a future that is immediately ready with an error value.

ok
Create a future that is immediately ready with a success value.

pending
Creates a future which never resolves, representing a computation that never finishes.

poll_fn
Creates a new future wrapping around a function returning [Poll].

ready
Create a future that is immediately ready with a value.