Struct tokio::io::driver::scheduled_io::ScheduledIo

pub(crate) struct ScheduledIo {
    readiness: AtomicUsize,
    waiters: Mutex<Waiters>,
}

Stored in the I/O driver resource slab.

Fields
readiness: AtomicUsize

Packs the resource’s readiness with the resource’s generation.

waiters: Mutex<Waiters>
Implementations

impl ScheduledIo

pub(crate) fn generation(&self) -> usize

pub(super) fn shutdown(&self)

Invoked when the IO driver is shut down; forces this ScheduledIo into a permanently ready state.

pub(super) fn set_readiness(
&self,
token: Option<usize>,
tick: Tick,
f: impl Fn(Ready) -> Ready
) -> Result<(), ()>

Sets the readiness on this ScheduledIo by invoking the given closure on the current value, returning the previous readiness value.

Arguments

token: the token for this ScheduledIo.

tick: whether setting the tick or trying to clear readiness for a specific tick.

f: a closure returning a new readiness value given the previous readiness.

Returns

If the given token’s generation no longer matches the ScheduledIo’s generation, then the corresponding IO resource has been removed and replaced with a new resource. In that case, this method returns Err. Otherwise, this returns the previous readiness.

pub(super) fn wake(&self, ready: Ready)

Notifies all pending waiters that have registered interest in ready.

There may be many waiters to notify. Waking the pending task must be done from outside of the lock otherwise there is a potential for a deadlock.

A stack array of wakers is created and filled with wakers to notify, the lock is released, and the wakers are notified. Because there may be more than 32 wakers to notify, if the stack array fills up, the lock is released, the array is cleared, and the iteration continues.

fn wake0(&self, ready: Ready, shutdown: bool)

pub(super) fn ready_event(&self, interest: Interest) -> ReadyEvent

pub(super) fn poll_readiness(
&self,
cx: &mut Context<'_>,
direction: Direction
) -> Poll<ReadyEvent>

Polls for readiness events in a given direction.

These are to support AsyncRead and AsyncWrite polling methods, which cannot use the async fn version. This uses reserved reader and writer slots.

pub(crate) fn clear_readiness(&self, event: ReadyEvent)

pub(crate) fn clear_wakers(&self)

impl ScheduledIo

pub(crate) async fn readiness(&self, interest: Interest) -> ReadyEvent

An async version of poll_readiness which uses a linked list of wakers.

fn readiness_fut(&self, interest: Interest) -> Readiness<'_>