Struct tokio::io::driver::Handle

pub(crate) struct Handle {
    inner: Weak<Inner>,
}

A reference to an I/O driver.

Fields
inner: Weak<Inner>
Implementations

impl Handle

pub(super) fn current() -> Self

Returns a handle to the current reactor.

Panics

This function panics if there is no current reactor set and rt feature flag is not enabled.

impl Handle

fn wakeup(&self)

Forces a reactor blocked in a call to turn to wakeup, or otherwise makes the next call to turn return immediately.

This method is intended to be used in situations where a notification needs to otherwise be sent to the main reactor. If the reactor is currently blocked inside of turn then it will wake up and soon return after this method has been called. If the reactor is not currently blocked in turn, then the next call to turn will not block and return immediately.

pub(super) fn inner(&self) -> Option<Arc<Inner>>