Function tokio::io::duplex

pub fn duplex(max_buf_size: usize) -> (DuplexStream, DuplexStream)

Create a new pair of DuplexStreams that act like a pair of connected sockets.

The max_buf_size argument is the maximum amount of bytes that can be written to a side before the write returns Poll::Pending.