Function tokio::io::repeat

pub fn repeat(byte: u8) -> Repeat

Creates an instance of an async reader that infinitely repeats one byte.

All reads from this reader will succeed by filling the specified buffer with the given byte.

This is an asynchronous version of std::io::repeat.

Examples

use tokio::io::{self, AsyncReadExt};

#[tokio::main]
async fn main() {
    let mut buffer = [0; 3];
    io::repeat(0b101).read_exact(&mut buffer).await.unwrap();
    assert_eq!(buffer, [0b101, 0b101, 0b101]);
}