Function tokio::io::stderr

pub fn stderr() -> Stderr

Constructs a new handle to the standard error of the current process.

The returned handle allows writing to standard error from the within the Tokio runtime.