Struct tokio::io::Join

pub struct Join<R, W> { /* private fields */ }

Joins two values implementing AsyncRead and AsyncWrite into a single handle.

Implementations

impl<R, W> Join<R, W>
where R: AsyncRead, W: AsyncWrite,

pub fn into_inner(self) -> (R, W)

Splits this Join back into its AsyncRead and AsyncWrite components.

pub fn reader(&self) -> &R

Returns a reference to the inner reader.

pub fn writer(&self) -> &W

Returns a reference to the inner writer.

pub fn reader_mut(&mut self) -> &mut R

Returns a mutable reference to the inner reader.

pub fn writer_mut(&mut self) -> &mut W

Returns a mutable reference to the inner writer.

pub fn reader_pin_mut(self: Pin<&mut Self>) -> Pin<&mut R>

Returns a pinned mutable reference to the inner reader.

pub fn writer_pin_mut(self: Pin<&mut Self>) -> Pin<&mut W>

Returns a pinned mutable reference to the inner writer.