tokio::io
Struct Lines

pub struct Lines<R> { /* private fields */ }

Reads lines from an AsyncBufRead.

A Lines can be turned into a Stream with LinesStream.

This type is usually created using the lines method.

Implementations

impl<R> Lines<R>
where R: AsyncBufRead + Unpin,

pub async fn next_line(&mut self) -> Result<Option<String>>

Returns the next line in the stream.

Cancel safety

This method is cancellation safe.

Examples

use tokio::io::AsyncBufReadExt;

let mut lines = my_buf_read.lines();

while let Some(line) = lines.next_line().await? {
    println!("length = {}", line.len())
}

pub fn get_mut(&mut self) -> &mut R

Obtains a mutable reference to the underlying reader.

pub fn get_ref(&mut self) -> &R

Obtains a reference to the underlying reader.

pub fn into_inner(self) -> R

Unwraps this Lines<R>, returning the underlying reader.

Note that any leftover data in the internal buffer is lost. Therefore, a following read from the underlying reader may lead to data loss.

impl<R> Lines<R>
where R: AsyncBufRead,

pub fn poll_next_line( self: Pin<&mut Self>, cx: &mut Context<'_>, ) -> Poll<Result<Option<String>>>

Polls for the next line in the stream.

This method returns:

Poll::Pending if the next line is not yet available.

Poll::Ready(Ok(Some(line))) if the next line is available.

Poll::Ready(Ok(None)) if there are no more lines in this stream.

Poll::Ready(Err(err)) if an IO error occurred while reading the next line.

When the method returns Poll::Pending, the Waker in the provided Context is scheduled to receive a wakeup when more bytes become available on the underlying IO resource. Note that on multiple calls to poll_next_line, only the Waker from the Context passed to the most recent call is scheduled to receive a wakeup.