Struct tokio::io::PollEvented

pub struct PollEvented<E: Evented> { /* fields omitted */ }

Associates an I/O resource that implements the std::io::Read and/or std::io::Write traits with the reactor that drives it.

PollEvented uses Registration internally to take a type that implements mio::Evented as well as std::io::Read and or std::io::Write and associate it with a reactor that will drive it.

Once the mio::Evented type is wrapped by PollEvented, it can be used from within the future’s execution model. As such, the PollEvented type provides AsyncRead and AsyncWrite implementations using the underlying I/O resource as well as readiness events provided by the reactor.

Note: While PollEvented is Sync (if the underlying I/O type is Sync), the caller must ensure that there are at most two tasks that use a PollEvented instance concurrently. One for reading and one for writing. While violating this requirement is “safe” from a Rust memory model point of view, it will result in unexpected behavior in the form of lost notifications and tasks hanging.

Readiness events

Besides just providing AsyncRead and AsyncWrite implementations, this type also supports access to the underlying readiness event stream. While similar in function to what Registration provides, the semantics are a bit different.

Two functions are provided to access the readiness events: poll_read_ready and poll_write_ready. These functions return the current readiness state of the PollEvented instance. If poll_read_ready indicates read readiness, immediately calling poll_read_ready again will also indicate read readiness.

When the operation is attempted and is unable to succeed due to the I/O resource not being ready, the caller must call clear_read_ready or clear_write_ready. This clears the readiness state until a new readiness event is received.

This allows the caller to implement additional functions. For example, TcpListener implements poll_accept by using poll_read_ready and clear_read_ready.

use tokio::io::PollEvented;

use futures::ready;
use mio::Ready;
use mio::net::{TcpStream, TcpListener};
use std::io;
use std::task::{Context, Poll};

struct MyListener {
    poll_evented: PollEvented<TcpListener>,
}

impl MyListener {
    pub fn poll_accept(&mut self, cx: &mut Context<'_>) -> Poll<Result<TcpStream, io::Error>> {
        let ready = Ready::readable();

        ready!(self.poll_evented.poll_read_ready(cx, ready))?;

        match self.poll_evented.get_ref().accept() {
            Ok((socket, _)) => Poll::Ready(Ok(socket)),
            Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {
                self.poll_evented.clear_read_ready(cx, ready)?;
                Poll::Pending
            }
            Err(e) => Poll::Ready(Err(e)),
        }
    }
}

Platform-specific events

PollEvented also allows receiving platform-specific mio::Ready events. These events are included as part of the read readiness event stream. The write readiness event stream is only for Ready::writable() events.

Implementations

impl<E> PollEvented<E> where
E: Evented,

pub fn new(io: E) -> Result<Self>

Creates a new PollEvented associated with the default reactor.

Panics

This function panics if thread-local runtime is not set.

The runtime is usually set implicitly when this function is called from a future driven by a tokio runtime, otherwise runtime can be set explicitly with Handle::enter function.

pub fn new_with_ready(io: E, ready: Ready) -> Result<Self>

Creates a new PollEvented associated with the default reactor, for specific mio::Ready state. new_with_ready should be used over new when you need control over the readiness state, such as when a file descriptor only allows reads. This does not add hup or error so if you are interested in those states, you will need to add them to the readiness state passed to this function.

An example to listen to read only

#[cfg(unix)]
    mio::Ready::from_usize(
        mio::Ready::readable().as_usize()
        | mio::unix::UnixReady::error().as_usize()
        | mio::unix::UnixReady::hup().as_usize()
    );

Panics

This function panics if thread-local runtime is not set.

The runtime is usually set implicitly when this function is called from a future driven by a tokio runtime, otherwise runtime can be set explicitly with Handle::enter function.

pub fn get_ref(&self) -> &E

Returns a shared reference to the underlying I/O object this readiness stream is wrapping.

pub fn get_mut(&mut self) -> &mut E

Returns a mutable reference to the underlying I/O object this readiness stream is wrapping.

pub fn into_inner(self) -> Result<E>

Consumes self, returning the inner I/O object

This function will deregister the I/O resource from the reactor before returning. If the deregistration operation fails, an error is returned.

Note that deregistering does not guarantee that the I/O resource can be registered with a different reactor. Some I/O resource types can only be associated with a single reactor instance for their lifetime.

pub fn poll_read_ready(
&self,
cx: &mut Context<'_>,
mask: Ready
) -> Poll<Result<Ready>>

Checks the I/O resource’s read readiness state.

The mask argument allows specifying what readiness to notify on. This can be any value, including platform specific readiness, except writable. HUP is always implicitly included on platforms that support it.

If the resource is not ready for a read then Poll::Pending is returned and the current task is notified once a new event is received.

The I/O resource will remain in a read-ready state until readiness is cleared by calling clear_read_ready.

Panics

This function panics if:

ready includes writable.

called from outside of a task context.

Warning

This method may not be called concurrently. It takes &self to allow calling it concurrently with poll_write_ready.

pub fn clear_read_ready(&self, cx: &mut Context<'_>, ready: Ready) -> Result<()>

Clears the I/O resource’s read readiness state and registers the current task to be notified once a read readiness event is received.

After calling this function, poll_read_ready will return Poll::Pending until a new read readiness event has been received.

The mask argument specifies the readiness bits to clear. This may not include writable or hup.

Panics

This function panics if:

ready includes writable or HUP

called from outside of a task context.

pub fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<Result<Ready>>

Checks the I/O resource’s write readiness state.

This always checks for writable readiness and also checks for HUP readiness on platforms that support it.

If the resource is not ready for a write then Poll::Pending is returned and the current task is notified once a new event is received.

The I/O resource will remain in a write-ready state until readiness is cleared by calling clear_write_ready.

Panics

This function panics if:

ready contains bits besides writable and hup.

called from outside of a task context.

Warning

This method may not be called concurrently. It takes &self to allow calling it concurrently with poll_read_ready.

pub fn clear_write_ready(&self, cx: &mut Context<'_>) -> Result<()>

Resets the I/O resource’s write readiness state and registers the current task to be notified once a write readiness event is received.

This only clears writable readiness. HUP (on platforms that support HUP) cannot be cleared as it is a final state.

After calling this function, poll_write_ready(Ready::writable()) will return NotReady until a new write readiness event has been received.

Panics

This function will panic if called from outside of a task context.