Struct tokio::io::Stdin

pub struct Stdin { /* private fields */ }

A handle to the standard input stream of a process.

The handle implements the AsyncRead trait, but beware that concurrent reads of Stdin must be executed with care.

This handle is best used for non-interactive uses, such as when a file is piped into the application. For technical reasons, stdin is implemented by using an ordinary blocking read on a separate thread, and it is impossible to cancel that read. This can make shutdown of the runtime hang until the user presses enter.

For interactive uses, it is recommended to spawn a thread dedicated to user input and use blocking IO directly in that thread.

Created by the stdin function.