tokio::io

特性 AsyncBufReadExt

pub trait AsyncBufReadExt: AsyncBufRead {
    // Provided methods
    fn read_until<'a>(
        &'a mut self,
        byte: u8,
        buf: &'a mut Vec<u8>,
    ) -> ReadUntil<'a, Self>
       where Self: Unpin { ... }
    fn read_line<'a>(&'a mut self, buf: &'a mut String) -> ReadLine<'a, Self>
       where Self: Unpin { ... }
    fn split(self, byte: u8) -> Split<Self>
       where Self: Sized + Unpin { ... }
    fn fill_buf(&mut self) -> FillBuf<'_, Self>
       where Self: Unpin { ... }
    fn consume(&mut self, amt: usize)
       where Self: Unpin { ... }
    fn lines(self) -> Lines<Self>
       where Self: Sized { ... }
}

一个扩展 trait，为 AsyncBufRead 类型添加实用方法。

提供方法

fn read_until<'a>( &'a mut self, byte: u8, buf: &'a mut Vec<u8>, ) -> ReadUntil<'a, Self>
where Self: Unpin,

读取所有字节到 buf，直到遇到分隔字节 byte 或 EOF。

Equivalent to:

async fn read_until(&mut self, byte: u8, buf: &mut Vec<u8>) -> io::Result<usize>;

此函数将从底层流读取字节，直到找到分隔字节或 EOF。找到后，到此为止的所有字节（包括分隔字节本身，若找到的话）将被追加到 buf。

若成功，此函数将返回读取的总字节数。

如果此函数返回 Ok(0)，则流已达到 EOF。

错误

此函数会忽略所有 ErrorKind::Interrupted 错误， 并返回 fill_buf 所返回的其他任何错误。

如果遇到 I/O 错误，则到目前为止已读取的所有字节都将保留在 buf 中， 并且其长度会相应调整。

Cancel safety

如果该方法在 tokio::select! 语句中作为事件使用，且另一个分支先完成，则部分数据可能已被读取。 已部分读取的字节会附加到 buf，可以再次调用该方法以继续读取， 直到遇到 byte。

此方法返回已读取的总字节数。如果取消对 read_until 的调用， 然后再次调用以继续读取，则计数器会重置。

示例

std::io::Cursor 是实现 BufRead 的类型。 在此示例中，我们使用 Cursor 按连字符分隔的方式读取字节切片中的所有字节：

use tokio::io::AsyncBufReadExt;

use std::io::Cursor;

let mut cursor = Cursor::new(b"lorem-ipsum");
let mut buf = vec![];

// cursor is at 'l'
let num_bytes = cursor.read_until(b'-', &mut buf)
    .await
    .expect("reading from cursor won't fail");

assert_eq!(num_bytes, 6);
assert_eq!(buf, b"lorem-");
buf.clear();

// cursor is at 'i'
let num_bytes = cursor.read_until(b'-', &mut buf)
    .await
    .expect("reading from cursor won't fail");

assert_eq!(num_bytes, 5);
assert_eq!(buf, b"ipsum");
buf.clear();

// cursor is at EOF
let num_bytes = cursor.read_until(b'-', &mut buf)
    .await
    .expect("reading from cursor won't fail");
assert_eq!(num_bytes, 0);
assert_eq!(buf, b"");

fn read_line<'a>(&'a mut self, buf: &'a mut String) -> ReadLine<'a, Self>
where Self: Unpin,

读取所有字节直到遇到换行符（0xA 字节），并将它们追加到提供的缓冲区中。

Equivalent to:

async fn read_line(&mut self, buf: &mut String) -> io::Result<usize>;

此函数会从底层流读取字节，直到找到换行分隔符（0xA 字节）或 EOF。 找到后，所有字节（如果找到分隔符，则包括分隔符）将附加到 buf。

若成功，此函数将返回读取的总字节数。

如果此函数返回 Ok(0)，则流已达到 EOF。

错误

此函数的错误语义与 read_until 相同， 并且如果读取的字节不是有效的 UTF-8，也会返回错误。 如果遇到 I/O 错误，则在所有已读取数据均为有效 UTF-8 的情况下， buf 中可能已包含部分已读取的字节。

Cancel safety

此方法不是可取消安全的。如果该方法在 tokio::select! 语句中作为事件使用，且另一个分支先完成，则部分数据可能已被读取， 并且该部分数据会丢失。取消调用时，buf 的内容不提供任何保证。 当前实现会将 buf 替换为空字符串，但未来可能会有所变化。

由于字符串只能包含有效 UTF-8 的要求，此函数的行为与 read_until 不相同。 如果你需要一个可取消安全的 read_line，有三种选择：

Call read_until with a newline character and manually perform the utf-8 check.

The stream returned by lines has a cancellation safe next_line method.

Use tokio_util::codec::LinesCodec.

示例

std::io::Cursor 是实现 AsyncBufRead 的类型。 在此示例中，我们使用 Cursor 读取字节切片中的所有行：

use tokio::io::AsyncBufReadExt;

use std::io::Cursor;

let mut cursor = Cursor::new(b"foo\nbar");
let mut buf = String::new();

// cursor is at 'f'
let num_bytes = cursor.read_line(&mut buf)
    .await
    .expect("reading from cursor won't fail");

assert_eq!(num_bytes, 4);
assert_eq!(buf, "foo\n");
buf.clear();

// cursor is at 'b'
let num_bytes = cursor.read_line(&mut buf)
    .await
    .expect("reading from cursor won't fail");

assert_eq!(num_bytes, 3);
assert_eq!(buf, "bar");
buf.clear();

// cursor is at EOF
let num_bytes = cursor.read_line(&mut buf)
    .await
    .expect("reading from cursor won't fail");

assert_eq!(num_bytes, 0);
assert_eq!(buf, "");

fn split(self, byte: u8) -> Split<Self>
where Self: Sized + Unpin,

返回一个流，其内容按字节 byte 切分。

此方法是 BufRead::split 的异步等价方法。

此函数返回的流将产生 io::Result<Option<Vec<u8>>> 的实例。 每个返回的向量不包含末尾的分隔符字节。

错误

流中每个项的错误语义与 AsyncBufReadExt::read_until 相同。

示例

use tokio::io::AsyncBufReadExt;

let mut segments = my_buf_read.split(b'f');

while let Some(segment) = segments.next_segment().await? {
    println!("length = {}", segment.len())
}

fn fill_buf(&mut self) -> FillBuf<'_, Self>
where Self: Unpin,

返回内部缓冲的内容，如果内部缓冲为空，则从内部读取器读取更多数据填充。

此函数是较低级别的调用。 It needs to be paired with the consume method to function properly. When calling this method, none of the contents will be “read” in the sense that later calling read may return the same contents. As such, consume must be called with the number of bytes that are consumed from this buffer to ensure that the bytes are never returned twice.

返回空缓冲表示流已到达 EOF。

Equivalent to:

async fn fill_buf(&mut self) -> io::Result<&[u8]>;

错误

如果底层读取器被读取但返回了错误，此函数将返回 I/O 错误。

Cancel safety

此方法是取消安全的。若你将其作为事件用在 tokio::select! 语句中，且其他分支先完成，则保证尚未读取任何数据。

fn consume(&mut self, amt: usize)
where Self: Unpin,

通知此缓冲区已有 amt 个字节被消费， 因此在调用 read 时不应再返回这些字节。

此函数是较低级别的调用。 It needs to be paired with the fill_buf method to function properly. This function does not perform any I/O, it simply informs this object that some amount of its buffer, returned from fill_buf, has been consumed and should no longer be returned. As such, this function may do odd things if fill_buf isn’t called before calling it.

amt 必须小于 fill_buf 返回的缓冲区中的字节数。

fn lines(self) -> Lines<Self>
where Self: Sized,

返回此读取器各行的流。 This method is the async equivalent to BufRead::lines.

此函数返回的流将产生 io::Result<Option<String>> 的实例。 每个返回的字符串不包含末尾的换行字节（0xA 字节）或 CRLF（0xD, 0xA 字节）。

错误

流中每行的错误语义与 AsyncBufReadExt::read_line 相同。

示例

std::io::Cursor 是实现 BufRead 的类型。 在此示例中，我们使用 Cursor 遍历字节切片中的所有行。

use tokio::io::AsyncBufReadExt;

use std::io::Cursor;

let cursor = Cursor::new(b"lorem\nipsum\r\ndolor");

let mut lines = cursor.lines();

assert_eq!(lines.next_line().await.unwrap(), Some(String::from("lorem")));
assert_eq!(lines.next_line().await.unwrap(), Some(String::from("ipsum")));
assert_eq!(lines.next_line().await.unwrap(), Some(String::from("dolor")));
assert_eq!(lines.next_line().await.unwrap(), None);

动态兼容性

此 trait 不是dyn 兼容的。

在较早版本的 Rust 中，dyn 兼容性被称为"对象安全"，因此此 trait 不是对象安全的。