tokio::io

特性 AsyncReadExt

pub trait AsyncReadExt: AsyncRead {

    // Provided methods
    fn chain<R>(self, next: R) -> Chain<Self, R>
       where Self: Sized,
             R: AsyncRead { ... }
    fn read<'a>(&'a mut self, buf: &'a mut [u8]) -> Read<'a, Self>
       where Self: Unpin { ... }
    fn read_buf<'a, B>(&'a mut self, buf: &'a mut B) -> ReadBuf<'a, Self, B>
       where Self: Unpin,
             B: BufMut + ?Sized { ... }
    fn read_exact<'a>(&'a mut self, buf: &'a mut [u8]) -> ReadExact<'a, Self>
       where Self: Unpin { ... }
    fn read_u8(&mut self) -> ReadU8<&mut Self>
       where Self: Unpin { ... }
    fn read_i8(&mut self) -> ReadI8<&mut Self>
       where Self: Unpin { ... }
    fn read_u16(&mut self) -> ReadU16<&mut Self>
       where Self: Unpin { ... }
    fn read_i16(&mut self) -> ReadI16<&mut Self>
       where Self: Unpin { ... }
    fn read_u32(&mut self) -> ReadU32<&mut Self>
       where Self: Unpin { ... }
    fn read_i32(&mut self) -> ReadI32<&mut Self>
       where Self: Unpin { ... }
    fn read_u64(&mut self) -> ReadU64<&mut Self>
       where Self: Unpin { ... }
    fn read_i64(&mut self) -> ReadI64<&mut Self>
       where Self: Unpin { ... }
    fn read_u128(&mut self) -> ReadU128<&mut Self>
       where Self: Unpin { ... }
    fn read_i128(&mut self) -> ReadI128<&mut Self>
       where Self: Unpin { ... }
    fn read_f32(&mut self) -> ReadF32<&mut Self>
       where Self: Unpin { ... }
    fn read_f64(&mut self) -> ReadF64<&mut Self>
       where Self: Unpin { ... }
    fn read_u16_le(&mut self) -> ReadU16Le<&mut Self>
       where Self: Unpin { ... }
    fn read_i16_le(&mut self) -> ReadI16Le<&mut Self>
       where Self: Unpin { ... }
    fn read_u32_le(&mut self) -> ReadU32Le<&mut Self>
       where Self: Unpin { ... }
    fn read_i32_le(&mut self) -> ReadI32Le<&mut Self>
       where Self: Unpin { ... }
    fn read_u64_le(&mut self) -> ReadU64Le<&mut Self>
       where Self: Unpin { ... }
    fn read_i64_le(&mut self) -> ReadI64Le<&mut Self>
       where Self: Unpin { ... }
    fn read_u128_le(&mut self) -> ReadU128Le<&mut Self>
       where Self: Unpin { ... }
    fn read_i128_le(&mut self) -> ReadI128Le<&mut Self>
       where Self: Unpin { ... }
    fn read_f32_le(&mut self) -> ReadF32Le<&mut Self>
       where Self: Unpin { ... }
    fn read_f64_le(&mut self) -> ReadF64Le<&mut Self>
       where Self: Unpin { ... }
    fn read_to_end<'a>(
        &'a mut self,
        buf: &'a mut Vec<u8>,
    ) -> ReadToEnd<'a, Self>
       where Self: Unpin { ... }
    fn read_to_string<'a>(
        &'a mut self,
        dst: &'a mut String,
    ) -> ReadToString<'a, Self>
       where Self: Unpin { ... }
    fn take(self, limit: u64) -> Take<Self>
       where Self: Sized { ... }

}

从源读取字节。

作为扩展 trait 实现，向所有 AsyncRead 类型添加实用方法。 调用者通常会导入此 trait 而不是 AsyncRead。

use tokio::fs::File;
use tokio::io::{self, AsyncReadExt};

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut f = File::open("foo.txt").await?;
    let mut buffer = [0; 10];

    // The `read` method is defined by this trait.
    let n = f.read(&mut buffer[..]).await?;

    Ok(())
}

有关更多详细信息，请参阅 模块 文档。

提供方法

fn chain<R>(self, next: R) -> Chain<Self, R>
where Self: Sized, R: AsyncRead,

创建一个新的 AsyncRead 实例，将此流与 next 链接起来。

返回的 AsyncRead 实例将首先从此对象读取所有字节直到 EOF。之后的输出等同于 next 的输出。

示例

File 实现 AsyncRead：

use tokio::fs::File;
use tokio::io::{self, AsyncReadExt};

#[tokio::main]
async fn main() -> io::Result<()> {
    let f1 = File::open("foo.txt").await?;
    let f2 = File::open("bar.txt").await?;

    let mut handle = f1.chain(f2);
    let mut buffer = String::new();

    // read the value into a String. We could use any AsyncRead
    // method here, this is just one example.
    handle.read_to_string(&mut buffer).await?;
    Ok(())
}

fn read<'a>(&'a mut self, buf: &'a mut [u8]) -> Read<'a, Self>
where Self: Unpin,

从此源拉取一些字节到指定的缓冲中，返回读取的字节数。

等价于：

async fn read(&mut self, buf: &mut [u8]) -> io::Result<usize>;

此方法不保证是立即完成还是异步完成。

Return

若此方法的返回值为 Ok(n)，则必须保证 0 <= n <= buf.len()。非零的 n 值表示缓冲 buf 已从该源填充了 n 个字节的数据。若 n 为 0，则可能表示以下两种情况之一：

This reader has reached its “end of file” and will likely no longer be able to produce bytes. Note that this does not mean that the reader will always no longer be able to produce bytes.

The buffer specified was 0 bytes in length.

调用此函数时，不保证 buf 的内容，实现不能依赖 buf 内容的任何属性。建议实现仅向 buf 写入数据，而不是读取其内容。

相应地，但是，此方法的调用者不能假设实现使用 buf 的方式的任何保证。打算写入缓冲的代码也可能从其中读取。 请你负责确保在调用 read 之前 buf 已初始化。

错误

若此函数遇到任何形式的 I/O 或其他错误，则返回错误变体。若返回错误，则必须保证尚未读取任何字节。

Cancel safety

此方法是取消安全的。若你将其作为事件用在 tokio::select! 语句中，且其他分支先完成，则保证尚未读取任何数据。

示例

File 实现 Read：

use tokio::fs::File;
use tokio::io::{self, AsyncReadExt};

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut f = File::open("foo.txt").await?;
    let mut buffer = [0; 10];

    // read up to 10 bytes
    let n = f.read(&mut buffer[..]).await?;

    println!("The bytes: {:?}", &buffer[..n]);
    Ok(())
}

fn read_buf<'a, B>(&'a mut self, buf: &'a mut B) -> ReadBuf<'a, Self, B>
where Self: Unpin, B: BufMut + ?Sized,

从此源拉取一些字节到指定的缓冲中，前进缓冲的内部游标。

等价于：

async fn read_buf<B: BufMut>(&mut self, buf: &mut B) -> io::Result<usize>;

通常，即使提供的缓冲还有更多空间，也只会发出单个 read 系统调用。

此方法不保证是立即完成还是异步完成。

Return

非零的 n 值表示缓冲 buf 已从该源填充了 n 个字节的数据。若 n 为 0，则可能表示以下两种情况之一：

This reader has reached its “end of file” and will likely no longer be able to produce bytes. Note that this does not mean that the reader will always no longer be able to produce bytes.

The buffer specified had a remaining capacity of zero.

错误

若此函数遇到任何形式的 I/O 或其他错误，则返回错误变体。若返回错误，则必须保证尚未读取任何字节。

Cancel safety

此方法是取消安全的。若你将其作为事件用在 tokio::select! 语句中，且其他分支先完成，则保证尚未读取任何数据。

示例

File 实现了 Read，BytesMut 实现了 BufMut：

use tokio::fs::File;
use tokio::io::{self, AsyncReadExt};

use bytes::BytesMut;

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut f = File::open("foo.txt").await?;
    let mut buffer = BytesMut::with_capacity(10);

    assert!(buffer.is_empty());
    assert!(buffer.capacity() >= 10);

    // note that the return value is not needed to access the data
    // that was read as `buffer`'s internal cursor is updated.
    //
    // this might read more than 10 bytes if the capacity of `buffer`
    // is larger than 10.
    f.read_buf(&mut buffer).await?;

    println!("The bytes: {:?}", &buffer[..]);
    Ok(())
}

fn read_exact<'a>(&'a mut self, buf: &'a mut [u8]) -> ReadExact<'a, Self>
where Self: Unpin,

读取填充 buf 所需的精确字节数。

等价于：

async fn read_exact(&mut self, buf: &mut [u8]) -> io::Result<usize>;

此函数读取必要的字节数，以完全填充指定的缓冲 buf。

错误

若操作在完全填满缓冲之前遇到“文件结束”，则返回 ErrorKind::UnexpectedEof 类型的错误。这种情况下 buf 的内容是未指定的。

若遇到任何其他读取错误，操作将立即返回。这种情况下 buf 的内容是未指定的。

若此操作返回错误，则读取的字节数是未指定的，但绝不会读取超过完全填充缓冲所需的字节数。

Cancel safety

此方法不是可取消安全的。如果在 tokio::select! 语句中作为事件使用且另一个分支先完成， 则某些数据可能已经 被读入 buf。

示例

File 实现 Read：

use tokio::fs::File;
use tokio::io::{self, AsyncReadExt};

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut f = File::open("foo.txt").await?;
    let len = 10;
    let mut buffer = vec![0; len];

    // read exactly 10 bytes
    f.read_exact(&mut buffer).await?;
    Ok(())
}

fn read_u8(&mut self) -> ReadU8<&mut Self>
where Self: Unpin,

从底层 reader 读取一个无符号 8 位整数。

等价于：

async fn read_u8(&mut self) -> io::Result<u8>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则保证尚未读取任何数据。

示例

从 AsyncRead 读取无符号 8 位整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![2, 5]);

assert_eq!(2, reader.read_u8().await?);
assert_eq!(5, reader.read_u8().await?);

Ok(())

fn read_i8(&mut self) -> ReadI8<&mut Self>
where Self: Unpin,

从底层 reader 读取一个有符号 8 位整数。

等价于：

async fn read_i8(&mut self) -> io::Result<i8>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则保证尚未读取任何数据。

示例

从 AsyncRead 读取无符号 8 位整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x02, 0xfb]);

assert_eq!(2, reader.read_i8().await?);
assert_eq!(-5, reader.read_i8().await?);

Ok(())

fn read_u16(&mut self) -> ReadU16<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个无符号 16 位整数。

等价于：

async fn read_u16(&mut self) -> io::Result<u16>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 16 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![2, 5, 3, 0]);

assert_eq!(517, reader.read_u16().await?);
assert_eq!(768, reader.read_u16().await?);
Ok(())

fn read_i16(&mut self) -> ReadI16<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个有符号 16 位整数。

等价于：

async fn read_i16(&mut self) -> io::Result<i16>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 16 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x00, 0xc1, 0xff, 0x7c]);

assert_eq!(193, reader.read_i16().await?);
assert_eq!(-132, reader.read_i16().await?);
Ok(())

fn read_u32(&mut self) -> ReadU32<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个无符号 32 位整数。

等价于：

async fn read_u32(&mut self) -> io::Result<u32>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 32 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x00, 0x00, 0x01, 0x0b]);

assert_eq!(267, reader.read_u32().await?);
Ok(())

fn read_i32(&mut self) -> ReadI32<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个有符号 32 位整数。

等价于：

async fn read_i32(&mut self) -> io::Result<i32>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 32 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0xff, 0xff, 0x7a, 0x33]);

assert_eq!(-34253, reader.read_i32().await?);
Ok(())

fn read_u64(&mut self) -> ReadU64<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个无符号 64 位整数。

等价于：

async fn read_u64(&mut self) -> io::Result<u64>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 64 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0x00, 0x03, 0x43, 0x95, 0x4d, 0x60, 0x86, 0x83
]);

assert_eq!(918733457491587, reader.read_u64().await?);
Ok(())

fn read_i64(&mut self) -> ReadI64<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个有符号 64 位整数。

等价于：

async fn read_i64(&mut self) -> io::Result<i64>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 64 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x80, 0, 0, 0, 0, 0, 0, 0]);

assert_eq!(i64::MIN, reader.read_i64().await?);
Ok(())

fn read_u128(&mut self) -> ReadU128<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个无符号 128 位整数。

等价于：

async fn read_u128(&mut self) -> io::Result<u128>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 128 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
        0x00, 0x03, 0x43, 0x95, 0x4d, 0x60, 0x86, 0x83,
    0x00, 0x03, 0x43, 0x95, 0x4d, 0x60, 0x86, 0x83
]);

assert_eq!(16947640962301618749969007319746179, reader.read_u128().await?);
Ok(())

fn read_i128(&mut self) -> ReadI128<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个有符号 128 位整数。

等价于：

async fn read_i128(&mut self) -> io::Result<i128>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 128 位大端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0x80, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0
]);

assert_eq!(i128::MIN, reader.read_i128().await?);
Ok(())

fn read_f32(&mut self) -> ReadF32<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个 32 位浮点类型。

等价于：

async fn read_f32(&mut self) -> io::Result<f32>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取 32 位浮点类型：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0xff, 0x7f, 0xff, 0xff]);

assert_eq!(f32::MIN, reader.read_f32().await?);
Ok(())

fn read_f64(&mut self) -> ReadF64<&mut Self>
where Self: Unpin,

从底层读取器以大端顺序读取一个 64 位浮点类型。

等价于：

async fn read_f64(&mut self) -> io::Result<f64>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取 64 位浮点类型：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0xff, 0xef, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
]);

assert_eq!(f64::MIN, reader.read_f64().await?);
Ok(())

fn read_u16_le(&mut self) -> ReadU16Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个无符号 16 位整数。

等价于：

async fn read_u16_le(&mut self) -> io::Result<u16>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 16 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![2, 5, 3, 0]);

assert_eq!(1282, reader.read_u16_le().await?);
assert_eq!(3, reader.read_u16_le().await?);
Ok(())

fn read_i16_le(&mut self) -> ReadI16Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个有符号 16 位整数。

等价于：

async fn read_i16_le(&mut self) -> io::Result<i16>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 16 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x00, 0xc1, 0xff, 0x7c]);

assert_eq!(-16128, reader.read_i16_le().await?);
assert_eq!(31999, reader.read_i16_le().await?);
Ok(())

fn read_u32_le(&mut self) -> ReadU32Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个无符号 32 位整数。

等价于：

async fn read_u32_le(&mut self) -> io::Result<u32>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 32 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x00, 0x00, 0x01, 0x0b]);

assert_eq!(184614912, reader.read_u32_le().await?);
Ok(())

fn read_i32_le(&mut self) -> ReadI32Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个有符号 32 位整数。

等价于：

async fn read_i32_le(&mut self) -> io::Result<i32>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 32 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0xff, 0xff, 0x7a, 0x33]);

assert_eq!(863698943, reader.read_i32_le().await?);
Ok(())

fn read_u64_le(&mut self) -> ReadU64Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个无符号 64 位整数。

等价于：

async fn read_u64_le(&mut self) -> io::Result<u64>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 64 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0x00, 0x03, 0x43, 0x95, 0x4d, 0x60, 0x86, 0x83
]);

assert_eq!(9477368352180732672, reader.read_u64_le().await?);
Ok(())

fn read_i64_le(&mut self) -> ReadI64Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个有符号 64 位整数。

等价于：

async fn read_i64_le(&mut self) -> io::Result<i64>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 64 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0x80, 0, 0, 0, 0, 0, 0, 0]);

assert_eq!(128, reader.read_i64_le().await?);
Ok(())

fn read_u128_le(&mut self) -> ReadU128Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个无符号 128 位整数。

等价于：

async fn read_u128_le(&mut self) -> io::Result<u128>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取无符号 128 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0x00, 0x03, 0x43, 0x95, 0x4d, 0x60, 0x86, 0x83,
    0x00, 0x03, 0x43, 0x95, 0x4d, 0x60, 0x86, 0x83
]);

assert_eq!(174826588484952389081207917399662330624, reader.read_u128_le().await?);
Ok(())

fn read_i128_le(&mut self) -> ReadI128Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个有符号 128 位整数。

等价于：

async fn read_i128_le(&mut self) -> io::Result<i128>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取有符号 128 位小端整数：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0x80, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0
]);

assert_eq!(128, reader.read_i128_le().await?);
Ok(())

fn read_f32_le(&mut self) -> ReadF32Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个 32 位浮点类型。

等价于：

async fn read_f32_le(&mut self) -> io::Result<f32>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取 32 位浮点类型：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![0xff, 0xff, 0x7f, 0xff]);

assert_eq!(f32::MIN, reader.read_f32_le().await?);
Ok(())

fn read_f64_le(&mut self) -> ReadF64Le<&mut Self>
where Self: Unpin,

从底层读取器以小端顺序读取一个 64 位浮点类型。

等价于：

async fn read_f64_le(&mut self) -> io::Result<f64>;

建议使用带缓冲的读取器，以避免过多的系统调用。

错误

此方法返回与 AsyncReadExt::read_exact 相同的错误。

Cancel safety

此方法不是取消安全的。若将此方法用作 tokio::select! 语句中的事件，且其他分支先完成，则可能会丢失一些数据。

示例

从 AsyncRead 读取 64 位浮点类型：

use tokio::io::{self, AsyncReadExt};

use std::io::Cursor;

let mut reader = Cursor::new(vec![
    0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xef, 0xff
]);

assert_eq!(f64::MIN, reader.read_f64_le().await?);
Ok(())

fn read_to_end<'a>(&'a mut self, buf: &'a mut Vec<u8>) -> ReadToEnd<'a, Self>
where Self: Unpin,

从此源读取所有字节直到 EOF，将它们放入 buf。

等价于：

async fn read_to_end(&mut self, buf: &mut Vec<u8>) -> io::Result<usize>;

从此源读取的所有字节都将附加到指定的 缓冲区 buf。此函数会持续调用 read() 来 将更多数据附加到 buf，直到 read() 返回 Ok(0)。

如果成功，则返回读取的总字节数。

错误

若遇到读取错误，read_to_end 操作将立即完成。已读取的所有字节将追加到 buf。

示例

File 实现 Read：

use tokio::io::{self, AsyncReadExt};
use tokio::fs::File;

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut f = File::open("foo.txt").await?;
    let mut buffer = Vec::new();

    // read the whole file
    f.read_to_end(&mut buffer).await?;
    Ok(())
}

（另请参阅用于从文件读取的便捷函数 tokio::fs::read。）

fn read_to_string<'a>( &'a mut self, dst: &'a mut String, ) -> ReadToString<'a, Self>
where Self: Unpin,

从此源读取所有字节直到 EOF，将它们追加到 buf。

等价于：

async fn read_to_string(&mut self, buf: &mut String) -> io::Result<usize>;

若成功，返回已读取并追加到 buf 的字节数。

错误

若此流中的数据不是有效的 UTF-8，则返回错误且 buf 保持不变。

有关其他错误语义，请参阅 read_to_end。

示例

File 实现 Read：

use tokio::io::{self, AsyncReadExt};
use tokio::fs::File;

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut f = File::open("foo.txt").await?;
    let mut buffer = String::new();

    f.read_to_string(&mut buffer).await?;
    Ok(())
}

（另请参阅用于从文件读取的便捷函数 crate::fs::read_to_string。）

fn take(self, limit: u64) -> Take<Self>
where Self: Sized,

创建一个适配器，从其中最多读取 limit 个字节。

此函数返回 AsyncRead 的一个新实例，最多读取 limit 个字节，之后将始终返回 EOF（Ok(0)）。 Any read errors will not count towards the number of bytes read and future calls to read() may succeed.

示例

File 实现 Read：

use tokio::io::{self, AsyncReadExt};
use tokio::fs::File;

#[tokio::main]
async fn main() -> io::Result<()> {
    let f = File::open("foo.txt").await?;
    let mut buffer = [0; 5];

    // read at most five bytes
    let mut handle = f.take(5);

    handle.read(&mut buffer).await?;
    Ok(())
}

动态兼容性

此 trait 不是dyn 兼容的。

在较早版本的 Rust 中，dyn 兼容性被称为"对象安全"，因此此 trait 不是对象安全的。