Function tokio::io::util::write_all_buf::write_all_buf

pub(crate) fn write_all_buf<'a, W, B>(
    writer: &'a mut W, 
    buf: &'a mut B
) -> WriteAllBuf<'a, W, B> where
    W: AsyncWrite + Unpin,
    B: Buf, 

Tries to write some bytes from the given buf to the writer in an asynchronous manner, returning a future.