tokio::loom::std
Module sync

Modules

atomic 🔒

Structs

Arc 🔒

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

WaitTimeoutResult 🔒

A type indicating whether a timed wait on a condition variable returned due to a time out or not.

Weak 🔒

Weak is a version of Arc that holds a non-owning reference to the managed allocation. The allocation is accessed by calling upgrade on the Weak pointer, which returns an Option<Arc<T>>.