Module tokio::loom::std::sync::atomic

Structs

AtomicBool

A boolean type which can be safely shared between threads.

AtomicU64

An integer type which can be safely shared between threads.

Enums

Ordering

Atomic memory orderings

Functions

fence

An atomic fence.

spin_loop_hintDeprecated

Signals the processor that it is inside a busy-wait spin-loop (“spin lock”).