tokio

宏 join

macro_rules! join {
    ($(biased;)? $($future:expr),*) => { ... };
}

等待多个并发分支，当所有分支完成时返回。

join! 宏必须在 async 函数、闭包和块内使用。

join! 宏接受一个 async 表达式列表，并在同一任务上并发地执行它们。 每个 async 表达式求值为一个 future，来自每个表达式的 future 在当前任务上多路复用。

对于返回 Result 的异步表达式，join! 将等待 所有分支完成，无论是否有任何分支以 Err 完成。 使用 try_join! 在遇到 Err 时提早返回。

Notes

所提供的 future 内联存储，无需分配 Vec。

Runtime characteristics

由于所有异步表达式都在当前任务上运行，这些表达式 能够并发运行，但不能并行运行。这意味着所有 表达式都在同一线程上运行，如果某个分支阻塞了线程， 所有其他表达式都将无法继续。如果需要并行执行， 请使用 tokio::spawn 生成每个异步表达式， 然后将其 join handle 传递给 join!。

Fairness

默认情况下，每次唤醒时，join! 生成的 future 会轮换首先 轮询哪个包含的 future。

可以通过在宏使用的开头添加 biased; 来覆盖此行为。 有关详细信息，请参阅示例。这将使 join 按 future 出现的顺序（从上到下）进行轮询。

如果您的 future 之间的交互中已知的 poll 顺序很重要，则可能需要这样做。

但此模式有一个重要的注意事项。您需要自行确保 future 的 poll 顺序是公平的。例如，如果您 join 一个流和一个 shutdown future，并且该流每次 poll 需要处理大量消息， 那么该流可能会一直阻止 shutdown future 被 poll，导致 shutdown future 永远无法执行。u should place the shutdown future earlier in the join! list to ensure that it is always polled, and will not be delayed due to the stream future taking a long time to return Poll::Pending.

示例

两个分支的基本 join

async fn do_stuff_async() {
    // async work
}

async fn more_async_work() {
    // more here
}

let (first, second) = tokio::join!(
    do_stuff_async(),
    more_async_work());

// do something with the values

使用 biased; 模式控制轮询顺序。

async fn do_stuff_async() {
    // async work
}

async fn more_async_work() {
    // more here
}

let (first, second) = tokio::join!(
    biased;
    do_stuff_async(),
    more_async_work()
);

// do something with the values