tokio

Module net

tokio 的 TCP/UDP/Unix 绑定。

此模块包含 TCP/UDP/Unix 网络类型，类似于标准库，可用于实现网络协议。

Organization

TcpListener and TcpStream provide functionality for communication over TCP

UdpSocket provides functionality for communication over UDP

UnixListener and UnixStream provide functionality for communication over a Unix Domain Stream Socket (available on Unix only)

UnixDatagram provides functionality for communication over Unix Domain Datagram Socket (available on Unix only)

tokio::net::unix::pipe for FIFO pipes (available on Unix only)

tokio::net::windows::named_pipe for Named Pipes (available on Windows only)

对于 tokio::net 中不可用的 IO 资源，你可以使用 AsyncFd。

模块

tcp

TCP utility types.

windows

Windows specific network types.

结构体

TcpListener

A TCP socket server, listening for connections.

TcpSocket

A TCP socket that has not yet been converted to a TcpStream or TcpListener.

TcpStream

A TCP stream between a local and a remote socket.

UdpSocket

A UDP socket.

trait

ToSocketAddrs

Converts or resolves without blocking to one or more SocketAddr values.

函数

lookup_host

Performs a DNS resolution.