Module tokio::net::tcp::split_owned

TcpStream owned split support.

A TcpStream can be split into an OwnedReadHalf and a OwnedWriteHalf with the TcpStream::into_split method. OwnedReadHalf implements AsyncRead while OwnedWriteHalf implements AsyncWrite.

Compared to the generic split of AsyncRead + AsyncWrite, this specialized split has no associated overhead and enforces all invariants at the type level.

Structs

OwnedReadHalf
Owned read half of a TcpStream, created by into_split.

OwnedWriteHalf
Owned write half of a TcpStream, created by into_split.

ReuniteError
Error indicating two halves were not from the same socket, and thus could not be reunited.

Functions

reunite

split_owned