Struct tokio::net::tcp::Incoming

#[must_use = "streams do nothing unless polled"]pub struct Incoming<'a> { /* fields omitted */ }

Stream returned by the TcpListener::incoming function representing the stream of sockets received from a listener.

Implementations

impl<'_> Incoming<'_>

pub fn poll_accept(
self: Pin<&mut Self>,
cx: &mut Context
) -> Poll<Result<TcpStream>>

Attempts to poll TcpStream by polling inner TcpListener to accept connection.

If TcpListener isn't ready yet, Poll::Pending is returned and current task will be notified by a waker.