Struct tokio::net::udp::RecvHalf

pub struct RecvHalf(_);

The recv half after split.

Use recv_from or recv to receive datagrams.

Implementations

impl RecvHalf

pub fn reunite(self, other: SendHalf) -> Result<UdpSocket, ReuniteError>

Attempts to put the two "halves" of a UdpSocket back together and recover the original socket. Succeeds only if the two "halves" originated from the same call to UdpSocket::split.

pub async fn recv_from<'_, '_>(
&'_ mut self,
buf: &'_ mut [u8]
) -> Result<(usize, SocketAddr)>

Returns a future that receives a single datagram on the socket. On success, the future resolves to the number of bytes read and the origin.

The function must be called with valid byte array buf of sufficient size to hold the message bytes. If a message is too long to fit in the supplied buffer, excess bytes may be discarded.

pub async fn recv<'_, '_>(&'_ mut self, buf: &'_ mut [u8]) -> Result<usize>

Returns a future that receives a single datagram message on the socket from the remote address to which it is connected. On success, the future will resolve to the number of bytes read.

The function must be called with valid byte array buf of sufficient size to hold the message bytes. If a message is too long to fit in the supplied buffer, excess bytes may be discarded.

The connect method will connect this socket to a remote address. The future will fail if the socket is not connected.