Struct tokio::net::unix::UnixDatagram

pub struct UnixDatagram { /* fields omitted */ }

An I/O object representing a Unix datagram socket.

Methods

impl UnixDatagram

pub fn bind<P>(path: P) -> Result<UnixDatagram, Error> where
P: AsRef<Path>,

Creates a new UnixDatagram bound to the specified path.

pub fn pair() -> Result<(UnixDatagram, UnixDatagram), Error>

Creates an unnamed pair of connected sockets.

This function will create a pair of interconnected Unix sockets for communicating back and forth between one another. Each socket will be associated with the default event loop's handle.

pub fn from_std(
datagram: UnixDatagram,
handle: &Handle
) -> Result<UnixDatagram, Error>

Consumes a UnixDatagram in the standard library and returns a nonblocking UnixDatagram from this crate.

The returned datagram will be associated with the given event loop specified by handle and is ready to perform I/O.

pub fn unbound() -> Result<UnixDatagram, Error>

Creates a new UnixDatagram which is not bound to any address.

pub fn connect<P>(&self, path: P) -> Result<(), Error> where
P: AsRef<Path>,

Connects the socket to the specified address.

The send method may be used to send data to the specified address. recv and recv_from will only receive data from that address.

pub fn poll_read_ready(&self, ready: Ready) -> Result<Async<Ready>, Error>

Test whether this socket is ready to be read or not.

pub fn poll_write_ready(&self) -> Result<Async<Ready>, Error>

Test whether this socket is ready to be written to or not.

pub fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local address that this socket is bound to.

pub fn peer_addr(&self) -> Result<SocketAddr, Error>

Returns the address of this socket's peer.

The connect method will connect the socket to a peer.

pub fn poll_recv_from(
&self,
buf: &mut [u8]
) -> Result<Async<(usize, SocketAddr)>, Error>

Receives data from the socket.

On success, returns the number of bytes read and the address from whence the data came.

pub fn poll_recv(&self, buf: &mut [u8]) -> Result<Async<usize>, Error>

Receives data from the socket.

On success, returns the number of bytes read.

pub fn recv_dgram<T>(self, buf: T) -> RecvDgram<T> where
T: AsMut<[u8]>,

Returns a future for receiving a datagram. See the documentation on RecvDgram for details.

pub fn poll_send_to<P>(
&self,
buf: &[u8],
path: P
) -> Result<Async<usize>, Error> where
P: AsRef<Path>,

Sends data on the socket to the specified address.

On success, returns the number of bytes written.

pub fn poll_send(&self, buf: &[u8]) -> Result<Async<usize>, Error>

Sends data on the socket to the socket's peer.

The peer address may be set by the connect method, and this method will return an error if the socket has not already been connected.

On success, returns the number of bytes written.

pub fn send_dgram<T, P>(self, buf: T, path: P) -> SendDgram<T, P> where
P: AsRef<Path>,
T: AsRef<[u8]>,

Returns a future sending the data in buf to the socket at path.

pub fn take_error(&self) -> Result<Option<Error>, Error>

Returns the value of the SO_ERROR option.

pub fn shutdown(&self, how: Shutdown) -> Result<(), Error>

Shut down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).