Module tokio::net::windows::named_pipe

Tokio support for Windows named pipes.

Structs

ClientOptions

A builder suitable for building and interacting with named pipes from the client side.

NamedPipeClient

A Windows named pipe client.

NamedPipeServer

A Windows named pipe server.

PipeInfo

Information about a named pipe.

ServerOptions

A builder structure for construct a named pipe with named pipe-specific options. This is required to use for named pipe servers who wants to modify pipe-related options.

Enums

PipeEnd

Indicates the end of a named pipe.

PipeMode

The pipe mode of a named pipe.