Trait tokio::park::Unpark

pub(crate) trait Unpark: Sync + Send + 'static {
    fn unpark(&self);
}

Unblock a thread blocked by the associated Park instance.

Required methods

fn unpark(&self)

Unblocks a thread that is blocked by the associated Park handle.

Calling unpark atomically makes available the unpark token, if it is not already available.

Panics

This function should not panic, but ultimately, panics are left as an implementation detail. Refer to the documentation for the specific Unpark implementation.