Enum tokio::prelude::future::Either

pub enum Either<A, B> {
    A(A),
    B(B),
}

Combines two different futures yielding the same item and error types into a single type.

Variants

A(A)

First branch of the type

B(B)

Second branch of the type

Methods

impl<T, A, B> Either<(T, A), (T, B)>

pub fn split(self) -> (T, Either<A, B>)

Splits out the homogeneous type from an either of tuples.

This method is typically useful when combined with the Future::select2 combinator.