Function tokio::prelude::stream::channel

pub fn channel<T, E>() -> (Sender<T, E>, Receiver<T, E>)

Deprecated since 0.1.4
: use sync::mpsc::channel instead

Creates an in-memory channel implementation of the Stream trait.

This method creates a concrete implementation of the Stream trait which can be used to send values across threads in a streaming fashion. This channel is unique in that it implements back pressure to ensure that the sender never outpaces the receiver. The Sender::send method will only allow sending one message and the next message can only be sent once the first was consumed.

The Receiver returned implements the Stream trait and has access to any number of the associated combinators for transforming the result.