Struct tokio::prelude::stream::Sender

pub struct Sender<T, E> { /* fields omitted */ }

Deprecated since 0.1.4
: use sync::mpsc::channel instead

The transmission end of a channel which is used to send values.

This is created by the channel method in the stream module.

Methods

impl<T, E> Sender<T, E>

pub fn send(self, t: Result<T, E>) -> FutureSender<T, E>

Deprecated since 0.1.4
: use sync::mpsc::channel instead

Sends a new value along this channel to the receiver.

This method consumes the sender and returns a future which will resolve to the sender again when the value sent has been consumed.