Struct tokio::prelude::stream::Wait

#[must_use = "iterators do nothing unless advanced"]

pub struct Wait<S> { /* fields omitted */ }

A stream combinator which converts an asynchronous stream to a blocking iterator.

Created by the Stream::wait method, this function transforms any stream into a standard iterator. This is implemented by blocking the current thread while items on the underlying stream aren't ready yet.

Methods

impl<S> Wait<S>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized,

pub fn get_ref(&self) -> &S

Acquires a reference to the underlying stream that this combinator is pulling from.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized,

pub fn get_mut(&mut self) -> &mut S

Acquires a mutable reference to the underlying stream that this combinator is pulling from.

Note that care must be taken to avoid tampering with the state of the stream which may otherwise confuse this combinator.

pub fn into_inner(self) -> S

Consumes this combinator, returning the underlying stream.

Note that this may discard intermediate state of this combinator, so care should be taken to avoid losing resources when this is called.