Trait tokio::prelude::task::Executor

pub trait Executor: 'static + Send + Sync {
    fn execute(&self, r: Run);
}

Deprecated

A trait representing requests to poll futures.

This trait is an argument to the Spawn::execute which is used to run a future to completion. An executor will receive requests to run a future and an executor is responsible for ensuring that happens in a timely fashion.

Note that this trait is likely to be deprecated and/or renamed to avoid clashing with the future::Executor trait. If you've got a use case for this or would like to comment on the name please let us know!

Required methods

fn execute(&self, r: Run)

Deprecated

Requests that Run is executed soon on the given executor.

Loading content...