Trait tokio::prelude::AsyncWriteExt

pub trait AsyncWriteExt: AsyncWrite {
    default fn write_async(&'a mut self, buf: &'a [u8]) -> Write<'a, Self> { ... }

    default fn write_all_async(
        &'a mut self, 
        buf: &'a [u8]
    ) -> WriteAll<'a, Self> { ... }

    default fn flush_async(&mut self) -> Flush<Self> { ... }
}

An extension trait which adds utility methods to AsyncWrite types.

Provided methods

default fn write_async(&'a mut self, buf: &'a [u8]) -> Write<'a, Self>

Write data into this object.

Creates a future that will write the entire contents of the buffer buf into this AsyncWrite.

The returned future will not complete until all the data has been written.

Examples

#![feature(async_await, await_macro, futures_api)]
tokio::run_async(async {
// The extension trait can also be imported with
// `use tokio::prelude::*`.
use tokio::prelude::AsyncWriteExt;
use std::io::Cursor;

let mut buf = [0u8; 5];
let mut writer = Cursor::new(&mut buf[..]);

let n = await!(writer.write_async(&[1, 2, 3, 4])).unwrap();

assert_eq!(writer.into_inner()[..n], [1, 2, 3, 4, 0][..n]);
});

default fn write_all_async(&'a mut self, buf: &'a [u8]) -> WriteAll<'a, Self>

Write an entire buffer into this object.

Creates a future that will write the entire contents of the buffer buf into this AsyncWrite.

The returned future will not complete until all the data has been written.

Examples

#![feature(async_await, await_macro, futures_api)]
tokio::run_async(async {
// The extension trait can also be imported with
// `use tokio::prelude::*`.
use tokio::prelude::AsyncWriteExt;
use std::io::Cursor;

let mut buf = [0u8; 5];
let mut writer = Cursor::new(&mut buf[..]);

await!(writer.write_all_async(&[1, 2, 3, 4])).unwrap();

assert_eq!(writer.into_inner(), [1, 2, 3, 4, 0]);
});

default fn flush_async(&mut self) -> Flush<Self>

Creates a future which will entirely flush this AsyncWrite.

Examples

#![feature(async_await, await_macro, futures_api)]
tokio::run_async(async {
// The extension trait can also be imported with
// `use tokio::prelude::*`.
use tokio::prelude::AsyncWriteExt;
use std::io::{BufWriter, Cursor};

let mut output = [0u8; 5];

{
    let mut writer = Cursor::new(&mut output[..]);
    let mut buffered = BufWriter::new(writer);
    await!(buffered.write_all_async(&[1, 2])).unwrap();
    await!(buffered.write_all_async(&[3, 4])).unwrap();
    await!(buffered.flush_async()).unwrap();
}

assert_eq!(output, [1, 2, 3, 4, 0]);
});

Loading content...