Trait tokio::prelude::IntoFuture

pub trait IntoFuture where
    <Self::Future as Future>::Item == Self::Item,
    <Self::Future as Future>::Error == Self::Error, {
    type Future: Future;
    type Item;
    type Error;
    fn into_future(self) -> Self::Future;
}

Class of types which can be converted into a future.

This trait is very similar to the IntoIterator trait and is intended to be used in a very similar fashion.

Associated Types

type Future: Future

The future that this type can be converted into.

type Item

The item that the future may resolve with.

type Error

The error that the future may resolve with.

Loading content...
Required methods

fn into_future(self) -> Self::Future

Consumes this object and produces a future.

Loading content...