Trait tokio::process::imp::orphan::Wait

pub(crate) trait Wait {
    // Required methods
    fn id(&self) -> u32;
    fn try_wait(&mut self) -> Result<Option<ExitStatus>>;
}

An interface for waiting on a process to exit.

Required Methods

fn id(&self) -> u32

Get the identifier for this process or diagnostics.

fn try_wait(&mut self) -> Result<Option<ExitStatus>>

Try waiting for a process to exit in a non-blocking manner.