Module tokio::reactor

The core reactor driving all I/O.

This module contains the Reactor reactor type which is the event loop for all I/O happening in tokio. This core reactor (or event loop) is used to drive I/O resources.

The Handle struct, created by handle, is a reference to the event loop and can be used when constructing I/O objects.

Lastly PollEvented can be used to construct I/O objects that interact with the event loop, e.g. TcpStream in the net module.

Structs

Handle
A handle to an event loop.

PollEvented
A concrete implementation of a stream of readiness notifications for I/O objects that originates from an event loop.

Reactor
The core reactor, or event loop.

SetFallbackError
Error returned from Handle::set_fallback.

Turn
Return value from the turn method on Reactor.