Struct tokio::reactor::Reactor

pub struct Reactor { /* fields omitted */ }

The core reactor, or event loop.

The event loop is the main source of blocking in an application which drives all other I/O events and notifications happening. Each event loop can have multiple handles pointing to it, each of which can then be used to create various I/O objects to interact with the event loop in interesting ways.

Methods

impl Reactor

pub fn new() -> Result<Reactor>

Creates a new event loop, returning any error that happened during the creation.

pub fn handle(&self) -> Handle

Returns a handle to this event loop which cannot be sent across threads but can be used as a proxy to the event loop itself.

Handles are cloneable and clones always refer to the same event loop. This handle is typically passed into functions that create I/O objects to bind them to this event loop.

pub fn turn(&mut self, max_wait: Option<Duration>) -> Turn

Performs one iteration of the event loop, blocking on waiting for events for at most max_wait (forever if None).

This method is the primary method of running this reactor and processing I/O events that occur. This method executes one iteration of an event loop, blocking at most once waiting for events to happen.

If a max_wait is specified then the method should block no longer than the duration specified, but this shouldn't be used as a super-precise timer but rather a "ballpark approximation"

Return value

This function returns an instance of Turn which as of today has no extra information with it and can be safely discarded. In the future this return value may contain information about what happened while this reactor blocked.