Module tokio::runtime::basic_scheduler

Structs

BasicScheduler

Executes tasks on the current thread

Context

Thread-local context.

Inner

The inner scheduler that owns the task queue and the main parker P.

InnerGuard

Used to ensure we always place the Inner value back into its slot in BasicScheduler, even if the future panics.

Shared

Scheduler state shared between threads.

Spawner

Tasks

Enums

RemoteMsg

A remote scheduler entry.

Constants

INITIAL_CAPACITY

Initial queue capacity.

MAX_TASKS_PER_TICK

REMOTE_FIRST_INTERVAL

How often to check the remote queue first.

Statics

CURRENT

Functions

enter

Enters the scheduler context. This sets the queue and other necessary scheduler state in the thread-local.