tokio::runtime::blocking

Module pool

Thread pool for blocking operations

Structs

BlockingPool 🔒

Inner 🔒

Shared 🔒

Spawner 🔒

SpawnerMetrics 🔒

Task 🔒

Enums

Mandatory 🔒

SpawnError 🔒

Constants

KEEP_ALIVE 🔒

Functions

is_temporary_os_thread_error 🔒

spawn_blocking 🔒

Runs the provided function on an executor dedicated to blocking operations. Tasks will be scheduled as non-mandatory, meaning they may not get executed in case of runtime shutdown.

spawn_mandatory_blocking 🔒

Runs the provided function on an executor dedicated to blocking operations. Tasks will be scheduled as mandatory, meaning they are guaranteed to run unless a shutdown is already taking place. In case a shutdown is already taking place, None will be returned.