Function tokio::runtime::current_thread::block_on_all

pub fn block_on_all<F>(future: F) -> Result<F::Item, F::Error> where
    F: Future, 

Run the provided future to completion using a runtime running on the current thread.

This first creates a new [Runtime], and calls [Runtime::block_on] with the provided future, which blocks the current thread until the provided future completes. It then calls [Runtime::run] to wait for any other spawned futures to resolve.