Struct tokio::runtime::current_thread::Builder

pub struct Builder { /* fields omitted */ }

Builds a Single-threaded runtime with custom configuration values.

Methods can be chained in order to set the configuration values. The Runtime is constructed by calling build.

New instances of Builder are obtained via Builder::new.

See function level documentation for details on the various configuration settings.

Examples

extern crate tokio;
extern crate tokio_timer;

use tokio::runtime::current_thread::Builder;
use tokio_timer::clock::Clock;

// build Runtime
let runtime = Builder::new()
    .clock(Clock::new())
    .build();
// ... call runtime.run(...)

Methods

impl Builder

pub fn new() -> Builder

Returns a new runtime builder initialized with default configuration values.

Configuration methods can be chained on the return value.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn clock(&mut self, clock: Clock) -> &mut Self

Set the Clock instance that will be used by the runtime.

pub fn build(&mut self) -> Result<Runtime>

Create the configured Runtime.