Type Alias tokio::runtime::driver::ProcessDriver

type ProcessDriver = Driver;

Aliased Type

struct ProcessDriver {
    signal_ready: bool,
    events: Events,
    poll: Poll,
}

Fields
signal_ready: bool

True when an event with the signal token is received

events: Events

Reuse the mio::Events value across calls to poll.

poll: Poll

The system event queue.

Implementations

impl Driver

pub(crate) fn new(nevents: usize) -> Result<(Driver, Handle)>

Creates a new event loop, returning any error that happened during the creation.

pub(crate) fn park(&mut self, rt_handle: &Handle)

pub(crate) fn park_timeout(&mut self, rt_handle: &Handle, duration: Duration)

pub(crate) fn shutdown(&mut self, rt_handle: &Handle)

fn turn(&mut self, handle: &Handle, max_wait: Option<Duration>)