Module tokio::runtime::enter

Structs

Enter
Represents an executor context.

Enums

EnterContext

Constants

ENTERED

Functions

enter
Marks the current thread as being within the dynamic extent of an executor.

try_enter
Tries to enter a runtime context, returns None if already in a runtime context.