tokio::runtime::scheduler::current_thread

Struct CoreGuard

struct CoreGuard<'a> {
    context: Context,
    scheduler: &'a CurrentThread,
}

Used to ensure we always place the Core value back into its slot in CurrentThread, even if the future panics.

Fields
context: Contextscheduler: &'a CurrentThread
Implementations

impl CoreGuard<'_>

fn block_on<F: Future>(self, future: F) -> F::Output

fn enter<F, R>(self, f: F) -> R
where F: FnOnce(Box<Core>, &Context) -> (Box<Core>, R),

Enters the scheduler context. This sets the queue and other necessary scheduler state in the thread-local.