tokio::runtime::scheduler

Enum Handle

pub(crate) enum Handle {
    CurrentThread(Arc<Handle>),
    MultiThread(Arc<Handle>),
}

Variants

CurrentThread(Arc<Handle>)

MultiThread(Arc<Handle>)

Implementations

impl Handle

pub(crate) fn driver(&self) -> &Handle

impl Handle

pub(crate) fn current() -> Handle

pub(crate) fn blocking_spawner(&self) -> &Spawner

pub(crate) fn is_local(&self) -> bool

pub(crate) fn timer_flavor(&self) -> TimerFlavor

pub(crate) fn can_spawn_local_on_local_runtime(&self) -> bool

Returns true if this is a local runtime and the runtime is owned by the current thread.

pub(crate) fn spawn<F>( &self, future: F, id: Id, spawned_at: SpawnLocation, ) -> JoinHandle<F::Output>
where F: Future + Send + 'static, F::Output: Send + 'static,

pub(crate) unsafe fn spawn_local<F>( &self, future: F, id: Id, spawned_at: SpawnLocation, ) -> JoinHandle<F::Output>
where F: Future + 'static, F::Output: 'static,

Spawn a local task

Safety

This should only be called in LocalRuntime if the runtime has been verified to be owned by the current thread.

pub(crate) fn shutdown(&self)

pub(crate) fn seed_generator(&self) -> &RngSeedGenerator

pub(crate) fn as_current_thread(&self) -> &Arc<Handle>

pub(crate) fn hooks(&self) -> &TaskHooks

impl Handle

pub(crate) fn num_workers(&self) -> usize

pub(crate) fn num_alive_tasks(&self) -> usize

pub(crate) fn injection_queue_depth(&self) -> usize

pub(crate) fn worker_metrics(&self, worker: usize) -> &WorkerMetrics