tokio::runtime::scheduler
Module multi_thread

Multi-threaded runtime

Modules

counters 🔒

handle 🔒

idle 🔒

Coordinates idling workers

overflow 🔒

park 🔒

Parks the runtime.

queue 🔒

Run-queue structures to support a work-stealing scheduler

stats 🔒

trace_mock 🔒

worker 🔒

A scheduler is initialized with a fixed number of workers. Each worker is driven by a thread. Each worker has a “core” which contains data such as the run queue and other state. When block_in_place is called, the worker’s “core” is handed off to a new thread allowing the scheduler to continue to make progress while the originating thread blocks.

Structs

MultiThread 🔒

Work-stealing based thread pool for executing futures.