Struct tokio::runtime::scheduler::multi_thread::queue::Inner

pub(crate) struct Inner<T: 'static> {
    head: AtomicU64,
    tail: AtomicU32,
    buffer: Box<[UnsafeCell<MaybeUninit<Notified<T>>>; 256]>,
}

Fields
head: AtomicU64

Concurrently updated by many threads.

Contains two UnsignedShort values. The LSB byte is the “real” head of the queue. The UnsignedShort in the MSB is set by a stealer in process of stealing values. It represents the first value being stolen in the batch. The UnsignedShort indices are intentionally wider than strictly required for buffer indexing in order to provide ABA mitigation and make it possible to distinguish between full and empty buffers.

When both UnsignedShort values are the same, there is no active stealer.

Tracking an in-progress stealer prevents a wrapping scenario.

tail: AtomicU32

Only updated by producer thread but read by many threads.

buffer: Box<[UnsafeCell<MaybeUninit<Notified<T>>>; 256]>

Elements

Implementations

impl<T> Inner<T>

fn remaining_slots(&self) -> usize

fn len(&self) -> u32

fn is_empty(&self) -> bool