Struct tokio::runtime::TaskExecutor

pub struct TaskExecutor { /* fields omitted */ }

Executes futures on the runtime

All futures spawned using this executor will be submitted to the associated Runtime's executor. This executor is usually a thread pool.

For more details, see the module level documentation.

Methods

impl TaskExecutor

pub fn spawn<F>(&self, future: F) where
F: Future<Item = (), Error = ()> + Send + 'static,

Spawn a future onto the Tokio runtime.

This spawns the given future onto the runtime's executor, usually a thread pool. The thread pool is then responsible for polling the future until it completes.

See module level documentation for more details.

Examples

use tokio::runtime::Runtime;

// Create the runtime
let mut rt = Runtime::new().unwrap();
let executor = rt.executor();

// Spawn a future onto the runtime
executor.spawn(future::lazy(|| {
    println!("now running on a worker thread");
    Ok(())
}));

Panics

This function panics if the spawn fails. Failure occurs if the executor is currently at capacity and is unable to spawn a new future.