Struct tokio::runtime::task::core::Core

#[repr(C)]
pub(super) struct Core<T: Future, S> {
    pub(super) scheduler: S,
    pub(super) task_id: Id,
    pub(super) stage: CoreStage<T>,
}

The core of the task.

Holds the future or output, depending on the stage of execution.

Any changes to the layout of this struct must also be reflected in the const fns in raw.rs.

Fields
scheduler: S

Scheduler used to drive this future.

task_id: Id

The task’s ID, used for populating JoinErrors.

stage: CoreStage<T>

Either the future or the output.

Implementations

impl<T: Future, S: Schedule> Core<T, S>

pub(super) fn poll(&self, cx: Context<'_>) -> Poll<T::Output>

Polls the future.

Safety

The caller must ensure it is safe to mutate the state field. This requires ensuring mutual exclusion between any concurrent thread that might modify the future or output field.

The mutual exclusion is implemented by Harness and the Lifecycle component of the task state.

self must also be pinned. This is handled by storing the task on the heap.

pub(super) fn drop_future_or_output(&self)

Drops the future.

Safety

The caller must ensure it is safe to mutate the stage field.

pub(super) fn store_output(&self, output: Result<T::Output, JoinError>)

Stores the task output.

Safety

The caller must ensure it is safe to mutate the stage field.

pub(super) fn take_output(&self) -> Result<T::Output, JoinError>

Takes the task output.

Safety

The caller must ensure it is safe to mutate the stage field.

unsafe fn set_stage(&self, stage: Stage<T>)