Function tokio::runtime::task::unowned

pub(crate) fn unowned<T, S>(
    task: T,
    scheduler: S,
    id: Id,
) -> (UnownedTask<S>, JoinHandle<T::Output>)
where
    S: Schedule,
    T: Send + Future + 'static,
    T::Output: Send + 'static,

Creates a new task with an associated join handle. This method is used only when the task is not going to be stored in an OwnedTasks list.

Currently only blocking tasks use this method.