Type Alias tokio::runtime::task::list::List

type List<S> = ShardedList<Task<S>, <Task<S> as Link>::Target>;

Aliased Type

struct List<S> {
    lists: Box<[Mutex<LinkedList<Task<S>, Header>>], Global>,
    count: AtomicUsize,
    shard_mask: usize,
}

Fields
lists: Box<[Mutex<LinkedList<Task<S>, Header>>], Global>count: AtomicUsizeshard_mask: usize
Implementations

impl<L, T> ShardedList<L, T>

pub(crate) fn new(sharded_size: usize) -> Self

Creates a new and empty sharded linked list with the specified size.

impl<L: ShardedListItem> ShardedList<L, L::Target>

pub(crate) fn pop_back(&self, shard_id: usize) -> Option<L::Handle>

Removes the last element from a list specified by shard_id and returns it, or None if it is empty.

pub(crate) unsafe fn remove( &self, node: NonNull<L::Target> ) -> Option<L::Handle>

Removes the specified node from the list.

Safety

The caller must ensure that exactly one of the following is true:

node is currently contained by self,

node is not contained by any list,

node is currently contained by some other GuardedLinkedList.

pub(crate) fn lock_shard(&self, val: &L::Handle) -> ShardGuard<'_, L, L::Target>

Gets the lock of ShardedList, makes us have the write permission.

pub(crate) fn len(&self) -> usize

Gets the count of elements in this list.

pub(crate) fn is_empty(&self) -> bool

Returns whether the linked list does not contain any node.

pub(crate) fn shard_size(&self) -> usize

Gets the shard size of this SharedList.

Used to help us to decide the parameter shard_id of the pop_back method.

fn shard_inner( &self, id: usize ) -> MutexGuard<'_, LinkedList<L, <L as Link>::Target>>