tokio::runtime::task
Module raw

Structs

OffsetHelper 🔒

Calling get_trailer_offset directly in vtable doesn’t work because it prevents the vtable from being promoted to a static reference.

RawTask 🔒

Raw task handle

Vtable 🔒

Functions

dealloc 🔒 ⚠

drop_abort_handle 🔒 ⚠

drop_join_handle_slow 🔒 ⚠

get_core_offset 🔒

Compute the offset of the Core<T, S> field in Cell<T, S> using the #[repr(C)] algorithm.

get_id_offset 🔒

Compute the offset of the Id field in Cell<T, S> using the #[repr(C)] algorithm.

get_trailer_offset 🔒

Compute the offset of the Trailer field in Cell<T, S> using the #[repr(C)] algorithm.

poll 🔒 ⚠

schedule 🔒 ⚠

shutdown 🔒 ⚠

try_read_output 🔒 ⚠

vtable 🔒

Get the vtable for the requested T and S generics.