Struct tokio::runtime::task::Task

#[repr(transparent)]pub(crate) struct Task<S: 'static> {
    raw: RawTask,
    _p: PhantomData<S>,
}

An owned handle to the task, tracked by ref count

Fields
raw: RawTask_p: PhantomData<S>
Implementations

impl<S: 'static> Task<S>

pub(crate) unsafe fn from_raw(ptr: NonNull<Header>) -> Task<S>

pub(crate) fn header(&self) -> &Header

impl<S: 'static> Task<S>

pub(crate) fn into_raw(self) -> NonNull<Header>

impl<S: Schedule> Task<S>

pub(crate) fn shutdown(&self)

Pre-emptively cancel the task as part of the shutdown process.