Struct tokio::runtime::thread_pool::worker::Core

struct Core {
    tick: u8,
    lifo_slot: Option<Notified<Arc<Shared>>>,
    run_queue: Local<Arc<Shared>>,
    is_searching: bool,
    is_shutdown: bool,
    park: Option<Parker>,
    stats: WorkerStatsBatcher,
    rand: FastRand,
}

Core data

Fields
tick: u8

Used to schedule bookkeeping tasks every so often.

lifo_slot: Option<Notified<Arc<Shared>>>

When a task is scheduled from a worker, it is stored in this slot. The worker will check this slot for a task before checking the run queue. This effectively results in the last scheduled task to be run next (LIFO). This is an optimization for message passing patterns and helps to reduce latency.

run_queue: Local<Arc<Shared>>

The worker-local run queue.

is_searching: bool

True if the worker is currently searching for more work. Searching involves attempting to steal from other workers.

is_shutdown: bool

True if the scheduler is being shutdown

park: Option<Parker>

Parker

Stored in an Option as the parker is added / removed to make the borrow checker happy.

stats: WorkerStatsBatcher

Batching stats so they can be submitted to RuntimeStats.

rand: FastRand

Fast random number generator.

Implementations

impl Core

fn tick(&mut self)

Increment the tick

fn next_task(&mut self, worker: &Worker) -> Option<Notified<Arc<Shared>>>

Return the next notified task available to this worker.

fn next_local_task(&mut self) -> Option<Notified<Arc<Shared>>>

fn steal_work(&mut self, worker: &Worker) -> Option<Notified<Arc<Shared>>>

fn transition_to_searching(&mut self, worker: &Worker) -> bool

fn transition_from_searching(&mut self, worker: &Worker)

fn transition_to_parked(&mut self, worker: &Worker) -> bool

Prepares the worker state for parking.

Returns true if the transition happend, false if there is work to do first.

fn transition_from_parked(&mut self, worker: &Worker) -> bool

Returns true if the transition happened.

fn maintenance(&mut self, worker: &Worker)

Runs maintenance work such as checking the pool’s state.

fn pre_shutdown(&mut self, worker: &Worker)

Signals all tasks to shut down, and waits for them to complete. Must run before we enter the single-threaded phase of shutdown processing.

fn shutdown(&mut self)

Shuts down the core.