Struct tokio::runtime::thread_pool::worker::Shared

pub(super) struct Shared {
    remotes: Box<[Remote]>,
    inject: Inject<Arc<Worker>>,
    idle: Idle,
    shutdown_workers: Mutex<Vec<(Box<Core>, Arc<Worker>)>>,
}

State shared across all workers

Fields
remotes: Box<[Remote]>

Per-worker remote state. All other workers have access to this and is how they communicate between each other.

inject: Inject<Arc<Worker>>

Submit work to the scheduler while not currently on a worker thread.

idle: Idle

Coordinates idle workers

shutdown_workers: Mutex<Vec<(Box<Core>, Arc<Worker>)>>

Workers have have observed the shutdown signal

The core is not placed back in the worker to avoid it from being stolen by a thread that was spawned as part of block_in_place.

Implementations

impl Shared

pub(super) fn schedule(
&self,
task: Notified<Arc<Worker>>,
is_yield: bool
)

fn schedule_local(
&self,
core: &mut Core,
task: Notified<Arc<Worker>>,
is_yield: bool
)

pub(super) fn close(
&self
)

fn notify_parked(&self)

fn notify_all(&self)

fn notify_if_work_pending(&self)

fn transition_worker_from_searching(&self)

fn shutdown(&self, core: Box<Core>, worker: Arc<Worker>)

Signals that a worker has observed the shutdown signal and has replaced its core back into its handle.

If all workers have reached this point, the final cleanup is performed.

fn ptr_eq(&self, other: &Shared) -> bool