tokio::signal

Module unix

Unix-specific types for signal handling.

This module is only defined on Unix platforms and contains the primary Signal type for receiving notifications of signals.

Structs

OsExtraData 🔒

OsStorage 🔒

Signal

An listener for receiving a particular type of OS signal.

SignalInfo 🔒

SignalKind

Represents the specific kind of signal to listen for.

Functions

action 🔒

Our global signal handler for all signals registered by this module.

ctrl_c 🔒

signal

Creates a new listener which will receive notifications when the current process receives the specified signal kind.

signal_enable 🔒

Enables this module to receive signal notifications for the signal provided.

signal_with_handle 🔒