Function tokio::stream::iter

pub fn iter<I>(i: I) -> Iter<I::IntoIter> where
    I: IntoIterator, 

Converts an Iterator into a Stream which is always ready to yield the next value.

Iterators in Rust don't express the ability to block, so this adapter simply always calls iter.next() and returns that.

use tokio::stream::{self, StreamExt};

let mut stream = stream::iter(vec![17, 19]);

assert_eq!(stream.next().await, Some(17));
assert_eq!(stream.next().await, Some(19));
assert_eq!(stream.next().await, None);