Trait tokio::stream::FromStream

pub trait FromStream<T>: FromStreamPriv<T> { }

Convert from a Stream.

This trait is not intended to be used directly. Instead, call StreamExt::collect().

Implementing

Currently, this trait may not be implemented by third parties. The trait is sealed in order to make changes in the future. Stabilization is pending enhancements to the Rust language.