Struct tokio::sync::batch_semaphore::Semaphore

pub(crate) struct Semaphore {
    waiters: Mutex<Waitlist>,
    permits: AtomicUsize,
}

An asynchronous counting semaphore which permits waiting on multiple permits at once.

Fields
waiters: Mutex<Waitlist>permits: AtomicUsize

The current number of available permits in the semaphore.

Implementations

impl Semaphore

pub(crate) const MAX_PERMITS: usize

The maximum number of permits which a semaphore can hold.

Note that this reserves three bits of flags in the permit counter, but we only actually use one of them. However, the previous semaphore implementation used three bits, so we will continue to reserve them to avoid a breaking change if additional flags need to be added in the future.

const CLOSED: usize

const PERMIT_SHIFT: usize

pub(crate) fn new(permits: usize) -> Self

Creates a new semaphore with the initial number of permits

Maximum number of permits on 32-bit platforms is 1<<29.

pub(crate) fn available_permits(&self) -> usize

Returns the current number of available permits.

pub(crate) fn release(&self, added: usize)

Adds added new permits to the semaphore.

The maximum number of permits is usize::MAX >> 3, and this function will panic if the limit is exceeded.

pub(crate) fn close(&self)

Closes the semaphore. This prevents the semaphore from issuing new permits and notifies all pending waiters.

pub(crate) fn is_closed(&self) -> bool

Returns true if the semaphore is closed.

pub(crate) fn try_acquire(
&self,
num_permits: u32
) -> Result<(), TryAcquireError>

pub(crate) fn acquire(&self, num_permits: u32) -> Acquire<'_>

fn add_permits_locked(&self, rem: usize, waiters: MutexGuard<'_, Waitlist>)

Release rem permits to the semaphore’s wait list, starting from the end of the queue.

If rem exceeds the number of permits needed by the wait list, the remainder are assigned back to the semaphore.

fn poll_acquire(
&self,
cx: &mut Context<'_>,
num_permits: u32,
node: Pin<&mut Waiter>,
queued: bool
) -> Poll<Result<(), AcquireError>>