Module tokio::sync::mpsc::bounded

Structs

OwnedPermit

Owned permit to send one value into the channel.

Permit

Permits to send one value into the channel.

PermitIterator

An Iterator of Permit that can be used to hold n slots in the channel.

Receiver

Receives values from the associated Sender.

Semaphore 🔒

Channel semaphore is a tuple of the semaphore implementation and a usize representing the channel bound.

Sender

Sends values to the associated Receiver.

WeakSender

A sender that does not prevent the channel from being closed.

Functions

channel

Creates a bounded mpsc channel for communicating between asynchronous tasks with backpressure.