Trait tokio::sync::mpsc::chan::Semaphore

pub(crate) trait Semaphore {
    type Permit;
    fn new_permit() -> Self::Permit;

    fn drop_permit(&self, permit: &mut Self::Permit);

    fn is_idle(&self) -> bool;

    fn add_permit(&self);

    fn poll_acquire(
        &self, 
        cx: &mut Context<'_>, 
        permit: &mut Self::Permit
    ) -> Poll<Result<(), ClosedError>>;

    fn try_acquire(&self, permit: &mut Self::Permit) -> Result<(), TrySendError>;

    fn forget(&self, permit: &mut Self::Permit);

    fn close(&self);
}

Associated Types

type Permit

Required methods

fn new_permit() -> Self::Permit

fn drop_permit(&self, permit: &mut Self::Permit)

The permit is dropped without a value being sent. In this case, the permit must be returned to the semaphore.

fn is_idle(&self) -> bool

fn add_permit(&self)

fn poll_acquire(
&self,
cx: &mut Context<'_>,
permit: &mut Self::Permit
) -> Poll<Result<(), ClosedError>>

fn try_acquire(&self, permit: &mut Self::Permit) -> Result<(), TrySendError>

fn forget(&self, permit: &mut Self::Permit)

A value was sent into the channel and the permit held by tx is dropped. In this case, the permit should not immeditely be returned to the semaphore. Instead, the permit is returnred to the semaphore once the sent value is read by the rx handle.

fn close(&self)