Struct tokio::sync::mpsc::list::Rx

pub(crate) struct Rx<T> {
    head: NonNull<Block<T>>,
    index: usize,
    free_head: NonNull<Block<T>>,
}

List queue receive handle

Fields
head: NonNull<Block<T>>

Pointer to the block being processed

index: usize

Next slot index to process

free_head: NonNull<Block<T>>

Pointer to the next block pending release

Implementations

impl<T> Rx<T>

pub(crate) fn pop(&mut self, tx: &Tx<T>) -> Option<Read<T>>

Pops the next value off the queue

fn try_advancing_head(&mut self) -> bool

Tries advancing the block pointer to the block referenced by self.index.

Returns true if successful, false if there is no next block to load.

fn reclaim_blocks(&mut self, tx: &Tx<T>)

pub(super) unsafe fn free_blocks(
&mut self
)

Effectively Drop all the blocks. Should only be called once, when the list is dropping.