tokio::sync::mpsc
Module unbounded

Structs

Semaphore 🔒

No capacity

UnboundedReceiver

Receive values from the associated UnboundedSender.

UnboundedSender

Send values to the associated UnboundedReceiver.

WeakUnboundedSender

An unbounded sender that does not prevent the channel from being closed.

Functions

unbounded_channel

Creates an unbounded mpsc channel for communicating between asynchronous tasks without backpressure.