tokio::sync

Module mutex

Structs

MappedMutexGuard

A handle to a held Mutex that has had a function applied to it via MutexGuard::map.

MappedMutexGuardInner 🔒

A helper type used when taking apart a MappedMutexGuard without running its Drop implementation.

Mutex

An asynchronous Mutex-like type.

MutexGuard

A handle to a held Mutex. The guard can be held across any .await point as it is Send.

MutexGuardInner 🔒

A helper type used when taking apart a MutexGuard without running its Drop implementation.

OwnedMappedMutexGuard

A owned handle to a held Mutex that has had a function applied to it via OwnedMutexGuard::map.

OwnedMappedMutexGuardInner 🔒

A helper type used when taking apart a OwnedMappedMutexGuard without running its Drop implementation.

OwnedMutexGuard

An owned handle to a held Mutex.

OwnedMutexGuardInner 🔒

A helper type used when taking apart a OwnedMutexGuard without running its Drop implementation.

TryLockError

Error returned from the Mutex::try_lock, RwLock::try_read and RwLock::try_write functions.