tokio::sync
Module notify

Structs

AtomicNotification 🔒

Notification for a Waiter. This struct is equivalent to Option<Notification>, but uses AtomicUsize inside for atomic operations.

Notified

Future returned from Notify::notified().

Notify

Notifies a single task to wake up.

NotifyWaitersList 🔒

List used in Notify::notify_waiters. It wraps a guarded linked list and gates the access to it on notify.waiters mutex. It also empties the list on drop.

Waiter 🔒

Enums

Notification 🔒

NotifyOneStrategy 🔒

State 🔒

Constants

EMPTY 🔒

Initial “idle” state.

NOTIFICATION_ALL 🔒

NOTIFICATION_LAST 🔒

NOTIFICATION_NONE 🔒

NOTIFICATION_ONE 🔒

NOTIFIED 🔒

Pending notification.

NOTIFY_WAITERS_CALLS_MASK 🔒

NOTIFY_WAITERS_SHIFT 🔒

STATE_MASK 🔒

WAITING 🔒

One or more threads are currently waiting to be notified.

Functions

atomic_inc_num_notify_waiters_calls 🔒

get_num_notify_waiters_calls 🔒

get_state 🔒

inc_num_notify_waiters_calls 🔒

is_unpin 🔒

notify_locked 🔒

set_state 🔒

Type Aliases

GuardedWaitList 🔒

WaitList 🔒