Module tokio::sync::semaphore_ll

Thread-safe, asynchronous counting semaphore.

A Semaphore instance holds a set of permits. Permits are used to synchronize access to a shared resource.

Before accessing the shared resource, callers acquire a permit from the semaphore. Once the permit is acquired, the caller then enters the critical section. If no permits are available, then acquiring the semaphore returns Pending. The task is woken once a permit becomes available.

Structs

AcquireError
Error returned by Permit::poll_acquire.

Permit
A semaphore permit

SemState
Semaphore state

Semaphore
Futures-aware semaphore.

Waiter
Node used to notify the semaphore waiter when permit is available.

WaiterState
State for an individual waker node

Enums

PermitState
Permit state

TryAcquireError
Error returned by Permit::try_acquire.

Constants

CLOSED
Semaphore has been closed, no more permits will be issued.

CLOSED_FLAG
Signal the semaphore is closed

DROPPED
The permit that owns the Waiter dropped.

MAX_PERMITS
Maximum number of permits a semaphore can manage

NUM_FLAG
Flag differentiating between available permits and waiter pointers.

NUM_SHIFT
When representing "numbers", the state has to be shifted this much (to get rid of the flag bit).

PERMIT_MASK
Masks the waiter state to only contain bits tracking number of requested permits.

PERMIT_ONE
Represents "one requested permit" in the waiter state

PERMIT_SHIFT
How much to shift a permit count to pack it into the waker state

QUEUED
Waiter node is in the semaphore queue

Functions

to_try_acquire