Struct tokio::sync::semaphore_ll::Permit

pub(crate) struct Permit {
    waiter: Option<Box<Waiter>>,
    state: PermitState,
}

A semaphore permit

Tracks the lifecycle of a semaphore permit.

An instance of Permit is intended to be used with a single instance of Semaphore. Using a single instance of Permit with multiple semaphore instances will result in unexpected behavior.

Permit does not release the permit back to the semaphore on drop. It is the user’s responsibility to ensure that Permit::release is called before dropping the permit.

Fields
waiter: Option<Box<Waiter>>state: PermitState
Implementations

impl Permit

pub(crate) fn new() -> Permit

Creates a new Permit.

The permit begins in the “unacquired” state.

pub(crate) fn is_acquired(&self) -> bool

Returns true if the permit has been acquired

pub(crate) fn poll_acquire(
&mut self,
cx: &mut Context<'_>,
num_permits: u16,
semaphore: &Semaphore
) -> Poll<Result<(), AcquireError>>

Tries to acquire the permit. If no permits are available, the current task is notified once a new permit becomes available.

pub(crate) fn try_acquire(
&mut self,
num_permits: u16,
semaphore: &Semaphore
) -> Result<(), TryAcquireError>

Tries to acquire the permit.

pub(crate) fn release(&mut self, n: u16, semaphore: &Semaphore)

Releases a permit back to the semaphore

pub(crate) fn forget(&mut self, n: u16) -> u16

Forgets the permit without releasing it back to the semaphore.

After calling forget, poll_acquire is able to acquire new permit from the sempahore.

Repeatedly calling forget without associated calls to add_permit will result in the semaphore losing all permits.

Will forget at most the number of acquired permits. This number is returned.