Struct tokio::sync::semaphore_ll::Semaphore

pub(crate) struct Semaphore {
    state: AtomicUsize,
    head: UnsafeCell<NonNull<Waiter>>,
    rx_lock: AtomicUsize,
    stub: Box<Waiter>,
}

Futures-aware semaphore.

Fields
state: AtomicUsize

Tracks both the waiter queue tail pointer and the number of remaining permits.

head: UnsafeCell<NonNull<Waiter>>

waiter queue head pointer.

rx_lock: AtomicUsize

Coordinates access to the queue head.

stub: Box<Waiter>

Stub waiter node used as part of the MPSC channel algorithm.

Implementations

impl Semaphore

pub(crate) fn new(permits: usize) -> Semaphore

Creates a new semaphore with the initial number of permits

Panics

Panics if permits is zero.

pub(crate) fn available_permits(&self) -> usize

Returns the current number of available permits

fn poll_acquire(
&self,
cx: &mut Context,
num_permits: u16,
permit: &mut Permit
) -> Poll<Result<(), AcquireError>>

Tries to acquire the requested number of permits, registering the waiter if not enough permits are available.

fn try_acquire(&self, num_permits: u16) -> Result<(), TryAcquireError>

fn poll_acquire2<F>(
&self,
num_permits: u16,
get_waiter: F
) -> Poll<Result<(), AcquireError>> where
F: FnMut() -> Option<NonNull<Waiter>>,

Polls for a permit

Tries to acquire available permits first. If unable to acquire a sufficient number of permits, the caller's waiter is pushed onto the semaphore's wait queue.

pub(crate) fn close(&self)

Closes the semaphore. This prevents the semaphore from issuing new permits and notifies all pending waiters.

pub(crate) fn add_permits(&self, n: usize)

Adds n new permits to the semaphore.

fn add_permits_locked(&self, rem: usize, closed: bool)

fn add_permits_locked2(&self, n: usize, closed: bool)

Releases a specific amount of permits to the semaphore

This function is called by add_permits after the add lock has been acquired.

fn remove_queued(&self, waiter: NonNull<Waiter>, closed: bool)

The wait node has had all of its permits assigned and has been removed from the wait queue.

Attempt to remove the QUEUED bit from the node. If additional permits are concurrently requested, the node must be pushed back into the wait queued.

unsafe fn push_stub(&self, closed: bool)

unsafe fn push_waiter(&self, waiter: NonNull<Waiter>, closed: bool)

fn stub(&self) -> NonNull<Waiter>