Struct tokio::sync::semaphore_ll::Waiter

struct Waiter {
    state: AtomicUsize,
    waker: AtomicWaker,
    next: AtomicPtr<Waiter>,
}

Node used to notify the semaphore waiter when permit is available.

Fields
state: AtomicUsize

Stores waiter state.

See WaiterState for more details.

waker: AtomicWaker

Task to wake when a permit is made available.

next: AtomicPtr<Waiter>

Next pointer in the queue of waiting senders.

Implementations

impl Waiter

fn new() -> Waiter

fn permits_to_acquire(&self) -> Result<usize, AcquireError>

fn try_inc_permits_to_acquire(&self, n: usize) -> bool

Only increments the number of permits if the waiter is currently queued.

Returns

true if the number of permits to acquire has been incremented. false otherwise. On false, the caller should use Semaphore::poll_acquire.

fn try_dec_permits_to_acquire(&self, n: usize) -> usize

Try to decrement the number of permits to acquire. This returns the actual number of permits that were decremented. The delta betweeen n and the return has been assigned to the permit and the caller must assign these back to the semaphore.

fn to_queued(&self, num_permits: usize) -> bool

Store the number of remaining permits needed to satisfy the waiter and transition to the "QUEUED" state.

Returns

true if the QUEUED bit was set as part of the transition.

fn set_permits_to_acquire(&self, num_permits: usize)

Set the number of permits to acquire.

This function is only called when the waiter is being inserted into the wait queue. Because of this, there are no concurrent threads that can modify the state and using store is safe.

fn assign_permits(&self, n: &mut usize, closed: bool) -> bool

Assign permits to the waiter.

Returns true if the waiter should be removed from the queue

fn revert_to_idle(&self)

fn store_next(&self, next: NonNull<Waiter>)