Struct tokio::sync::MappedMutexGuard

pub struct MappedMutexGuard<'a, T: ?Sized> { /* private fields */ }

A handle to a held Mutex that has had a function applied to it via MutexGuard::map.

This can be used to hold a subfield of the protected data.

Implementations

impl<'a, T: ?Sized> MappedMutexGuard<'a, T>

pub fn map<U, F>(this: Self, f: F) -> MappedMutexGuard<'a, U>where
F: FnOnce(&mut T) -> &mut U,

Makes a new MappedMutexGuard for a component of the locked data.

This operation cannot fail as the MappedMutexGuard passed in already locked the mutex.

This is an associated function that needs to be used as MappedMutexGuard::map(...). A method would interfere with methods of the same name on the contents of the locked data.

pub fn try_map<U, F>(this: Self, f: F) -> Result<MappedMutexGuard<'a, U>, Self>where
F: FnOnce(&mut T) -> Option<&mut U>,

Attempts to make a new MappedMutexGuard for a component of the locked data. The original guard is returned if the closure returns None.

This operation cannot fail as the MappedMutexGuard passed in already locked the mutex.

This is an associated function that needs to be used as MappedMutexGuard::try_map(...). A method would interfere with methods of the same name on the contents of the locked data.