tokio::sync

结构体 Mutex

pub struct Mutex<T: ?Sized> { /* private fields */ }

一个异步的、类似 Mutex 的类型。

This type acts similarly to std::sync::Mutex, with two major differences: lock is an async method so does not block, and the lock guard is designed to be held across .await points.

Tokio’s Mutex operates on a guaranteed FIFO basis. This means that the order in which tasks call the lock method is the exact order in which they will acquire the lock.

Which kind of mutex should you use?

Contrary to popular belief, it is ok and often preferred to use the ordinary Mutex from the standard library in asynchronous code.

The feature that the async mutex offers over the blocking mutex is the ability to keep it locked across an .await point. This makes the async mutex more expensive than the blocking mutex, so the blocking mutex should be preferred in the cases where it can be used. The primary use case for the async mutex is to provide shared mutable access to IO resources such as a database connection. If the value behind the mutex is just data, it’s usually appropriate to use a blocking mutex such as the one in the standard library or parking_lot.

Note that, although the compiler will not prevent the std Mutex from holding its guard across .await points in situations where the task is not movable between threads, this virtually never leads to correct concurrent code in practice as it can easily lead to deadlocks.

A common pattern is to wrap the Arc<Mutex<...>> in a struct that provides non-async methods for performing operations on the data within, and only lock the mutex inside these methods. The mini-redis example provides an illustration of this pattern.

Additionally, when you do want shared access to an IO resource, it is often better to spawn a task to manage the IO resource, and to use message passing to communicate with that task.

Examples:

use tokio::sync::Mutex;

use std::sync::Arc;

let data1 = Arc::new(Mutex::new(0));

let data2 = Arc::clone(&data1);

tokio::spawn(async move {

    let mut lock = data2.lock().await;

    *lock += 1;

});

let mut lock = data1.lock().await;

*lock += 1;

use tokio::sync::Mutex;

use std::sync::Arc;

let count = Arc::new(Mutex::new(0));

for i in 0..5 {

    let my_count = Arc::clone(&count);

    tokio::spawn(async move {

        for j in 0..10 {

            let mut lock = my_count.lock().await;

            *lock += 1;

            println!("{} {} {}", i, j, lock);

        }

    });

}

loop {

    if *count.lock().await >= 50 {

        break;

    }

}

println!("Count hit 50.");

There are a few things of note here to pay attention to in this example.

The mutex is wrapped in an Arc to allow it to be shared across threads.

Each spawned task obtains a lock and releases it on every iteration.

Mutation of the data protected by the Mutex is done by de-referencing the obtained lock as seen on lines 13 and 20.

Tokio’s Mutex works in a simple FIFO (first in, first out) style where all calls to lock complete in the order they were performed. In that way the Mutex is “fair” and predictable in how it distributes the locks to inner data. Locks are released and reacquired after every iteration, so basically, each thread goes to the back of the line after it increments the value once. Note that there’s some unpredictability to the timing between when the threads are started, but once they are going they alternate predictably. Finally, since there is only a single valid lock at any given time, there is no possibility of a race condition when mutating the inner value.

Note that in contrast to std::sync::Mutex, this implementation does not poison the mutex when a thread holding the MutexGuard panics. In such a case, the mutex will be unlocked. If the panic is caught, this might leave the data protected by the mutex in an inconsistent state.

实现

impl<T: ?Sized> Mutex<T>

pub fn new(t: T) -> Self
where T: Sized,

创建一个新的锁，初始状态为未锁定，可直接使用。

示例

use tokio::sync::Mutex;

let lock = Mutex::new(5);

pub const fn const_new(t: T) -> Self
where T: Sized,

创建一个新的锁，初始状态为未锁定，可直接使用。

使用 tracing 不稳定特性时，通过 const_new 创建的 Mutex 不会被插桩。因此，它不会出现在 tokio-console 中。如有需要，请改用 Mutex::new 来创建可插桩的对象。

示例

use tokio::sync::Mutex;

static LOCK: Mutex<i32> = Mutex::const_new(5);

pub async fn lock(&self) -> MutexGuard<'_, T>

锁定此互斥锁，使当前任务挂起直到获取到锁。获取到锁后，函数返回一个 MutexGuard。

如果互斥锁可立即获取，则此调用通常不会让出运行时。但在所有情况下这都不能保证。

Cancel safety

此方法使用队列按请求顺序公平分配锁。取消对 lock 的调用会丢失在队列中的位置。

示例

use tokio::sync::Mutex;

let mutex = Mutex::new(1);

let mut n = mutex.lock().await;

*n = 2;

pub fn blocking_lock(&self) -> MutexGuard<'_, T>

阻塞地锁定此 Mutex。获取到锁后，函数返回一个 MutexGuard。

此方法用于需要在异步代码和同步代码中都使用此互斥锁的场景。

Panics

如果在异步执行上下文中调用此函数会触发 panic。

If you find yourself in an asynchronous execution context and needing to call some (synchronous) function which performs one of these blocking_ operations, then consider wrapping that call inside spawn_blocking() (or block_in_place()).

示例

use std::sync::Arc;

use tokio::sync::Mutex;

#[tokio::main]

async fn main() {

    let mutex =  Arc::new(Mutex::new(1));

    let lock = mutex.lock().await;

    let mutex1 = Arc::clone(&mutex);

    let blocking_task = tokio::task::spawn_blocking(move || {

        // This shall block until the `lock` is released.

        let mut n = mutex1.blocking_lock();

        *n = 2;

    });

    assert_eq!(*lock, 1);

    // Release the lock.

    drop(lock);

    // Await the completion of the blocking task.

    blocking_task.await.unwrap();

    // Assert uncontended.

    let n = mutex.try_lock().unwrap();

    assert_eq!(*n, 2);

}

pub fn blocking_lock_owned(self: Arc<Self>) -> OwnedMutexGuard<T>

阻塞地锁定此 Mutex。获取到锁后，函数返回一个 OwnedMutexGuard。

此方法与 Mutex::blocking_lock 相同，只是返回的 guard 通过 Arc 而非借用引用 Mutex。因此，调用此方法时 Mutex 必须包装在 Arc 中，且 guard 将在 'static 生命周期内有效，因为它通过持有 Arc 保持 Mutex 存活。

Panics

如果在异步执行上下文中调用此函数会触发 panic。

If you find yourself in an asynchronous execution context and needing to call some (synchronous) function which performs one of these blocking_ operations, then consider wrapping that call inside spawn_blocking() (or block_in_place()).

示例

use std::sync::Arc;

use tokio::sync::Mutex;

#[tokio::main]

async fn main() {

    let mutex =  Arc::new(Mutex::new(1));

    let lock = mutex.lock().await;

    let mutex1 = Arc::clone(&mutex);

    let blocking_task = tokio::task::spawn_blocking(move || {

        // This shall block until the `lock` is released.

        let mut n = mutex1.blocking_lock_owned();

        *n = 2;

    });

    assert_eq!(*lock, 1);

    // Release the lock.

    drop(lock);

    // Await the completion of the blocking task.

    blocking_task.await.unwrap();

    // Assert uncontended.

    let n = mutex.try_lock().unwrap();

    assert_eq!(*n, 2);

}

pub async fn lock_owned(self: Arc<Self>) -> OwnedMutexGuard<T>

锁定此互斥锁，使当前任务挂起直到获取到锁。获取到锁后，返回一个 OwnedMutexGuard。

如果互斥锁可立即获取，则此调用通常不会让出运行时。但在所有情况下这都不能保证。

此方法与 Mutex::lock 相同，只是返回的 guard 通过 Arc 而非借用引用 Mutex。因此，调用此方法时 Mutex 必须包装在 Arc 中，且 guard 将在 'static 生命周期内有效，因为它通过持有 Arc 保持 Mutex 存活。

Cancel safety

此方法使用队列按请求顺序公平分配锁。取消对 lock_owned 的调用会丢失在队列中的位置。

示例

use tokio::sync::Mutex;

use std::sync::Arc;

let mutex = Arc::new(Mutex::new(1));

let mut n = mutex.clone().lock_owned().await;

*n = 2;

pub fn try_lock(&self) -> Result<MutexGuard<'_, T>, TryLockError>

尝试获取锁，如果锁当前被其他位置持有则返回 TryLockError。

示例

use tokio::sync::Mutex;

let mutex = Mutex::new(1);

let n = mutex.try_lock()?;

assert_eq!(*n, 1);

pub fn get_mut(&mut self) -> &mut T

返回对底层数据的可变引用。

由于此调用可变地借用 Mutex，不需要实际执行加锁 —— 可变借用静态保证不存在任何锁。

示例

use tokio::sync::Mutex;

fn main() {

    let mut mutex = Mutex::new(1);

    let n = mutex.get_mut();

    *n = 2;

}

pub fn try_lock_owned( self: Arc<Self>, ) -> Result<OwnedMutexGuard<T>, TryLockError>

尝试获取锁，如果锁当前被其他位置持有则返回 TryLockError。

此方法与 Mutex::try_lock 相同，只是返回的 guard 通过 Arc 而非借用引用 Mutex。因此，调用此方法时 Mutex 必须包装在 Arc 中，且 guard 将在 'static 生命周期内有效，因为它通过持有 Arc 保持 Mutex 存活。

示例

use tokio::sync::Mutex;

use std::sync::Arc;

let mutex = Arc::new(Mutex::new(1));

let n = mutex.clone().try_lock_owned()?;

assert_eq!(*n, 1);

pub fn into_inner(self) -> T
where T: Sized,

消耗互斥锁，返回底层数据。

示例

use tokio::sync::Mutex;

let mutex = Mutex::new(1);

let n = mutex.into_inner();

assert_eq!(n, 1);