Struct tokio::sync::OwnedMutexGuard

pub struct OwnedMutexGuard<T: ?Sized> { /* fields omitted */ }

An owned handle to a held Mutex.

This guard is only available from a Mutex that is wrapped in an Arc. It is identical to MutexGuard, except that rather than borrowing the Mutex, it clones the Arc, incrementing the reference count. This means that unlike MutexGuard, it will have the 'static lifetime.

As long as you have this guard, you have exclusive access to the underlying T. The guard internally keeps a reference-couned pointer to the original Mutex, so even if the lock goes away, the guard remains valid.

The lock is automatically released whenever the guard is dropped, at which point lock will succeed yet again.