Struct tokio::task::futures::TaskLocalFuture

pub struct TaskLocalFuture<T, F>where
    T: 'static,{ /* private fields */ }

A future that sets a value T of a task local for the future F during its execution.

The value of the task-local must be 'static and will be dropped on the completion of the future.

Created by the function LocalKey::scope.

Examples

tokio::task_local! {
    static NUMBER: u32;
}

NUMBER.scope(1, async move {
    println!("task local value: {}", NUMBER.get());
}).await;