tokio::task

Module local

Runs !Send futures on the current thread.

Structs

Context 🔒

State available from the thread-local.

LocalData 🔒

LocalDataEnterGuard 🔒

A guard for LocalData::enter()

LocalEnterGuard

Context guard for LocalSet

LocalSet

A set of tasks which are executed on the same thread.

LocalState 🔒

Tracks the LocalSet state that must only be accessed from the thread that created the LocalSet.

RunUntil 🔒

Shared 🔒

LocalSet state shared between threads.

Constants

CURRENT 🔒

INITIAL_CAPACITY 🔒

Initial queue capacity.

MAX_TASKS_PER_TICK 🔒

Max number of tasks to poll per tick.

REMOTE_FIRST_INTERVAL 🔒

How often it check the remote queue first.

Functions

spawn_local

Spawns a !Send future on the current LocalSet or LocalRuntime.

spawn_local_inner 🔒