Struct tokio::time::driver::entry::TimerShared

pub(crate) struct TimerShared {
    state: StateCell,
    driver_state: CachePadded<TimerSharedPadded>,
    _p: PhantomPinned,
}

The shared state structure of a timer. This structure is shared between the frontend (Entry) and driver backend.

Note that this structure is located inside the TimerEntry structure.

Fields
state: StateCell

Current state. This records whether the timer entry is currently under the ownership of the driver, and if not, its current state (not complete, fired, error, etc).

driver_state: CachePadded<TimerSharedPadded>

Data manipulated by the driver thread itself, only.

_p: PhantomPinned
Implementations

impl TimerShared

pub(super) fn new() -> Self

pub(super) fn cached_when(&self) -> u64

Gets the cached time-of-expiration value.

pub(super) unsafe fn sync_when(&self) -> u64

Gets the true time-of-expiration value, and copies it into the cached time-of-expiration value.

SAFETY: Must be called with the driver lock held, and when this entry is not in any timer wheel lists.

unsafe fn set_cached_when(&self, when: u64)

Sets the cached time-of-expiration value.

SAFETY: Must be called with the driver lock held, and when this entry is not in any timer wheel lists.

pub(super) fn true_when(&self) -> u64

Returns the true time-of-expiration value, with relaxed memory ordering.

pub(super) unsafe fn set_expiration(&self, t: u64)

Sets the true time-of-expiration value, even if it is less than the current expiration or the timer is deregistered.

SAFETY: Must only be called with the driver lock held and the entry not in the timer wheel.

pub(super) fn extend_expiration(&self, t: u64) -> Result<(), ()>

Sets the true time-of-expiration only if it is after the current.

pub(super) fn handle(&self) -> TimerHandle

Returns a TimerHandle for this timer.

pub(super) fn might_be_registered(&self) -> bool

Returns true if the state of this timer indicates that the timer might be registered with the driver. This check is performed with relaxed ordering, but is conservative - if it returns false, the timer is definitely not registered.