Function tokio::time::delay_until

pub fn delay_until(deadline: Instant) -> Delay

Waits until deadline is reached.

No work is performed while awaiting on the delay to complete. The delay operates at millisecond granularity and should not be used for tasks that require high-resolution timers.

Cancellation

Canceling a delay is done by dropping the returned future. No additional cleanup work is required.