Function tokio::time::resume

pub fn resume()

Resume time

Clears the saved Instant::now() value. Subsequent calls to Instant::now() will return the value returned by the system call.

Panics

Panics if time is not frozen or if called from outside of the Tokio runtime.