tokio::time
Struct Instant

pub struct Instant { /* private fields */ }

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

Instants are always guaranteed to be no less than any previously measured instant when created, and are often useful for tasks such as measuring benchmarks or timing how long an operation takes.

Note, however, that instants are not guaranteed to be steady. In other words, each tick of the underlying clock may not be the same length (e.g. some seconds may be longer than others). An instant may jump forwards or experience time dilation (slow down or speed up), but it will never go backwards.

Instants are opaque types that can only be compared to one another. There is no method to get “the number of seconds” from an instant. Instead, it only allows measuring the duration between two instants (or comparing two instants).

The size of an Instant struct may vary depending on the target operating system.

Note

This type wraps the inner std variant and is used to align the Tokio clock for uses of now(). This can be useful for testing where you can take advantage of time::pause() and time::advance().

Implementations

impl Instant

pub fn now() -> Instant

Returns an instant corresponding to “now”.

Examples

use tokio::time::Instant;

let now = Instant::now();

pub fn from_std(std: Instant) -> Instant

Create a tokio::time::Instant from a std::time::Instant.

pub fn into_std(self) -> Instant

Convert the value into a std::time::Instant.

pub fn duration_since(&self, earlier: Instant) -> Duration

Returns the amount of time elapsed from another instant to this one, or zero duration if that instant is later than this one.

pub fn checked_duration_since(&self, earlier: Instant) -> Option<Duration>

Returns the amount of time elapsed from another instant to this one, or None if that instant is later than this one.

Examples

use tokio::time::{Duration, Instant, sleep};

#[tokio::main]
async fn main() {
    let now = Instant::now();
    sleep(Duration::new(1, 0)).await;
    let new_now = Instant::now();
    println!("{:?}", new_now.checked_duration_since(now));
    println!("{:?}", now.checked_duration_since(new_now)); // None
}

pub fn saturating_duration_since(&self, earlier: Instant) -> Duration

Returns the amount of time elapsed from another instant to this one, or zero duration if that instant is later than this one.

Examples

use tokio::time::{Duration, Instant, sleep};

#[tokio::main]
async fn main() {
    let now = Instant::now();
    sleep(Duration::new(1, 0)).await;
    let new_now = Instant::now();
    println!("{:?}", new_now.saturating_duration_since(now));
    println!("{:?}", now.saturating_duration_since(new_now)); // 0ns
}

pub fn elapsed(&self) -> Duration

Returns the amount of time elapsed since this instant was created, or zero duration if this instant is in the future.

Examples

use tokio::time::{Duration, Instant, sleep};

#[tokio::main]
async fn main() {
    let instant = Instant::now();
    let three_secs = Duration::from_secs(3);
    sleep(three_secs).await;
    assert!(instant.elapsed() >= three_secs);
}

pub fn checked_add(&self, duration: Duration) -> Option<Instant>

Returns Some(t) where t is the time self + duration if t can be represented as Instant (which means it’s inside the bounds of the underlying data structure), None otherwise.

pub fn checked_sub(&self, duration: Duration) -> Option<Instant>

Returns Some(t) where t is the time self - duration if t can be represented as Instant (which means it’s inside the bounds of the underlying data structure), None otherwise.