Struct tokio::time::Timeout

#[must_use = "futures do nothing unless you `.await` or poll them"]pub struct Timeout<T> { /* fields omitted */ }

Future returned by timeout and timeout_at.

Implementations

impl<T> Timeout<T>

pub fn get_ref(&self) -> &T

Gets a reference to the underlying value in this timeout.

pub fn get_mut(&mut self) -> &mut T

Gets a mutable reference to the underlying value in this timeout.

pub fn into_inner(self) -> T

Consumes this timeout, returning the underlying value.