tokio::time

Module timeout

Allows a future to execute for a maximum amount of time.

See Timeout documentation for more details.

Structs

Timeout

Future returned by timeout and timeout_at.

Functions

poll_delay 🔒

timeout

Requires a Future to complete before the specified duration has elapsed.

timeout_at

Requires a Future to complete before the specified instant in time.