tokio::util::idle_notified_set
Function move_to_new_list

unsafe fn move_to_new_list<T>(
    from: &mut LinkedList<ListEntry<T>, <ListEntry<T> as Link>::Target>,
    to: &mut LinkedList<ListEntry<T>, <ListEntry<T> as Link>::Target>,
)

Safety

The mutex for the entries must be held, and the target list must be such that setting my_list to Neither is ok.