Type Alias tokio::util::idle_notified_set::LinkedList

type LinkedList<T> = LinkedList<ListEntry<T>, <ListEntry<T> as Link>::Target>;

Aliased Type

struct LinkedList<T> {
    head: Option<NonNull<ListEntry<T>>>,
    tail: Option<NonNull<ListEntry<T>>>,
    _marker: PhantomData<*const ListEntry<T>>,
}

Fields
head: Option<NonNull<ListEntry<T>>>

Linked list head

tail: Option<NonNull<ListEntry<T>>>

Linked list tail

_marker: PhantomData<*const ListEntry<T>>

Node type marker.

Implementations

impl<L, T> LinkedList<L, T>

pub(crate) const fn new() -> LinkedList<L, T>

Creates an empty linked list.

impl<L: Link> LinkedList<L, L::Target>

pub(crate) fn push_front(&mut self, val: L::Handle)

Adds an element first in the list.

pub(crate) fn pop_back(&mut self) -> Option<L::Handle>

Removes the last element from a list and returns it, or None if it is empty.

pub(crate) fn is_empty(&self) -> bool

Returns whether the linked list does not contain any node

pub(crate) unsafe fn remove( &mut self, node: NonNull<L::Target> ) -> Option<L::Handle>

Removes the specified node from the list

Safety

The caller must ensure that exactly one of the following is true:

node is currently contained by self,

node is not contained by any list,

node is currently contained by some other GuardedLinkedList and the caller has an exclusive access to that list. This condition is used by the linked list in sync::Notify.

impl<L: Link> LinkedList<L, L::Target>

pub(crate) fn last(&self) -> Option<&L::Target>

impl<T: Link> LinkedList<T, T::Target>

pub(crate) fn drain_filter<F>(&mut self, filter: F) -> DrainFilter<'_, T, F> where F: FnMut(&T::Target) -> bool,

impl<U, L: Link<Handle = NonNull<U>>> LinkedList<L, L::Target>

pub(crate) fn into_guarded( self, guard_handle: L::Handle ) -> GuardedLinkedList<L, L::Target>

Turns a linked list into the guarded version by linking the guard node with the head and tail nodes. Like with other nodes, you should guarantee that the guard node is pinned in memory.