Struct tokio::util::linked_list::GuardedLinkedList

pub(crate) struct GuardedLinkedList<L, T> {
    guard: NonNull<T>,
    _marker: PhantomData<*const L>,
}

An intrusive linked list, but instead of keeping pointers to the head and tail nodes, it uses a special guard node linked with those nodes. It means that the list is circular and every pointer of a node from the list is not None, including pointers from the guard node.

If a list is empty, then both pointers of the guard node are pointing at the guard node itself.

Fields
guard: NonNull<T>

Pointer to the guard node.

_marker: PhantomData<*const L>

Node type marker.

Implementations

impl<L: Link> GuardedLinkedList<L, L::Target>

fn tail(&self) -> Option<NonNull<L::Target>>

pub(crate) fn pop_back(&mut self) -> Option<L::Handle>

Removes the last element from a list and returns it, or None if it is empty.