tokio::util::sharded_list

Struct ShardedList

pub(crate) struct ShardedList<L: ShardedListItem> {
    lists: Box<[Mutex<LinkedList<L>>]>,
    added: MetricAtomicU64,
    count: MetricAtomicUsize,
    shard_mask: usize,
}

An intrusive linked list supporting highly concurrent updates.

It currently relies on LinkedList, so it is the caller’s responsibility to ensure the list is empty before dropping it.

Note: Due to its inner sharded design, the order of nodes cannot be guaranteed.

Fields
lists: Box<[Mutex<LinkedList<L>>]>added: MetricAtomicU64count: MetricAtomicUsizeshard_mask: usize
Implementations

impl<L: ShardedListItem> ShardedList<L>

pub(crate) fn new(sharded_size: usize) -> Self

Creates a new and empty sharded linked list with the specified size.

pub(crate) fn pop_back(&self, shard_id: usize) -> Option<L::Handle>

Removes the last element from a list specified by shard_id and returns it, or None if it is empty.

pub(crate) unsafe fn remove( &self, node: NonNull<L::Target>, ) -> Option<L::Handle>

Removes the specified node from the list.

Safety

The caller must ensure that exactly one of the following is true:

node is currently contained by self,

node is not contained by any list,

node is currently contained by some other GuardedLinkedList.

pub(crate) fn lock_shard(&self, val: &L::Handle) -> ShardGuard<'_, L>

Gets the lock of ShardedList, makes us have the write permission.

pub(crate) fn len(&self) -> usize

Gets the count of elements in this list.

pub(crate) fn is_empty(&self) -> bool

Returns whether the linked list does not contain any node.

pub(crate) fn shard_size(&self) -> usize

Gets the shard size of this SharedList.

Used to help us to decide the parameter shard_id of the pop_back method.

fn shard_inner(&self, id: usize) -> MutexGuard<'_, LinkedList<L>>