tokio::util::sharded_list
Trait ShardedListItem

pub(crate) unsafe trait ShardedListItem: Link {
    // Required method
    unsafe fn get_shard_id(target: NonNull<Self::Target>) -> usize;
}

Determines which linked list an item should be stored in.

Safety

Implementations must guarantee that the id of an item does not change from call to call.

Required Methods

unsafe fn get_shard_id(target: NonNull<Self::Target>) -> usize

Safety

The provided pointer must point at a valid list item.

Object Safety

This trait is not object safe.