Module tokio::util::slab

Structs

Address

References a slot in the slab. Indexing a slot using an Address is memory safe even if the slot has been released or the page has been deallocated. However, it is not guaranteed that the slot has not been reused and is now represents a different value.

Allocator

Allocate values in the associated slab.

CachedPage

Page

A page in the slab.

Ref

A reference to a value stored in the slab.

Slab

Amortized allocation for homogeneous data types.

Slot

A slot in the slab. Contains slot-specific metadata.

Slots

Page state.

Value

Value paired with a reference to the page.

Constants

NUM_PAGES

Maximum number of pages a slab can contain.

PAGE_INDEX_SHIFT

PAGE_INITIAL_SIZE

Minimum number of slots a page can contain.

Traits

Entry

An entry in the slab.

Functions

debug