Module tokio::util::slab::addr

Tracks the location of an entry in a slab.

Index packing

A slab index consists of multiple indices packed into a single usize value that correspond to different parts of the slab.

The least significant MAX_PAGES + INITIAL_PAGE_SIZE.trailing_zeros() + 1 bits store the address within a shard, starting at 0 for the first slot on the first page. To index a slot within a shard, we first find the index of the page that the address falls on, and then the offset of the slot within that page.

Since every page is twice as large as the previous page, and all page sizes are powers of two, we can determine the page index that contains a given address by shifting the address down by the smallest page size and looking at how many twos places necessary to represent that number, telling us what power of two page size it fits inside of. We can determine the number of twos places by counting the number of leading zeros (unused twos places) in the number's binary representation, and subtracting that count from the total number of bits in a word.

Once we know what page contains an address, we can subtract the size of all previous pages from the address to determine the offset within the page.

After the page address, the next MAX_THREADS.trailing_zeros() + 1 least significant bits are the thread ID. These are used to index the array of shards to find which shard a slot belongs to. If an entry is being removed and the thread ID of its index matches that of the current thread, we can use the remove_local fast path; otherwise, we have to use the synchronized remove_remote path.

Finally, a generation value is packed into the index. The RESERVED_BITS most significant bits are left unused, and the remaining bits between the last bit of the thread ID and the first reserved bit are used to store the generation. The generation is used as part of an atomic read-modify-write loop every time a ScheduledIo's readiness is modified, or when the resource is removed, to guard against the ABA problem.

Visualized:

    ┌──────────┬───────────────┬──────────────────┬──────────────────────────┐
    │ reserved │  generation   │    thread ID     │         address          │
    └▲─────────┴▲──────────────┴▲─────────────────┴▲────────────────────────▲┘
     │          │               │                  │                        │
bits(usize)     │       bits(MAX_THREADS)          │                        0
                │                                  │
     bits(usize) - RESERVED       MAX_PAGES + bits(INITIAL_PAGE_SIZE)

Structs

Address
References the location at which an entry is stored in a slab.

Constants

GENERATION
Masks the generation

PAGE_INDEX_SHIFT

RESERVED

SLOT
Address in the shard

THREAD
Masks the thread identifier