tokio::util
Module sync_wrapper

This module contains a type that can make Send + !Sync types Sync by disallowing all immutable access to the value.

A similar primitive is provided in the sync_wrapper crate.

Structs

SyncWrapper 🔒