Crate tokio

A runtime for writing reliable, asynchronous, and slim applications.

Tokio is an event-driven, non-blocking I/O platform for writing asynchronous applications with the Rust programming language. At a high level, it provides a few major components:

Tools for working with asynchronous tasks, including synchronization primitives and channels and timeouts, delays, and intervals.

APIs for performing asynchronous I/O, including TCP and UDP sockets, filesystem operations, and process and signal management.

A runtime for executing asynchronous code, including a task scheduler, an I/O driver backed by the operating system's event queue (epoll, kqueue, IOCP, etc...), and a high performance timer.

Guide level documentation is found on the website.

A Tour of Tokio

Tokio consists of a number of modules that provide a range of functionality essential for implementing asynchronous applications in Rust. In this section, we will take a brief tour of Tokio, summarizing the major APIs and their uses.

Note that Tokio uses Cargo feature flags to allow users to control what features are present, so that unused code can be eliminated. This documentation also lists what feature flags are necessary to enable each API.

The easiest way to get started is to enable all features. Do this by enabling the full feature flag:

tokio = { version = "0.2", features = ["full"] }

Working With Tasks

Asynchronous programs in Rust are based around lightweight, non-blocking units of execution called tasks. The tokio::task module provides important tools for working with tasks:

The spawn function and JoinHandle type, for scheduling a new task on the Tokio runtime and awaiting the output of a spawned task, respectively,

Functions for running blocking operations in an asynchronous task context.

The tokio::task module is present only when the "rt-core" feature flag is enabled.

The tokio::sync module contains synchronization primitives to use when need to communicate or share data. These include:

channels (oneshot, mpsc, and watch), for sending values between tasks,

a non-blocking Mutex, for controlling access to a shared, mutable value,

an asynchronous Barrier type, for multiple tasks to synchronize before beginning a computation.

The tokio::sync module is present only when the "sync" feature flag is enabled.

The tokio::time module provides utilities for tracking time and scheduling work. This includes functions for setting timeouts for tasks, delaying work to run in the future, or repeating an operation at an interval.

In order to use tokio::time, the "time" feature flag must be enabled.

Finally, Tokio provides a runtime for executing asynchronous tasks. Most applications can use the #[tokio::main] macro to run their code on the Tokio runtime. In use-cases where manual control over the runtime is required, the tokio::runtime module provides APIs for configuring and managing runtimes.

Using the runtime requires the "rt-core" or "rt-threaded" feature flags, to enable the basic single-threaded scheduler and the thread-pool scheduler, respectively. See the runtime module documentation for details. In addition, the "macros" feature flag enables the #[tokio::main] and #[tokio::test] attributes.

Asynchronous IO

As well as scheduling and running tasks, Tokio provides everything you need to perform input and output asynchronously.

The tokio::io module provides Tokio's asynchronous core I/O primitives, the AsyncRead, AsyncWrite, and AsyncBufRead traits. In addition, when the "io-util" feature flag is enabled, it also provides combinators and functions for working with these traits, forming as an asynchronous counterpart to std::io. When the "io-driver" feature flag is enabled, it also provides utilities for library authors implementing I/O resources.

Tokio also includes APIs for performing various kinds of I/O and interacting with the operating system asynchronously. These include:

tokio::net, which contains non-blocking versions of TCP, UDP, and Unix Domain Sockets (enabled by the "net" feature flag),

tokio::fs, similar to std::fs but for performing filesystem I/O asynchronously (enabled by the "fs" feature flag),

tokio::signal, for asynchronously handling Unix and Windows OS signals (enabled by the "signal" feature flag),

tokio::process, for spawning and managing child processes (enabled by the "process" feature flag).

Examples

A simple TCP echo server:

use tokio::net::TcpListener;
use tokio::prelude::*;

#[tokio::main]
async fn main() -> Result<(), Box<dyn std::error::Error>> {
    let mut listener = TcpListener::bind("127.0.0.1:8080").await?;

    loop {
        let (mut socket, _) = listener.accept().await?;

        tokio::spawn(async move {
            let mut buf = [0; 1024];

            // In a loop, read data from the socket and write the data back.
            loop {
                let n = match socket.read(&mut buf).await {
                    // socket closed
                    Ok(n) if n == 0 => return,
                    Ok(n) => n,
                    Err(e) => {
                        println!("failed to read from socket; err = {:?}", e);
                        return;
                    }
                };

                // Write the data back
                if let Err(e) = socket.write_all(&buf[0..n]).await {
                    println!("failed to write to socket; err = {:?}", e);
                    return;
                }
            }
        });
    }
}

Modules

io
Traits, helpers, and type definitions for asynchronous I/O functionality.

net
TCP/UDP/Unix bindings for tokio.

prelude
A "prelude" for users of the tokio crate.

runtime
The Tokio runtime.

task
Asynchronous green-threads.

Functions

spawn
Spawns a new asynchronous task, returning a JoinHandle] for it.