Module tokio::coop

Opt-in yield points for improved cooperative scheduling.

A single call to poll on a top-level task may potentially do a lot of work before it returns Poll::Pending. If a task runs for a long period of time without yielding back to the executor, it can starve other tasks waiting on that executor to execute them, or drive underlying resources. Since Rust does not have a runtime, it is difficult to forcibly preempt a long-running task. Instead, this module provides an opt-in mechanism for futures to collaborate with the executor to avoid starvation.

Consider a future like this one:

async fn drop_all<I: Stream + Unpin>(mut input: I) {
    while let Some(_) = input.next().await {}
}

It may look harmless, but consider what happens under heavy load if the input stream is always ready. If we spawn drop_all, the task will never yield, and will starve other tasks and resources on the same executor. With opt-in yield points, this problem is alleviated:

async fn drop_all<I: Stream + Unpin>(mut input: I) {
    while let Some(_) = input.next().await {
        tokio::coop::proceed().await;
    }
}

The proceed future will coordinate with the executor to make sure that every so often control is yielded back to the executor so it can run other tasks.

Placing yield points

Voluntary yield points should be placed after at least some work has been done. If they are not, a future sufficiently deep in the task hierarchy may end up never getting to run because of the number of yield points that inevitably appear before it is reached. In general, you will want yield points to only appear in “leaf” futures – those that do not themselves poll other futures. By doing this, you avoid double-counting each iteration of the outer future against the cooperating budget.

Structs

Budget

Opaque type tracking the amount of “work” a task may still do before yielding back to the scheduler.

Constants

CURRENT

Functions

budget

Run the given closure with a cooperative task budget. When the function returns, the budget is reset to the value prior to calling the function.

has_budget_remaining

poll_proceed

Returns Poll::Pending if the current task has exceeded its budget and should yield.

stop

Forcibly remove the budgeting constraints early.

with_budget