Module tokio::fs

Asynchronous filesystem manipulation operations.

This module contains basic methods and types for manipulating the contents of the local filesystem from within the context of the Tokio runtime.

Unlike most other Tokio APIs, the filesystem APIs must be used from the context of the Tokio runtime as they require Tokio specific features to function.

Modules

file

Types for working with File.

os

OS-specific functionality.

Structs

File

A reference to an open file on the filesystem.

OpenOptions

Options and flags which can be used to configure how a file is opened.

ReadFile

A future used to open a file and read its entire contents into a buffer.

WriteFile

A future used to open a file for writing and write the entire contents of some data to it.

Functions

create_dir

Creates a new, empty directory at the provided path

create_dir_all

Recursively create a directory and all of its parent components if they are missing.

hard_link

Creates a new hard link on the filesystem.

metadata

Queries the file system metadata for a path.

read

Creates a future which will open a file for reading and read the entire contents into a buffer and return said buffer.

read_dir

Returns a stream over the entries within a directory.

read_link

Reads a symbolic link, returning the file that the link points to.

remove_dir

Removes an existing, empty directory.

remove_file

Removes a file from the filesystem.

rename

Rename a file or directory to a new name, replacing the original file if to already exists.

set_permissions

Changes the permissions found on a file or a directory.

symlink_metadata

Queries the file system metadata for a path.

write

Creates a future that will open a file for writing and write the entire contents of contents to it.