Function tokio::fs::write

pub async fn write<C: AsRef<[u8]> + Unpin>(
    path: impl AsRef<Path>, 
    contents: C
) -> Result<()>

Creates a future that will open a file for writing and write the entire contents of contents to it.

This is the async equivalent of std::fs::write.

Examples

use tokio::fs;

fs::write("foo.txt", b"Hello world!").await?;