Struct tokio::fs::File

pub struct File { /* private fields */ }

A reference to an open file on the filesystem.

This is a specialized version of std::fs::File for usage from the Tokio runtime.

An instance of a File can be read and/or written depending on what options it was opened with. Files also implement AsyncSeek to alter the logical cursor that the file contains internally.

A file will not be closed immediately when it goes out of scope if there are any IO operations that have not yet completed. To ensure that a file is closed immediately when it is dropped, you should call flush before dropping it. Note that this does not ensure that the file has been fully written to disk; the operating system might keep the changes around in an in-memory buffer. See the sync_all method for telling the OS to write the data to disk.

Reading and writing to a File is usually done using the convenience methods found on the AsyncReadExt and AsyncWriteExt traits.

Examples

Create a new file and asynchronously write bytes to it:

use tokio::fs::File;
use tokio::io::AsyncWriteExt; // for write_all()

let mut file = File::create("foo.txt").await?;
file.write_all(b"hello, world!").await?;

Read the contents of a file into a buffer:

use tokio::fs::File;
use tokio::io::AsyncReadExt; // for read_to_end()

let mut file = File::open("foo.txt").await?;

let mut contents = vec![];
file.read_to_end(&mut contents).await?;

println!("len = {}", contents.len());

Implementations

impl File

pub async fn open(path: impl AsRef<Path>) -> Result<File>

Attempts to open a file in read-only mode.

See OpenOptions for more details.

Errors

This function will return an error if called from outside of the Tokio runtime or if path does not already exist. Other errors may also be returned according to OpenOptions::open.

Examples

use tokio::fs::File;
use tokio::io::AsyncReadExt;

let mut file = File::open("foo.txt").await?;

let mut contents = vec![];
file.read_to_end(&mut contents).await?;

println!("len = {}", contents.len());

The read_to_end method is defined on the AsyncReadExt trait.

pub async fn create(path: impl AsRef<Path>) -> Result<File>

Opens a file in write-only mode.

This function will create a file if it does not exist, and will truncate it if it does.

See OpenOptions for more details.

Errors

Results in an error if called from outside of the Tokio runtime or if the underlying create call results in an error.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::create("foo.txt").await?;
file.write_all(b"hello, world!").await?;

The write_all method is defined on the AsyncWriteExt trait.

pub async fn create_new<P: AsRef<Path>>(path: P) -> Result<File>

Opens a file in read-write mode.

This function will create a file if it does not exist, or return an error if it does. This way, if the call succeeds, the file returned is guaranteed to be new.

This option is useful because it is atomic. Otherwise between checking whether a file exists and creating a new one, the file may have been created by another process (a TOCTOU race condition / attack).

This can also be written using File::options().read(true).write(true).create_new(true).open(...).

See OpenOptions for more details.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::create_new("foo.txt").await?;
file.write_all(b"hello, world!").await?;

The write_all method is defined on the AsyncWriteExt trait.

pub fn options() -> OpenOptions

Returns a new OpenOptions object.

This function returns a new OpenOptions object that you can use to open or create a file with specific options if open() or create() are not appropriate.

It is equivalent to OpenOptions::new(), but allows you to write more readable code. Instead of OpenOptions::new().append(true).open("example.log"), you can write File::options().append(true).open("example.log"). This also avoids the need to import OpenOptions.

See the OpenOptions::new function for more details.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut f = File::options().append(true).open("example.log").await?;
f.write_all(b"new line\n").await?;

pub fn from_std(std: StdFile) -> File

Converts a std::fs::File to a tokio::fs::File.

Examples

// This line could block. It is not recommended to do this on the Tokio
// runtime.
let std_file = std::fs::File::open("foo.txt").unwrap();
let file = tokio::fs::File::from_std(std_file);

pub async fn sync_all(&self) -> Result<()>

Attempts to sync all OS-internal metadata to disk.

This function will attempt to ensure that all in-core data reaches the filesystem before returning.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::create("foo.txt").await?;
file.write_all(b"hello, world!").await?;
file.sync_all().await?;

The write_all method is defined on the AsyncWriteExt trait.

pub async fn sync_data(&self) -> Result<()>

This function is similar to sync_all, except that it may not synchronize file metadata to the filesystem.

This is intended for use cases that must synchronize content, but don’t need the metadata on disk. The goal of this method is to reduce disk operations.

Note that some platforms may simply implement this in terms of sync_all.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::create("foo.txt").await?;
file.write_all(b"hello, world!").await?;
file.sync_data().await?;

The write_all method is defined on the AsyncWriteExt trait.

pub async fn set_len(&self, size: u64) -> Result<()>

Truncates or extends the underlying file, updating the size of this file to become size.

If the size is less than the current file’s size, then the file will be shrunk. If it is greater than the current file’s size, then the file will be extended to size and have all of the intermediate data filled in with 0s.

Errors

This function will return an error if the file is not opened for writing.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::create("foo.txt").await?;
file.write_all(b"hello, world!").await?;
file.set_len(10).await?;

The write_all method is defined on the AsyncWriteExt trait.

pub async fn metadata(&self) -> Result<Metadata>

Queries metadata about the underlying file.

Examples

use tokio::fs::File;

let file = File::open("foo.txt").await?;
let metadata = file.metadata().await?;

println!("{:?}", metadata);

pub async fn try_clone(&self) -> Result<File>

Creates a new File instance that shares the same underlying file handle as the existing File instance. Reads, writes, and seeks will affect both File instances simultaneously.

Examples

use tokio::fs::File;

let file = File::open("foo.txt").await?;
let file_clone = file.try_clone().await?;

pub async fn into_std(self) -> StdFile

Destructures File into a std::fs::File. This function is async to allow any in-flight operations to complete.

Use File::try_into_std to attempt conversion immediately.

Examples

use tokio::fs::File;

let tokio_file = File::open("foo.txt").await?;
let std_file = tokio_file.into_std().await;

pub fn try_into_std(self) -> Result<StdFile, Self>

Tries to immediately destructure File into a std::fs::File.

Errors

This function will return an error containing the file if some operation is in-flight.

Examples

use tokio::fs::File;

let tokio_file = File::open("foo.txt").await?;
let std_file = tokio_file.try_into_std().unwrap();

pub async fn set_permissions(&self, perm: Permissions) -> Result<()>

Changes the permissions on the underlying file.

Platform-specific behavior

This function currently corresponds to the fchmod function on Unix and the SetFileInformationByHandle function on Windows. Note that, this may change in the future.

Errors

This function will return an error if the user lacks permission change attributes on the underlying file. It may also return an error in other os-specific unspecified cases.

Examples

use tokio::fs::File;

let file = File::open("foo.txt").await?;
let mut perms = file.metadata().await?.permissions();
perms.set_readonly(true);
file.set_permissions(perms).await?;

pub fn set_max_buf_size(&mut self, max_buf_size: usize)

Set the maximum buffer size for the underlying AsyncRead / AsyncWrite operation.

Although Tokio uses a sensible default value for this buffer size, this function would be useful for changing that default depending on the situation.

Examples

use tokio::fs::File;
use tokio::io::AsyncWriteExt;

let mut file = File::open("foo.txt").await?;

// Set maximum buffer size to 8 MiB
file.set_max_buf_size(8 * 1024 * 1024);

let mut buf = vec![1; 1024 * 1024 * 1024];

// Write the 1 GiB buffer in chunks up to 8 MiB each.
file.write_all(&mut buf).await?;