Struct tokio::fs::OpenOptions

pub struct OpenOptions(_);

Options and flags which can be used to configure how a file is opened.

This builder exposes the ability to configure how a File is opened and what operations are permitted on the open file. The File::open and File::create methods are aliases for commonly used options using this builder.

Generally speaking, when using OpenOptions, you’ll first call new, then chain calls to methods to set each option, then call open, passing the path of the file you’re trying to open. This will give you a io::Result with a File inside that you can further operate on.

This is a specialized version of std::fs::OpenOptions for usage from the Tokio runtime.

From<std::fs::OpenOptions> is implemented for more advanced configuration than the methods provided here.

Examples

Opening a file to read:

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .read(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

Opening a file for both reading and writing, as well as creating it if it doesn’t exist:

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .read(true)
        .write(true)
        .create(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

Implementations

impl OpenOptions

pub fn new() -> OpenOptions

Creates a blank new set of options ready for configuration.

All options are initially set to false.

This is an async version of std::fs::OpenOptions::new

Examples

use tokio::fs::OpenOptions;

let mut options = OpenOptions::new();
let future = options.read(true).open("foo.txt");

pub fn read(&mut self, read: bool) -> &mut OpenOptions

Sets the option for read access.

This option, when true, will indicate that the file should be read-able if opened.

This is an async version of std::fs::OpenOptions::read

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .read(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

pub fn write(&mut self, write: bool) -> &mut OpenOptions

Sets the option for write access.

This option, when true, will indicate that the file should be write-able if opened.

This is an async version of std::fs::OpenOptions::write

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .write(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

pub fn append(&mut self, append: bool) -> &mut OpenOptions

Sets the option for the append mode.

This option, when true, means that writes will append to a file instead of overwriting previous contents. Note that setting .write(true).append(true) has the same effect as setting only .append(true).

For most filesystems, the operating system guarantees that all writes are atomic: no writes get mangled because another process writes at the same time.

One maybe obvious note when using append-mode: make sure that all data that belongs together is written to the file in one operation. This can be done by concatenating strings before passing them to write(), or using a buffered writer (with a buffer of adequate size), and calling flush() when the message is complete.

If a file is opened with both read and append access, beware that after opening, and after every write, the position for reading may be set at the end of the file. So, before writing, save the current position (using seek(SeekFrom::Current(0))), and restore it before the next read.

This is an async version of std::fs::OpenOptions::append

Note

This function doesn’t create the file if it doesn’t exist. Use the create method to do so.

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .append(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

pub fn truncate(&mut self, truncate: bool) -> &mut OpenOptions

Sets the option for truncating a previous file.

If a file is successfully opened with this option set it will truncate the file to 0 length if it already exists.

The file must be opened with write access for truncate to work.

This is an async version of std::fs::OpenOptions::truncate

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .write(true)
        .truncate(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

pub fn create(&mut self, create: bool) -> &mut OpenOptions

Sets the option for creating a new file.

This option indicates whether a new file will be created if the file does not yet already exist.

In order for the file to be created, write or append access must be used.

This is an async version of std::fs::OpenOptions::create

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .write(true)
        .create(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

pub fn create_new(&mut self, create_new: bool) -> &mut OpenOptions

Sets the option to always create a new file.

This option indicates whether a new file will be created. No file is allowed to exist at the target location, also no (dangling) symlink.

This option is useful because it is atomic. Otherwise between checking whether a file exists and creating a new one, the file may have been created by another process (a TOCTOU race condition / attack).

If .create_new(true) is set, .create() and .truncate() are ignored.

The file must be opened with write or append access in order to create a new file.

This is an async version of std::fs::OpenOptions::create_new

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new()
        .write(true)
        .create_new(true)
        .open("foo.txt")
        .await?;

    Ok(())
}

pub async fn open(&self, path: impl AsRef<Path>) -> Result<File>

Opens a file at path with the options specified by self.

This is an async version of std::fs::OpenOptions::open

Errors

This function will return an error under a number of different circumstances. Some of these error conditions are listed here, together with their ErrorKind. The mapping to ErrorKinds is not part of the compatibility contract of the function, especially the Other kind might change to more specific kinds in the future.

NotFound: The specified file does not exist and neither create or create_new is set.

NotFound: One of the directory components of the file path does not exist.

PermissionDenied: The user lacks permission to get the specified access rights for the file.

PermissionDenied: The user lacks permission to open one of the directory components of the specified path.

AlreadyExists: create_new was specified and the file already exists.

InvalidInput: Invalid combinations of open options (truncate without write access, no access mode set, etc.).

Other: One of the directory components of the specified file path was not, in fact, a directory.

Other: Filesystem-level errors: full disk, write permission requested on a read-only file system, exceeded disk quota, too many open files, too long filename, too many symbolic links in the specified path (Unix-like systems only), etc.

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let file = OpenOptions::new().open("foo.txt").await?;
    Ok(())
}

impl OpenOptions

pub fn mode(&mut self, mode: u32) -> &mut OpenOptions

Sets the mode bits that a new file will be created with.

If a new file is created as part of an OpenOptions::open call then this specified mode will be used as the permission bits for the new file. If no mode is set, the default of 0o666 will be used. The operating system masks out bits with the system’s umask, to produce the final permissions.

Examples

use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut options = OpenOptions::new();
    options.mode(0o644); // Give read/write for owner and read for others.
    let file = options.open("foo.txt").await?;

    Ok(())
}

pub fn custom_flags(&mut self, flags: i32) -> &mut OpenOptions

Passes custom flags to the flags argument of open.

The bits that define the access mode are masked out with O_ACCMODE, to ensure they do not interfere with the access mode set by Rusts options.

Custom flags can only set flags, not remove flags set by Rusts options. This options overwrites any previously set custom flags.

Examples

use libc;
use tokio::fs::OpenOptions;
use std::io;

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut options = OpenOptions::new();
    options.write(true);
    if cfg!(unix) {
        options.custom_flags(libc::O_NOFOLLOW);
    }
    let file = options.open("foo.txt").await?;

    Ok(())
}