Function tokio::io::copy_bidirectional_with_sizes

pub async fn copy_bidirectional_with_sizes<A, B>(
    a: &mut A,
    b: &mut B,
    a_to_b_buf_size: usize,
    b_to_a_buf_size: usize,
) -> Result<(u64, u64)>
where
    A: AsyncRead + AsyncWrite + Unpin + ?Sized,
    B: AsyncRead + AsyncWrite + Unpin + ?Sized,

Copies data in both directions between a and b using buffers of the specified size.

This method is the same as the copy_bidirectional(), except that it allows you to set the size of the internal buffers used when copying data.