Struct tokio::io::BufStream

pub struct BufStream<RW> { /* fields omitted */ }

Wraps a type that is AsyncWrite and AsyncRead, and buffers its input and output.

It can be excessively inefficient to work directly with something that implements AsyncWrite and AsyncRead. For example, every write, however small, has to traverse the syscall interface, and similarly, every read has to do the same. The BufWriter and BufReader types aid with these problems respectively, but do so in only one direction. BufStream wraps one in the other so that both directions are buffered. See their documentation for details.

Implementations

impl<RW: AsyncRead + AsyncWrite> BufStream<RW>

pub fn new(stream: RW) -> BufStream<RW>

Wraps a type in both BufWriter and BufReader.

See the documentation for those types and BufStream for details.

pub fn with_capacity(
reader_capacity: usize,
writer_capacity: usize,
stream: RW
) -> BufStream<RW>

Creates a BufStream with the specified BufReader capacity and BufWriter capacity.

See the documentation for those types and BufStream for details.

pub fn get_ref(&self) -> &RW

Gets a reference to the underlying I/O object.

It is inadvisable to directly read from the underlying I/O object.

pub fn get_mut(&mut self) -> &mut RW

Gets a mutable reference to the underlying I/O object.

It is inadvisable to directly read from the underlying I/O object.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut RW>

Gets a pinned mutable reference to the underlying I/O object.

It is inadvisable to directly read from the underlying I/O object.

pub fn into_inner(self) -> RW

Consumes this BufStream, returning the underlying I/O object.

Note that any leftover data in the internal buffer is lost.