Struct tokio::io::BufWriter

pub struct BufWriter<W> { /* private fields */ }

Wraps a writer and buffers its output.

It can be excessively inefficient to work directly with something that implements AsyncWrite. A BufWriter keeps an in-memory buffer of data and writes it to an underlying writer in large, infrequent batches.

BufWriter can improve the speed of programs that make small and repeated write calls to the same file or network socket. It does not help when writing very large amounts at once, or writing just one or a few times. It also provides no advantage when writing to a destination that is in memory, like a Vec<u8>.

When the BufWriter is dropped, the contents of its buffer will be discarded. Creating multiple instances of a BufWriter on the same stream can cause data loss. If you need to write out the contents of its buffer, you must manually call flush before the writer is dropped.

Implementations

impl<W: AsyncWrite> BufWriter<W>

pub fn new(inner: W) -> Self

Creates a new BufWriter with a default buffer capacity. The default is currently 8 KB, but may change in the future.

pub fn with_capacity(cap: usize, inner: W) -> Self

Creates a new BufWriter with the specified buffer capacity.

pub fn get_ref(&self) -> &W

Gets a reference to the underlying writer.

pub fn get_mut(&mut self) -> &mut W

Gets a mutable reference to the underlying writer.

It is inadvisable to directly write to the underlying writer.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut W>

Gets a pinned mutable reference to the underlying writer.

It is inadvisable to directly write to the underlying writer.

pub fn into_inner(self) -> W

Consumes this BufWriter, returning the underlying writer.

Note that any leftover data in the internal buffer is lost.

pub fn buffer(&self) -> &[u8]

Returns a reference to the internally buffered data.