Struct tokio::io::ReadBuf

pub struct ReadBuf<'a> { /* fields omitted */ }

A wrapper around a byte buffer that is incrementally filled and initialized.

This type is a sort of "double cursor". It tracks three regions in the buffer: a region at the beginning of the buffer that has been logically filled with data, a region that has been initialized at some point but not yet logically filled, and a region at the end that may be uninitialized. The filled region is guaranteed to be a subset of the initialized region.

In summary, the contents of the buffer can be visualized as:

[             capacity              ]
[ filled |         unfilled         ]
[    initialized    | uninitialized ]

It is undefined behavior to de-initialize any bytes from the uninitialized region, since it is merely unknown whether this region is uninitialized or not, and if part of it turns out to be initialized, it must stay initialized.

Implementations

impl<'a> ReadBuf<'a>

pub fn new(buf: &'a mut [u8]) -> ReadBuf<'a>

Creates a new ReadBuf from a fully initialized buffer.

pub fn uninit(buf: &'a mut [MaybeUninit<u8>]) -> ReadBuf<'a>

Creates a new ReadBuf from a fully uninitialized buffer.

Use assume_init if part of the buffer is known to be already inintialized.

pub fn capacity(&self) -> usize

Returns the total capacity of the buffer.

pub fn filled(&self) -> &[u8]

Returns a shared reference to the filled portion of the buffer.

pub fn filled_mut(&mut self) -> &mut [u8]

Returns a mutable reference to the filled portion of the buffer.

pub fn take(&mut self, n: usize) -> ReadBuf<'_>

Returns a new ReadBuf comprised of the unfilled section up to n.

pub fn initialized(&self) -> &[u8]

Returns a shared reference to the initialized portion of the buffer.

This includes the filled portion.

pub fn initialized_mut(&mut self) -> &mut [u8]

Returns a mutable reference to the initialized portion of the buffer.

This includes the filled portion.

pub unsafe fn unfilled_mut(&mut self) -> &mut [MaybeUninit<u8>]

Returns a mutable reference to the unfilled part of the buffer without ensuring that it has been fully initialized.

Safety

The caller must not de-initialize portions of the buffer that have already been initialized. This includes any bytes in the region marked as uninitialized by ReadBuf.

pub fn initialize_unfilled(&mut self) -> &mut [u8]

Returns a mutable reference to the unfilled part of the buffer, ensuring it is fully initialized.

Since ReadBuf tracks the region of the buffer that has been initialized, this is effectively "free" after the first use.

pub fn initialize_unfilled_to(&mut self, n: usize) -> &mut [u8]

Returns a mutable reference to the first n bytes of the unfilled part of the buffer, ensuring it is fully initialized.

Panics

Panics if self.remaining() is less than n.

pub fn remaining(&self) -> usize

Returns the number of bytes at the end of the slice that have not yet been filled.

pub fn clear(&mut self)

Clears the buffer, resetting the filled region to empty.

The number of initialized bytes is not changed, and the contents of the buffer are not modified.

pub fn advance(&mut self, n: usize)

Advances the size of the filled region of the buffer.

The number of initialized bytes is not changed.

Panics

Panics if the filled region of the buffer would become larger than the initialized region.

pub fn set_filled(&mut self, n: usize)

Sets the size of the filled region of the buffer.

The number of initialized bytes is not changed.

Note that this can be used to shrink the filled region of the buffer in addition to growing it (for example, by a AsyncRead implementation that compresses data in-place).

Panics

Panics if the filled region of the buffer would become larger than the intialized region.

pub unsafe fn assume_init(&mut self, n: usize)

Asserts that the first n unfilled bytes of the buffer are initialized.

ReadBuf assumes that bytes are never de-initialized, so this method does nothing when called with fewer bytes than are already known to be initialized.

Safety

The caller must ensure that n unfilled bytes of the buffer have already been initialized.

pub fn put_slice(&mut self, buf: &[u8])

Appends data to the buffer, advancing the written position and possibly also the initialized position.

Panics

Panics if self.remaining() is less than buf.len().