Struct tokio::io::Take

#[must_use = "streams do nothing unless you `.await` or poll them"]pub struct Take<R> { /* fields omitted */ }

Stream for the take method.

Implementations

impl<R: AsyncRead> Take<R>

pub fn limit(&self) -> u64

Returns the remaining number of bytes that can be read before this instance will return EOF.

Note

This instance may reach EOF after reading fewer bytes than indicated by this method if the underlying [AsyncRead] instance reaches EOF.

pub fn set_limit(&mut self, limit: u64)

Sets the number of bytes that can be read before this instance will return EOF. This is the same as constructing a new Take instance, so the amount of bytes read and the previous limit value don't matter when calling this method.

pub fn get_ref(&self) -> &R

Gets a reference to the underlying reader.

pub fn get_mut(&mut self) -> &mut R

Gets a mutable reference to the underlying reader.

Care should be taken to avoid modifying the internal I/O state of the underlying reader as doing so may corrupt the internal limit of this Take.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut R>

Gets a pinned mutable reference to the underlying reader.

Care should be taken to avoid modifying the internal I/O state of the underlying reader as doing so may corrupt the internal limit of this Take.

pub fn into_inner(self) -> R

Consumes the Take, returning the wrapped reader.