Trait tokio::io::AsyncSeek

pub trait AsyncSeek {
    fn start_seek(
        self: Pin<&mut Self>, 
        cx: &mut Context, 
        position: SeekFrom
    ) -> Poll<Result<()>>;

    fn poll_complete(
        self: Pin<&mut Self>, 
        cx: &mut Context
    ) -> Poll<Result<u64>>;
}

Seek bytes asynchronously.

This trait is analogous to the std::io::Seek trait, but integrates with the asynchronous task system. In particular, the start_seek method, unlike Seek::seek, will not block the calling thread.

Utilities for working with AsyncSeek values are provided by AsyncSeekExt.

Required methods

fn start_seek(
self: Pin<&mut Self>,
cx: &mut Context,
position: SeekFrom
) -> Poll<Result<()>>

Attempts to seek to an offset, in bytes, in a stream.

A seek beyond the end of a stream is allowed, but behavior is defined by the implementation.

If this function returns successfully, then the job has been submitted. To find out when it completes, call poll_complete.

fn poll_complete(self: Pin<&mut Self>, cx: &mut Context) -> Poll<Result<u64>>

Waits for a seek operation to complete.

If the seek operation completed successfully, this method returns the new position from the start of the stream. That position can be used later with [SeekFrom::Start].

Errors

Seeking to a negative offset is considered an error.

Panics

Calling this method without calling start_seek first is an error.

Loading content...