Trait tokio::io::util::vec_with_initialized::VecU8

pub(crate) unsafe trait VecU8: AsRef<Vec<u8>> + AsMut<Vec<u8>> { }

Something that looks like a Vec<u8>.

Safety

The implementor must guarantee that the vector returned by the as_mut and as_mut methods do not change from one call to another.