tokio::loom::std

Module sync

Modules

atomic 🔒

Structs

Arc 🔒

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

Condvar 🔒

A Condition Variable

MutexGuard 🔒

An RAII implementation of a “scoped lock” of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

RwLockReadGuard 🔒

RAII structure used to release the shared read access of a lock when dropped.

WaitTimeoutResult 🔒

A type indicating whether a timed wait on a condition variable returned due to a time out or not.

Weak 🔒

Weak is a version of Arc that holds a non-owning reference to the managed allocation.