tokio

Module net

TCP/UDP/Unix bindings for tokio.

This module contains the TCP/UDP/Unix networking types, similar to the standard library, which can be used to implement networking protocols.

Organization

TcpListener and TcpStream provide functionality for communication over TCP

UdpSocket provides functionality for communication over UDP

UnixListener and UnixStream provide functionality for communication over a Unix Domain Stream Socket (available on Unix only)

UnixDatagram provides functionality for communication over Unix Domain Datagram Socket (available on Unix only)

tokio::net::unix::pipe for FIFO pipes (available on Unix only)

tokio::net::windows::named_pipe for Named Pipes (available on Windows only)

For IO resources not available in tokio::net, you can use AsyncFd.

Modules

tcpnet

TCP utility types.

unixUnix and net

Unix specific network types.

windowsWindows and net

Windows specific network types.

Structs

TcpListenernet

A TCP socket server, listening for connections.

TcpSocketnet

A TCP socket that has not yet been converted to a TcpStream or TcpListener.

TcpStreamnet

A TCP stream between a local and a remote socket.

UdpSocketnet

A UDP socket.

UnixDatagramUnix and net

An I/O object representing a Unix datagram socket.

UnixListenerUnix and net

A Unix socket which can accept connections from other Unix sockets.

UnixSocketUnix and net

A Unix socket that has not yet been converted to a UnixStream, UnixDatagram, or UnixListener.

UnixStreamUnix and net

A structure representing a connected Unix socket.

Traits

ToSocketAddrs

Converts or resolves without blocking to one or more SocketAddr values.

Functions

lookup_hostnet

Performs a DNS resolution.