Module tokio::net

TCP/UDP/Unix bindings for tokio.

This module contains the TCP/UDP/Unix networking types, similar to the standard library, which can be used to implement networking protocols.

Organization

TcpListener and TcpStream provide functionality for communication over TCP

UdpSocket provides functionality for communication over UDP

UnixListener and UnixStream provide functionality for communication over a Unix Domain Stream Socket (available on Unix only)

UnixDatagram provides functionality for communication over Unix Domain Datagram Socket (available on Unix only)

Modules

tcp
TCP utility types

udp
UDP utility types.

unix
Unix domain socket utility types

Structs

TcpListener
A TCP socket server, listening for connections.

TcpStream
A TCP stream between a local and a remote socket.

UdpSocket
A UDP socket

UnixDatagram
An I/O object representing a Unix datagram socket.

UnixListener
A Unix socket which can accept connections from other Unix sockets.

UnixStream
A structure representing a connected Unix socket.

Traits

ToSocketAddrs
Converts or resolves without blocking to one or more SocketAddr values.

Functions

lookup_host
Performs a DNS resolution.