Module tokio::net

TCP/UDP/Unix bindings for tokio.

This module contains the TCP/UDP/Unix networking types, similar to the standard library, which can be used to implement networking protocols.

Organization

TcpListener and TcpStream provide functionality for communication over TCP

UdpSocket provides functionality for communication over UDP

UnixListener and UnixStream provide functionality for communication over a Unix Domain Stream Socket (available on Unix only)

UnixDatagram provides functionality for communication over Unix Domain Datagram Socket (available on Unix only)

Traits

ToSocketAddrs

Converts or resolves without blocking to one or more SocketAddr values.