Module tokio::net

TCP/UDP/Unix bindings for tokio.

This module contains the TCP/UDP/Unix networking types, similar to the standard library, which can be used to implement networking protocols.

Organization

TcpListener and TcpStream provide functionality for communication over TCP

UdpSocket and UdpFramed provide functionality for communication over UDP

UnixListener and UnixStream provide functionality for communication over a Unix Domain Socket (available on Unix only)

Modules

tcp
TCP bindings for tokio.

udp
UDP bindings for tokio.

unix
Unix domain socket bindings for tokio (only available on unix systems).

Structs

TcpListener
An I/O object representing a TCP socket listening for incoming connections.

TcpStream
An I/O object representing a TCP stream connected to a remote endpoint.

UdpFramed
A unified Stream and Sink interface to an underlying UdpSocket, using the Encoder and Decoder traits to encode and decode frames.

UdpSocket
An I/O object representing a UDP socket.

UnixListener
A Unix socket which can accept connections from other Unix sockets.

UnixStream
A structure representing a connected Unix socket.