Struct tokio::net::UdpSocket

pub struct UdpSocket { /* fields omitted */ }

A UDP socket

Implementations

impl UdpSocket

pub async fn bind<A: ToSocketAddrs>(addr: A) -> Result<UdpSocket>

This function will create a new UDP socket and attempt to bind it to the addr provided.

pub fn from_std(socket: UdpSocket) -> Result<UdpSocket>

Creates a new UdpSocket from the previously bound socket provided.

The socket given will be registered with the event loop that handle is associated with. This function requires that socket has previously been bound to an address to work correctly.

This can be used in conjunction with net2's UdpBuilder interface to configure a socket before it's handed off, such as setting options like reuse_address or binding to multiple addresses.

Panics

This function panics if thread-local runtime is not set.

The runtime is usually set implicitly when this function is called from a future driven by a tokio runtime, otherwise runtime can be set explicitly with Handle::enter function.

pub fn split(self) -> (RecvHalf, SendHalf)

Splits the UdpSocket into a receive half and a send half. The two parts can be used to receive and send datagrams concurrently, even from two different tasks.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this socket is bound to.

pub async fn connect<'_, A: ToSocketAddrs>(&'_ self, addr: A) -> Result<()>

Connects the UDP socket setting the default destination for send() and limiting packets that are read via recv from the address specified in addr.

pub async fn send<'_, '_>(&'_ mut self, buf: &'_ [u8]) -> Result<usize>

Returns a future that sends data on the socket to the remote address to which it is connected. On success, the future will resolve to the number of bytes written.

The connect method will connect this socket to a remote address. The future will resolve to an error if the socket is not connected.

pub async fn recv<'_, '_>(&'_ mut self, buf: &'_ mut [u8]) -> Result<usize>

Returns a future that receives a single datagram message on the socket from the remote address to which it is connected. On success, the future will resolve to the number of bytes read.

The function must be called with valid byte array buf of sufficient size to hold the message bytes. If a message is too long to fit in the supplied buffer, excess bytes may be discarded.

The connect method will connect this socket to a remote address. The future will fail if the socket is not connected.

pub async fn send_to<'_, '_, A: ToSocketAddrs>(
&'_ mut self,
buf: &'_ [u8],
target: A
) -> Result<usize>

Returns a future that sends data on the socket to the given address. On success, the future will resolve to the number of bytes written.

The future will resolve to an error if the IP version of the socket does not match that of target.

pub async fn recv_from<'_, '_>(
&'_ mut self,
buf: &'_ mut [u8]
) -> Result<(usize, SocketAddr)>

Returns a future that receives a single datagram on the socket. On success, the future resolves to the number of bytes read and the origin.

The function must be called with valid byte array buf of sufficient size to hold the message bytes. If a message is too long to fit in the supplied buffer, excess bytes may be discarded.

pub fn broadcast(&self) -> Result<bool>

Gets the value of the SO_BROADCAST option for this socket.

For more information about this option, see set_broadcast.

pub fn set_broadcast(&self, on: bool) -> Result<()>

Sets the value of the SO_BROADCAST option for this socket.

When enabled, this socket is allowed to send packets to a broadcast address.

pub fn multicast_loop_v4(&self) -> Result<bool>

Gets the value of the IP_MULTICAST_LOOP option for this socket.

For more information about this option, see set_multicast_loop_v4.

pub fn set_multicast_loop_v4(&self, on: bool) -> Result<()>

Sets the value of the IP_MULTICAST_LOOP option for this socket.

If enabled, multicast packets will be looped back to the local socket.

Note

This may not have any affect on IPv6 sockets.

pub fn multicast_ttl_v4(&self) -> Result<u32>

Gets the value of the IP_MULTICAST_TTL option for this socket.

For more information about this option, see set_multicast_ttl_v4.

pub fn set_multicast_ttl_v4(&self, ttl: u32) -> Result<()>

Sets the value of the IP_MULTICAST_TTL option for this socket.

Indicates the time-to-live value of outgoing multicast packets for this socket. The default value is 1 which means that multicast packets don't leave the local network unless explicitly requested.

Note

This may not have any affect on IPv6 sockets.

pub fn multicast_loop_v6(&self) -> Result<bool>

Gets the value of the IPV6_MULTICAST_LOOP option for this socket.

For more information about this option, see set_multicast_loop_v6.

pub fn set_multicast_loop_v6(&self, on: bool) -> Result<()>

Sets the value of the IPV6_MULTICAST_LOOP option for this socket.

Controls whether this socket sees the multicast packets it sends itself.

Note

This may not have any affect on IPv4 sockets.

pub fn ttl(&self) -> Result<u32>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

pub fn set_ttl(&self, ttl: u32) -> Result<()>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

pub fn join_multicast_v4(
&self,
multiaddr: Ipv4Addr,
interface: Ipv4Addr
) -> Result<()>

Executes an operation of the IP_ADD_MEMBERSHIP type.

This function specifies a new multicast group for this socket to join. The address must be a valid multicast address, and interface is the address of the local interface with which the system should join the multicast group. If it's equal to INADDR_ANY then an appropriate interface is chosen by the system.

pub fn join_multicast_v6(
&self,
multiaddr: &Ipv6Addr,
interface: u32
) -> Result<()>

Executes an operation of the IPV6_ADD_MEMBERSHIP type.

This function specifies a new multicast group for this socket to join. The address must be a valid multicast address, and interface is the index of the interface to join/leave (or 0 to indicate any interface).

pub fn leave_multicast_v4(
&self,
multiaddr: Ipv4Addr,
interface: Ipv4Addr
) -> Result<()>

Executes an operation of the IP_DROP_MEMBERSHIP type.

For more information about this option, see join_multicast_v4.

pub fn leave_multicast_v6(
&self,
multiaddr: &Ipv6Addr,
interface: u32
) -> Result<()>

Executes an operation of the IPV6_DROP_MEMBERSHIP type.

For more information about this option, see join_multicast_v6.