Struct tokio::net::UnixDatagram

pub struct UnixDatagram { /* fields omitted */ }

An I/O object representing a Unix datagram socket.

Implementations

impl UnixDatagram

pub fn bind<P>(path: P) -> Result<UnixDatagram> where
P: AsRef<Path>,

Creates a new UnixDatagram bound to the specified path.

pub fn pair() -> Result<(UnixDatagram, UnixDatagram)>

Creates an unnamed pair of connected sockets.

This function will create a pair of interconnected Unix sockets for communicating back and forth between one another. Each socket will be associated with the default event loop's handle.

pub fn from_std(datagram: UnixDatagram) -> Result<UnixDatagram>

Consumes a UnixDatagram in the standard library and returns a nonblocking UnixDatagram from this crate.

The returned datagram will be associated with the given event loop specified by handle and is ready to perform I/O.

Panics

This function panics if thread-local runtime is not set.

The runtime is usually set implicitly when this function is called from a future driven by a tokio runtime, otherwise runtime can be set explicitly with Handle::enter function.

pub fn unbound() -> Result<UnixDatagram>

Creates a new UnixDatagram which is not bound to any address.

pub fn connect<P: AsRef<Path>>(&self, path: P) -> Result<()>

Connects the socket to the specified address.

The send method may be used to send data to the specified address. recv and recv_from will only receive data from that address.

pub async fn send<'_, '_>(&'_ mut self, buf: &'_ [u8]) -> Result<usize>

Sends data on the socket to the socket's peer.

pub async fn recv<'_, '_>(&'_ mut self, buf: &'_ mut [u8]) -> Result<usize>

Receives data from the socket.

pub async fn send_to<'_, '_, P>(
&'_ mut self,
buf: &'_ [u8],
target: P
) -> Result<usize> where
P: AsRef<Path> + Unpin,

Sends data on the socket to the specified address.

pub async fn recv_from<'_, '_>(
&'_ mut self,
buf: &'_ mut [u8]
) -> Result<(usize, SocketAddr)>

Receives data from the socket.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this socket is bound to.

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the address of this socket's peer.

The connect method will connect the socket to a peer.

pub fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).

pub fn into_split(self) -> (OwnedRecvHalf, OwnedSendHalf)

Split a UnixDatagram into a receive half and a send half, which can be used to receice and send the datagram concurrently.