Module tokio::net::tcp

TCP utility types.

Structs

OwnedReadHalf

Owned read half of a TcpStream, created by into_split.

OwnedWriteHalf

Owned write half of a TcpStream, created by into_split.

ReadHalf

Borrowed read half of a TcpStream, created by split.

ReuniteError

Error indicating that two halves were not from the same socket, and thus could not be reunited.

WriteHalf

Borrowed write half of a TcpStream, created by split.