Module tokio::net::tcp

TCP utility types

Structs

Incoming
Stream returned by the TcpListener::incoming function representing the stream of sockets received from a listener.

OwnedReadHalf
Owned read half of a TcpStream, created by into_split.

OwnedWriteHalf
Owned write half of a TcpStream, created by into_split.

ReadHalf
Read half of a TcpStream, created by split.

ReuniteError
Error indicating two halves were not from the same socket, and thus could not be reunited.

WriteHalf
Write half of a TcpStream, created by split.