Module tokio::net::unix

Unix domain socket bindings for tokio (only available on unix systems).

Structs

ConnectFuture
Future returned by UnixStream::connect which will resolve to a UnixStream when the stream is connected.

Incoming
Stream of listeners

RecvDgram
A future for receiving datagrams from a Unix datagram socket.

SendDgram
A future for writing a buffer to a Unix datagram socket.

UCred
Credentials of a process

UnixDatagram
An I/O object representing a Unix datagram socket.

UnixListener
A Unix socket which can accept connections from other Unix sockets.

UnixStream
A structure representing a connected Unix socket.