tokio::net::unix

Module pipe

Unix pipe types.

Structs

OpenOptions

Options and flags which can be used to configure how a FIFO file is opened.

Receiver

Reading end of a Unix pipe.

Sender

Writing end of a Unix pipe.

Enums

PipeEnd 🔒

Functions

get_file_flags 🔒

Gets file descriptor’s flags by fcntl.

has_read_access 🔒

Checks for O_RDONLY or O_RDWR access mode.

has_write_access 🔒

Checks for O_WRONLY or O_RDWR access mode.

is_pipe 🔒

Checks if the file descriptor is a pipe or a FIFO.

pipe

Creates a new anonymous Unix pipe.

set_blocking 🔒

Removes O_NONBLOCK from fd’s flags.

set_nonblocking 🔒

Sets file descriptor’s flags with O_NONBLOCK by fcntl.