Trait tokio::park::Park

pub(crate) trait Park {
    type Unpark: Unpark;
    type Error;
    fn unpark(&self) -> Self::Unpark;

    fn park(&mut self) -> Result<(), Self::Error>;

    fn park_timeout(&mut self, duration: Duration) -> Result<(), Self::Error>;
}

Block the current thread.

Associated Types

type Unpark: Unpark

Unpark handle type for the Park implementation.

type Error

Error returned by park

Required methods

fn unpark(&self) -> Self::Unpark

Gets a new Unpark handle associated with this Park instance.

fn park(&mut self) -> Result<(), Self::Error>

Blocks the current thread unless or until the token is available.

A call to park does not guarantee that the thread will remain blocked forever, and callers should be prepared for this possibility. This function may wakeup spuriously for any reason.

Panics

This function should not panic, but ultimately, panics are left as an implementation detail. Refer to the documentation for the specific Park implementation

fn park_timeout(&mut self, duration: Duration) -> Result<(), Self::Error>

Parks the current thread for at most duration.

This function is the same as park but allows specifying a maximum time to block the thread for.

Same as park, there is no guarantee that the thread will remain blocked for any amount of time. Spurious wakeups are permitted for any reason.

Panics

This function should not panic, but ultimately, panics are left as an implementation detail. Refer to the documentation for the specific Park implementation