Module tokio::prelude

A "prelude" for users of the tokio crate.

This prelude is similar to the standard library's prelude in that you'll almost always want to import its entire contents, but unlike the standard library's prelude you'll have to do so manually:

use tokio::prelude::*;

The prelude may grow over time as additional items see ubiquitous use.

Re-exports

pub use crate::io;

pub use crate::io::AsyncBufReadExt as _;

pub use crate::io::AsyncReadExt as _;

pub use crate::io::AsyncSeekExt as _;

pub use crate::io::AsyncWriteExt as _;

Traits

AsyncBufRead
Reads bytes asynchronously.

AsyncRead
Reads bytes from a source.

AsyncWrite
Writes bytes asynchronously.