Module tokio::reactor

The core reactor driving all I/O.

This module contains the Core reactor type which is the event loop for all I/O happening in tokio. This core reactor (or event loop) is used to drive I/O resources.

The Handle and Remote structs are refences to the event loop, created by the handle and remote respectively, and are used to construct I/O objects. Remote is sendable, while Handle is not.

Lastly PollEvented can be used to construct I/O objects that interact with the event loop, e.g. TcpStream in the net module.

Structs

Core
The core reactor, or event loop.

Handle
A handle to an event loop, used to construct I/O objects, send messages, and otherwise interact indirectly with the event loop itself.

PollEvented
A concrete implementation of a stream of readiness notifications for I/O objects that originates from an event loop.

Remote
A remote handle to an event loop, for more information see Handle.