tokio::runtime::blocking::schedule

Struct BlockingSchedule

pub(crate) struct BlockingSchedule {
    hooks: TaskHarnessScheduleHooks,
}

task::Schedule implementation that does nothing (except some bookkeeping in test-util builds). This is unique to the blocking scheduler as tasks scheduled are not really futures but blocking operations.

We avoid storing the task by forgetting it in bind and re-materializing it in release.

Fields
hooks: TaskHarnessScheduleHooks
Implementations

impl BlockingSchedule

pub(crate) fn new(handle: &Handle) -> Self