Type Alias tokio::runtime::driver::SignalDriver

type SignalDriver = Driver;

Aliased Type

struct SignalDriver {
    io: Driver,
    receiver: UnixStream,
    inner: Arc<()>,
}

Fields
io: Driver

Thread parker. The Driver park implementation delegates to this.

receiver: UnixStream

A pipe for receiving wake events from the signal handler

inner: Arc<()>

Shared state. The driver keeps a strong ref and the handle keeps a weak ref. The weak ref is used to check if the driver is still active before trying to register a signal handler.