tokio::runtime::scheduler

Module current_thread

Structs

Context 🔒

Thread-local context.

Core 🔒

Data required for executing the scheduler. The struct is passed around to a function that will perform the scheduling work and acts as a capability token.

CoreGuard 🔒

Used to ensure we always place the Core value back into its slot in CurrentThread, even if the future panics.

CurrentThread 🔒

Executes tasks on the current thread

Handle 🔒

Handle to the current thread scheduler

Shared 🔒

Scheduler state shared between threads.

Constants

DEFAULT_GLOBAL_QUEUE_INTERVAL 🔒

Used if none is specified. This is a temporary constant and will be removed as we unify tuning logic between the multi-thread and current-thread schedulers.

INITIAL_CAPACITY 🔒

Initial queue capacity.

Functions

shutdown2 🔒

Type Aliases

Notified 🔒