tokio::runtime::scheduler::multi_thread

Module queue

Run-queue structures to support a work-stealing scheduler

Structs

Inner 🔒

Local 🔒

Producer handle. May only be used from a single thread.

Steal 🔒

Consumer handle. May be used from many threads.

Constants

LOCAL_QUEUE_CAPACITY 🔒

MASK 🔒

Functions

len 🔒

Calculate the length of the queue using the head and tail. The head can be the steal or real head.

local 🔒

Create a new local run-queue

make_fixed_size 🔒

pack 🔒

Join the two head values

unpack 🔒

Split the head value into the real head and the index a stealer is working on.

Type Aliases

AtomicUnsignedLong 🔒

AtomicUnsignedShort 🔒

UnsignedLong 🔒

UnsignedShort 🔒