Struct tokio::runtime::Handle

pub struct Handle { /* fields omitted */ }

Handle to the runtime

Methods

impl Handle

pub fn enter<F, R>(&self, f: F) -> R where
F: FnOnce() -> R,

Enter the runtime context

pub fn current() -> Self

Returns a Handle view over the currently running Runtime

Panic

A Runtime must have been started or this will panic

Examples

This allows for the current handle to be gotten when running in a #

use tokio::runtime::Handle;

let handle = Handle::current();
handle.spawn(async {
    println!("now running in the existing Runtime");
})

pub fn try_current() -> Result<Self, TryCurrentError>

Returns a Handle view over the currently running Runtime

Returns an error if no Runtime has been started

Contrary to current, this never panics

impl Handle

pub fn spawn<F>(&self, future: F) -> JoinHandle<F::Output> where
F: Future + Send + 'static,
F::Output: Send + 'static,

Spawns a future onto the Tokio runtime.

This spawns the given future onto the runtime's executor, usually a thread pool. The thread pool is then responsible for polling the future until it completes.

See module level documentation for more details.

Examples

use tokio::runtime::Runtime;

// Create the runtime
let rt = Runtime::new().unwrap();
let handle = rt.handle();

// Spawn a future onto the runtime
handle.spawn(async {
    println!("now running on a worker thread");
});

Panics

This function panics if the spawn fails. Failure occurs if the executor is currently at capacity and is unable to spawn a new future.