Module tokio::runtime::task

Re-exports

pub use self::error::JoinError;

pub use self::join::JoinHandle;

Modules

core

Core task module.

error

harness

join

raw

stack

state

waker

Structs

Notified

A task was notified

Task

An owned handle to the task, tracked by ref count

Traits

Schedule

Functions

joinable

Create a new task with an associated join handle

Type Definitions

Result

Task result sent back