Module tokio::runtime::task::core

Core task module.

Safety

The functions in this module are private to the task module. All of them should be considered unsafe to use, but are not marked as such since it would be too noisy.

Make sure to consult the relevant safety section of each function before use.

Structs

Cell
The task cell. Contains the components of the task.

Core
The core of the task.

Header
Crate public as this is also needed by the pool.

Trailer
Cold data is stored after the future.

Enums

Stage
Either the future or the output.