Module tokio::runtime::task::core

Core task module.

Safety

The functions in this module are private to the task module. All of them should be considered unsafe to use, but are not marked as such since it would be too noisy.

Make sure to consult the relevant safety section of each function before use.

Structs

Cell 🔒

The task cell. Contains the components of the task.

Core 🔒

The core of the task.

CoreStage 🔒

Header 🔒

Crate public as this is also needed by the pool.

TaskIdGuard 🔒

Set and clear the task id in the context when the future is executed or dropped, or when the output produced by the future is dropped.

Trailer 🔒

Cold data is stored after the future. Data is considered cold if it is only used during creation or shutdown of the task.

Enums

Stage 🔒

Either the future or the output.