Struct tokio::runtime::task::harness::Harness

pub(super) struct Harness<T: Future, S: 'static> {
    cell: NonNull<Cell<T, S>>,
}

Typed raw task handle

Fields
cell: NonNull<Cell<T, S>>
Implementations

impl<T, S> Harness<T, S> where
T: Future,
S: 'static,

pub(super) unsafe fn from_raw(
ptr: NonNull<Header>
) -> Harness<T, S>

fn header(&self) -> &Header

fn trailer(&self) -> &Trailer

fn core(&self) -> &Core<T, S>

impl<T, S> Harness<T, S> where
T: Future,
S: Schedule,

pub(super) fn poll(
self
)

Polls the inner future.

All necessary state checks and transitions are performed.

Panics raised while polling the future are handled.

pub(super) fn dealloc(
self
)

pub(super) fn try_read_output(
self,
dst: &mut Poll<Result<T::Output, JoinError>>,
waker: &Waker
)

Read the task output into dst.

fn set_join_waker(
&self,
waker: Waker,
snapshot: Snapshot
) -> Result<Snapshot, Snapshot>

pub(super) fn drop_join_handle_slow(
self
)

pub(super) fn wake_by_val(
self
)

pub(super) fn wake_by_ref(
&self
)

pub(super) fn drop_reference(
self
)

pub(super) fn shutdown(
self
)

Forcibly shutdown the task

Attempt to transition to Running in order to forcibly shutdown the task. If the task is currently running or in a state of completion, then there is nothing further to do. When the task completes running, it will notice the CANCELLED bit and finalize the task.

fn cancel_task(self)

fn complete(
self,
output: Result<T::Output, JoinError>,
is_join_interested: bool
)

fn transition_to_complete(&mut self)

Transitions the task's lifecycle to Complete. Notifies the JoinHandle if it still has interest in the completion.

fn wake_join(&self)

fn to_task(&self) -> Task<S>