Module tokio::runtime::task::state

Structs

Snapshot

Current state value.

State

Enums

TransitionToIdle

TransitionToNotifiedByRef

TransitionToNotifiedByVal

TransitionToRunning

Constants

CANCELLED

The task has been forcibly cancelled.

COMPLETE

The task is complete.

INITIAL_STATE

State a task is initialized with.

JOIN_INTEREST

The join handle is still around.

JOIN_WAKER

A join handle waker has been set.

LIFECYCLE_MASK

Extracts the task’s lifecycle value from the state.

NOTIFIED

Flag tracking if the task has been pushed into a run queue.

REF_COUNT_MASK

Bits used by the ref count portion of the state.

REF_COUNT_SHIFT

Number of positions to shift the ref count.

REF_ONE

One ref count.

RUNNING

The task is currently being run.

STATE_MASK

All bits.

Type Definitions

UpdateResult