Trait tokio::runtime::task::Schedule

pub(crate) trait Schedule: Sync + Sized + 'static {
    fn bind(task: Task<Self>) -> Self;

    fn release(&self, task: &Task<Self>) -> Option<Task<Self>>;

    fn schedule(&self, task: Notified<Self>);

    fn yield_now(&self, task: Notified<Self>) { ... }
}

Required methods

fn bind(task: Task<Self>) -> Self

Bind a task to the executor.

Guaranteed to be called from the thread that called poll on the task. The returned Schedule instance is associated with the task and is used as &self in the other methods on this trait.

fn release(&self, task: &Task<Self>) -> Option<Task<Self>>

The task has completed work and is ready to be released. The scheduler is free to drop it whenever.

If the scheduler can immediately release the task, it should return it as part of the function. This enables the task module to batch the ref-dec with other options.

fn schedule(&self, task: Notified<Self>)

Schedule the task

Loading content...
Provided methods

fn yield_now(&self, task: Notified<Self>)

Schedule the task to run in the near future, yielding the thread to other tasks.

Loading content...