Module tokio::runtime::thread_pool

Threadpool

Modules

atomic_cell

idle

Coordinates idling workers

worker

A scheduler is initialized with a fixed number of workers. Each worker is driven by a thread. Each worker has a “core” which contains data such as the run queue and other state. When block_in_place is called, the worker’s “core” is handed off to a new thread allowing the scheduler to continue to make progress while the originating thread blocks.

Structs

Spawner

Submits futures to the associated thread pool for execution.

ThreadPool

Work-stealing based thread pool for executing futures.