Struct tokio::runtime::thread_pool::worker::Shared

pub(super) struct Shared {
    remotes: Box<[Remote]>,
    inject: Inject<Arc<Shared>>,
    idle: Idle,
    owned: OwnedTasks<Arc<Shared>>,
    shutdown_cores: Mutex<Vec<Box<Core>>>,
    before_park: Option<Arc<dyn Fn() + Send + Sync>>,
    after_unpark: Option<Arc<dyn Fn() + Send + Sync>>,
    stats: RuntimeStats,
}

State shared across all workers

Fields
remotes: Box<[Remote]>

Per-worker remote state. All other workers have access to this and is how they communicate between each other.

inject: Inject<Arc<Shared>>

Submits work to the scheduler while not currently on a worker thread.

idle: Idle

Coordinates idle workers

owned: OwnedTasks<Arc<Shared>>

Collection of all active tasks spawned onto this executor.

shutdown_cores: Mutex<Vec<Box<Core>>>

Cores that have observed the shutdown signal

The core is not placed back in the worker to avoid it from being stolen by a thread that was spawned as part of block_in_place.

before_park: Option<Arc<dyn Fn() + Send + Sync>>

Callback for a worker parking itself

after_unpark: Option<Arc<dyn Fn() + Send + Sync>>

Callback for a worker unparking itself

stats: RuntimeStats

Collects stats from the runtime.

Implementations

impl Shared

pub(super) fn bind_new_task<T>(
me: &Arc<Self>,
future: T
) -> JoinHandle<T::Output> where
T: Future + Send + 'static,
T::Output: Send + 'static,

pub(crate) fn stats(&self) -> &RuntimeStats

pub(super) fn schedule(&self, task: Notified<Arc<Shared>>, is_yield: bool)

fn schedule_local(
&self,
core: &mut Core,
task: Notified<Arc<Shared>>,
is_yield: bool
)

pub(super) fn close(&self)

fn notify_parked(&self)

fn notify_all(&self)

fn notify_if_work_pending(&self)

fn transition_worker_from_searching(&self)

fn shutdown(&self, core: Box<Core>)

Signals that a worker has observed the shutdown signal and has replaced its core back into its handle.

If all workers have reached this point, the final cleanup is performed.

fn ptr_eq(&self, other: &Shared) -> bool