Module tokio::signal::unix

Unix-specific types for signal handling.

This module is only defined on Unix platforms and contains the primary Signal type for receiving notifications of signals.

Structs

Signal

An listener for receiving a particular type of OS signal.

SignalKind

Represents the specific kind of signal to listen for.

Functions

signal

Creates a new listener which will receive notifications when the current process receives the specified signal kind.