Struct tokio::signal::unix::SignalKind

pub struct SignalKind(_);

Represents the specific kind of signal to listen for.

Implementations

impl SignalKind

pub fn from_raw(signum: c_int) -> Self

Allows for listening to any valid OS signal.

For example, this can be used for listening for platform-specific signals.

// let signum = libc::OS_SPECIFIC_SIGNAL;
let kind = SignalKind::from_raw(signum);

pub fn alarm() -> Self

Represents the SIGALRM signal.

On Unix systems this signal is sent when a real-time timer has expired. By default, the process is terminated by this signal.

pub fn child() -> Self

Represents the SIGCHLD signal.

On Unix systems this signal is sent when the status of a child process has changed. By default, this signal is ignored.

pub fn hangup() -> Self

Represents the SIGHUP signal.

On Unix systems this signal is sent when the terminal is disconnected. By default, the process is terminated by this signal.

pub fn interrupt() -> Self

Represents the SIGINT signal.

On Unix systems this signal is sent to interrupt a program. By default, the process is terminated by this signal.

pub fn io() -> Self

Represents the SIGIO signal.

On Unix systems this signal is sent when I/O operations are possible on some file descriptor. By default, this signal is ignored.

pub fn pipe() -> Self

Represents the SIGPIPE signal.

On Unix systems this signal is sent when the process attempts to write to a pipe which has no reader. By default, the process is terminated by this signal.

pub fn quit() -> Self

Represents the SIGQUIT signal.

On Unix systems this signal is sent to issue a shutdown of the process, after which the OS will dump the process core. By default, the process is terminated by this signal.

pub fn terminate() -> Self

Represents the SIGTERM signal.

On Unix systems this signal is sent to issue a shutdown of the process. By default, the process is terminated by this signal.

pub fn user_defined1() -> Self

Represents the SIGUSR1 signal.

On Unix systems this is a user defined signal. By default, the process is terminated by this signal.

pub fn user_defined2() -> Self

Represents the SIGUSR2 signal.

On Unix systems this is a user defined signal. By default, the process is terminated by this signal.

pub fn window_change() -> Self

Represents the SIGWINCH signal.

On Unix systems this signal is sent when the terminal window is resized. By default, this signal is ignored.