Module tokio::stream

Stream utilities for Tokio.

A Stream is an asynchronous sequence of values. It can be thought of as an asynchronous version of the standard library’s Iterator trait.

This module provides helpers to work with them.

Structs

Empty

Stream for the empty function.

Iter

Stream for the iter function.

Once

Stream for the once function.

Pending

Stream for the pending function.

StreamMap

Combine many streams into one, indexing each source stream with a unique key.

Traits

FromStream

Convert from a Stream.

Stream

A stream of values produced asynchronously.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.

Functions

empty

Creates a stream that yields nothing.

iter

Converts an Iterator into a Stream which is always ready to yield the next value.

once

Creates a stream that emits an element exactly once.

pending

Creates a stream that is never ready