Trait tokio::stream::StreamExt

pub trait StreamExt: Stream {
    fn throttle(self, duration: Duration) -> Throttle<Self>
    where
        Self: Sized,
    { ... }

    fn timeout(self, timeout: Duration) -> Timeout<Self>
    where
        Self: Sized,
    { ... }
}

An extension trait for Stream that provides a variety of convenient combinator functions.

Currently, there are only timeout and throttle functions, but this will increase over time.

Users are not expected to implement this trait. All types that implement Stream already implement StreamExt.

This trait can be imported directly or via the Tokio prelude: use tokio::prelude::*.

Provided methods

fn throttle(self, duration: Duration) -> Throttle<Self> where
Self: Sized,

Throttle down the stream by enforcing a fixed delay between items.

Errors are also delayed.

Important traits for Timeout<T>
impl<T> Future for Timeout<T> where
T: Future, type Output = Result<<T as Future>::Output, Elapsed>;

fn timeout(self, timeout: Duration) -> Timeout<Self> where
Self: Sized,

Creates a new stream which allows self until timeout.

This combinator creates a new stream which wraps the receiving stream with a timeout. For each item, the returned stream is allowed to execute until it completes or timeout has elapsed, whichever happens first.

If an item completes before timeout then the stream will yield with that item. Otherwise the stream will yield to an error.

Examples

use tokio::prelude::*;

use std::time::Duration;

let mut stream = slow_stream()
    .timeout(Duration::from_secs(1));

while let Some(value) = stream.next().await {
    println!("value = {:?}", value);
}

Loading content...