Module tokio::sync::mpsc::block

Structs

Block
A block in a linked list.

Values

Enums

Read

Constants

BLOCK_MASK
Masks an index to get the block identifier

READY_MASK
Mask covering all bits used to track slot readiness.

RELEASED
Flag tracking that a block has gone through the sender's release routine.

SLOT_MASK
Masks an index to get the value offset in a block.

TX_CLOSED
Flag tracking all senders dropped.

Functions

is_ready
Returns true if the specificed slot has a value ready to be consumed.

is_tx_closed
Returns true if the closed flag has been set.

offset
Returns the offset into the block referenced by slot_index.

start_index
Returns the index of the first slot in the block referenced by slot_index.