Struct tokio::sync::OwnedRwLockReadGuard

pub struct OwnedRwLockReadGuard<T: ?Sized, U: ?Sized = T> { /* private fields */ }

Owned RAII structure used to release the shared read access of a lock when dropped.

This structure is created by the read_owned method on RwLock.

Implementations

impl<T: ?Sized, U: ?Sized> OwnedRwLockReadGuard<T, U>

pub fn map<F, V: ?Sized>(this: Self, f: F) -> OwnedRwLockReadGuard<T, V>where
F: FnOnce(&U) -> &V,

Makes a new OwnedRwLockReadGuard for a component of the locked data. This operation cannot fail as the OwnedRwLockReadGuard passed in already locked the data.

This is an associated function that needs to be used as OwnedRwLockReadGuard::map(...). A method would interfere with methods of the same name on the contents of the locked data.

Examples

use std::sync::Arc;
use tokio::sync::{RwLock, OwnedRwLockReadGuard};

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
struct Foo(u32);

let lock = Arc::new(RwLock::new(Foo(1)));

let guard = lock.read_owned().await;
let guard = OwnedRwLockReadGuard::map(guard, |f| &f.0);

assert_eq!(1, *guard);

pub fn try_map<F, V: ?Sized>(
this: Self,
f: F
) -> Result<OwnedRwLockReadGuard<T, V>, Self>where
F: FnOnce(&U) -> Option<&V>,

Attempts to make a new OwnedRwLockReadGuard for a component of the locked data. The original guard is returned if the closure returns None.

This operation cannot fail as the OwnedRwLockReadGuard passed in already locked the data.

This is an associated function that needs to be used as OwnedRwLockReadGuard::try_map(..). A method would interfere with methods of the same name on the contents of the locked data.

Examples

use std::sync::Arc;
use tokio::sync::{RwLock, OwnedRwLockReadGuard};

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
struct Foo(u32);

let lock = Arc::new(RwLock::new(Foo(1)));

let guard = lock.read_owned().await;
let guard = OwnedRwLockReadGuard::try_map(guard, |f| Some(&f.0)).expect("should not fail");

assert_eq!(1, *guard);