Struct tokio::sync::RwLock

pub struct RwLock<T: ?Sized> { /* fields omitted */ }

An asynchronous reader-writer lock

This type of lock allows a number of readers or at most one writer at any point in time. The write portion of this lock typically allows modification of the underlying data (exclusive access) and the read portion of this lock typically allows for read-only access (shared access).

In comparison, a Mutex does not distinguish between readers or writers that acquire the lock, therefore causing any tasks waiting for the lock to become available to yield. An RwLock will allow any number of readers to acquire the lock as long as a writer is not holding the lock.

The priority policy of Tokio's read-write lock is fair (or write-preferring), in order to ensure that readers cannot starve writers. Fairness is ensured using a first-in, first-out queue for the tasks awaiting the lock; if a task that wishes to acquire the write lock is at the head of the queue, read locks will not be given out until the write lock has been released. This is in contrast to the Rust standard library's std::sync::RwLock, where the priority policy is dependent on the operating system's implementation.

The type parameter T represents the data that this lock protects. It is required that T satisfies Send to be shared across threads. The RAII guards returned from the locking methods implement Deref (and DerefMut for the write methods) to allow access to the content of the lock.

Examples

use tokio::sync::RwLock;

#[tokio::main]
async fn main() {
    let lock = RwLock::new(5);

    // many reader locks can be held at once
    {
        let r1 = lock.read().await;
        let r2 = lock.read().await;
        assert_eq!(*r1, 5);
        assert_eq!(*r2, 5);
    } // read locks are dropped at this point

    // only one write lock may be held, however
    {
        let mut w = lock.write().await;
        *w += 1;
        assert_eq!(*w, 6);
    } // write lock is dropped here
}

Implementations

impl<T: ?Sized> RwLock<T>

pub fn new(value: T) -> RwLock<T> where
T: Sized,

Creates a new instance of an RwLock<T> which is unlocked.

Examples

use tokio::sync::RwLock;

let lock = RwLock::new(5);

pub async fn read<'_, '_>(&'_ self) -> RwLockReadGuard<'_, T>

Locks this rwlock with shared read access, causing the current task to yield until the lock has been acquired.

The calling task will yield until there are no more writers which hold the lock. There may be other readers currently inside the lock when this method returns.

Examples

use std::sync::Arc;
use tokio::sync::RwLock;

#[tokio::main]
async fn main() {
    let lock = Arc::new(RwLock::new(1));
    let c_lock = lock.clone();

    let n = lock.read().await;
    assert_eq!(*n, 1);

    tokio::spawn(async move {
        // While main has an active read lock, we acquire one too.
        let r = c_lock.read().await;
        assert_eq!(*r, 1);
    }).await.expect("The spawned task has paniced");

    // Drop the guard after the spawned task finishes.
    drop(n);
}

pub async fn write<'_, '_>(&'_ self) -> RwLockWriteGuard<'_, T>

Locks this rwlock with exclusive write access, causing the current task to yield until the lock has been acquired.

This function will not return while other writers or other readers currently have access to the lock.

Returns an RAII guard which will drop the write access of this rwlock when dropped.

Examples

use tokio::sync::RwLock;

#[tokio::main]
async fn main() {
  let lock = RwLock::new(1);

  let mut n = lock.write().await;
  *n = 2;
}

pub fn into_inner(self) -> T where
T: Sized,

Consumes the lock, returning the underlying data.