Struct tokio::sync::watch::Receiver

pub struct Receiver<T> { /* fields omitted */ }

Receives values from the associated Sender.

Instances are created by the channel function.

Implementations

impl<T> Receiver<T>

pub fn borrow(&self) -> Ref<T>

Returns a reference to the most recently sent value

Outstanding borrows hold a read lock. This means that long lived borrows could cause the send half to block. It is recommended to keep the borrow as short lived as possible.

Examples

use tokio::sync::watch;

let (_, rx) = watch::channel("hello");
assert_eq!(*rx.borrow(), "hello");

impl<T: Clone> Receiver<T>

pub async fn recv<'_>(&'_ mut self) -> Option<T>

Attempts to clone the latest value sent via the channel.

If this is the first time the function is called on a Receiver instance, then the function completes immediately with the current value held by the channel. On the next call, the function waits until a new value is sent in the channel.

None is returned if the Sender half is dropped.

Examples

use tokio::sync::watch;

#[tokio::main]
async fn main() {
    let (tx, mut rx) = watch::channel("hello");

    let v = rx.recv().await.unwrap();
    assert_eq!(v, "hello");

    tokio::spawn(async move {
        tx.broadcast("goodbye").unwrap();
    });

    // Waits for the new task to spawn and send the value.
    let v = rx.recv().await.unwrap();
    assert_eq!(v, "goodbye");

    let v = rx.recv().await;
    assert!(v.is_none());
}