Function tokio::task::yield_now

#[must_use = "yield_now does nothing unless polled/`await`-ed"]pub async fn yield_now()

Yields execution back to the Tokio runtime.

A task yields by awaiting on yield_now(), and may resume when that future completes (with no output.) The current task will be re-added as a pending task at the back of the pending queue. Any other pending tasks will be scheduled. No other waking is required for the task to continue.

See also the usage example in the task module.