Struct tokio::task::JoinSet

pub struct JoinSet<T> { /* private fields */ }

A collection of tasks spawned on a Tokio runtime.

A JoinSet can be used to await the completion of some or all of the tasks in the set. The set is not ordered, and the tasks will be returned in the order they complete.

All of the tasks must have the same return type T.

When the JoinSet is dropped, all tasks in the JoinSet are immediately aborted.

Examples

Spawn multiple tasks and wait for them.

use tokio::task::JoinSet;

#[tokio::main]
async fn main() {
    let mut set = JoinSet::new();

    for i in 0..10 {
        set.spawn(async move { i });
    }

    let mut seen = [false; 10];
    while let Some(res) = set.join_next().await {
        let idx = res.unwrap();
        seen[idx] = true;
    }

    for i in 0..10 {
        assert!(seen[i]);
    }
}

Implementations

impl<T> JoinSet<T>

pub fn new() -> Self

Create a new JoinSet.

pub fn len(&self) -> usize

Returns the number of tasks currently in the JoinSet.

pub fn is_empty(&self) -> bool

Returns whether the JoinSet is empty.

impl<T: 'static> JoinSet<T>

pub fn spawn<F>(&mut self, task: F) -> AbortHandle
where F: Future<Output = T> + Send + 'static, T: Send,

Spawn the provided task on the JoinSet, returning an AbortHandle that can be used to remotely cancel the task.

The provided future will start running in the background immediately when this method is called, even if you don’t await anything on this JoinSet.

Panics

This method panics if called outside of a Tokio runtime.

pub fn spawn_on<F>(&mut self, task: F, handle: &Handle) -> AbortHandle
where F: Future<Output = T> + Send + 'static, T: Send,

Spawn the provided task on the provided runtime and store it in this JoinSet returning an AbortHandle that can be used to remotely cancel the task.

The provided future will start running in the background immediately when this method is called, even if you don’t await anything on this JoinSet.

pub fn spawn_local<F>(&mut self, task: F) -> AbortHandle
where F: Future<Output = T> + 'static,

Spawn the provided task on the current LocalSet and store it in this JoinSet, returning an AbortHandle that can be used to remotely cancel the task.

The provided future will start running in the background immediately when this method is called, even if you don’t await anything on this JoinSet.

Panics

This method panics if it is called outside of a LocalSet.

pub fn spawn_local_on<F>( &mut self, task: F, local_set: &LocalSet ) -> AbortHandle
where F: Future<Output = T> + 'static,

Spawn the provided task on the provided LocalSet and store it in this JoinSet, returning an AbortHandle that can be used to remotely cancel the task.

Unlike the spawn_local method, this method may be used to spawn local tasks on a LocalSet that is not currently running. The provided future will start running whenever the LocalSet is next started.

pub fn spawn_blocking<F>(&mut self, f: F) -> AbortHandle
where F: FnOnce() -> T + Send + 'static, T: Send,

Spawn the blocking code on the blocking threadpool and store it in this JoinSet, returning an AbortHandle that can be used to remotely cancel the task.

Examples

Spawn multiple blocking tasks and wait for them.

use tokio::task::JoinSet;

#[tokio::main]
async fn main() {
    let mut set = JoinSet::new();

    for i in 0..10 {
        set.spawn_blocking(move || { i });
    }

    let mut seen = [false; 10];
    while let Some(res) = set.join_next().await {
        let idx = res.unwrap();
        seen[idx] = true;
    }

    for i in 0..10 {
        assert!(seen[i]);
    }
}

Panics

This method panics if called outside of a Tokio runtime.

pub fn spawn_blocking_on<F>(&mut self, f: F, handle: &Handle) -> AbortHandle
where F: FnOnce() -> T + Send + 'static, T: Send,

Spawn the blocking code on the blocking threadpool of the provided runtime and store it in this JoinSet, returning an AbortHandle that can be used to remotely cancel the task.

pub async fn join_next(&mut self) -> Option<Result<T, JoinError>>

Waits until one of the tasks in the set completes and returns its output.

Returns None if the set is empty.

Cancel Safety

This method is cancel safe. If join_next is used as the event in a tokio::select! statement and some other branch completes first, it is guaranteed that no tasks were removed from this JoinSet.

pub fn try_join_next(&mut self) -> Option<Result<T, JoinError>>

Tries to join one of the tasks in the set that has completed and return its output.

Returns None if the set is empty.

pub async fn shutdown(&mut self)

Aborts all tasks and waits for them to finish shutting down.

Calling this method is equivalent to calling abort_all and then calling join_next in a loop until it returns None.

This method ignores any panics in the tasks shutting down. When this call returns, the JoinSet will be empty.

pub fn abort_all(&mut self)

Aborts all tasks on this JoinSet.

This does not remove the tasks from the JoinSet. To wait for the tasks to complete cancellation, you should call join_next in a loop until the JoinSet is empty.

pub fn detach_all(&mut self)

Removes all tasks from this JoinSet without aborting them.

The tasks removed by this call will continue to run in the background even if the JoinSet is dropped.

pub fn poll_join_next( &mut self, cx: &mut Context<'_> ) -> Poll<Option<Result<T, JoinError>>>

Polls for one of the tasks in the set to complete.

If this returns Poll::Ready(Some(_)), then the task that completed is removed from the set.

When the method returns Poll::Pending, the Waker in the provided Context is scheduled to receive a wakeup when a task in the JoinSet completes. Note that on multiple calls to poll_join_next, only the Waker from the Context passed to the most recent call is scheduled to receive a wakeup.

Returns

This function returns:

Poll::Pending if the JoinSet is not empty but there is no task whose output is available right now.

Poll::Ready(Some(Ok(value))) if one of the tasks in this JoinSet has completed. The value is the return value of one of the tasks that completed.

Poll::Ready(Some(Err(err))) if one of the tasks in this JoinSet has panicked or been aborted. The err is the JoinError from the panicked/aborted task.

Poll::Ready(None) if the JoinSet is empty.

Note that this method may return Poll::Pending even if one of the tasks has completed. This can happen if the coop budget is reached.