Module tokio::time::interval

Structs

Interval

Interval returned by interval and interval_at.

Enums

MissedTickBehavior

Defines the behavior of an Interval when it misses a tick.

Functions

internal_interval_at 🔒

interval

Creates new Interval that yields with interval of period. The first tick completes immediately. The default MissedTickBehavior is Burst, but this can be configured by calling set_missed_tick_behavior.

interval_at

Creates new Interval that yields with interval of period with the first tick completing at start. The default MissedTickBehavior is Burst, but this can be configured by calling set_missed_tick_behavior.