Struct tokio::time::Instant

pub struct Instant { /* fields omitted */ }

A measurement of the system clock, useful for talking to external entities like the file system or other processes.

Implementations

impl Instant

pub fn now() -> Instant

Returns an instant corresponding to "now".

Examples

use tokio::time::Instant;

let now = Instant::now();

pub fn from_std(std: Instant) -> Instant

Create a tokio::time::Instant from a std::time::Instant.

pub fn into_std(self) -> Instant

Convert the value into a std::time::Instant.

pub fn duration_since(&self, earlier: Instant) -> Duration

Returns the amount of time elapsed from another instant to this one.

Panics

This function will panic if earlier is later than self.

pub fn checked_duration_since(&self, earlier: Instant) -> Option<Duration>

Returns the amount of time elapsed from another instant to this one, or None if that instant is later than this one.

Examples

use tokio::time::{Duration, Instant, delay_for};

#[tokio::main]
async fn main() {
    let now = Instant::now();
    delay_for(Duration::new(1, 0)).await;
    let new_now = Instant::now();
    println!("{:?}", new_now.checked_duration_since(now));
    println!("{:?}", now.checked_duration_since(new_now)); // None
}

pub fn saturating_duration_since(&self, earlier: Instant) -> Duration

Returns the amount of time elapsed from another instant to this one, or zero duration if that instant is earlier than this one.

Examples

use tokio::time::{Duration, Instant, delay_for};

#[tokio::main]
async fn main() {
    let now = Instant::now();
    delay_for(Duration::new(1, 0)).await;
    let new_now = Instant::now();
    println!("{:?}", new_now.saturating_duration_since(now));
    println!("{:?}", now.saturating_duration_since(new_now)); // 0ns
}

pub fn elapsed(&self) -> Duration

Returns the amount of time elapsed since this instant was created.

Panics

This function may panic if the current time is earlier than this instant, which is something that can happen if an Instant is produced synthetically.

Examples

use tokio::time::{Duration, Instant, delay_for};

#[tokio::main]
async fn main() {
    let instant = Instant::now();
    let three_secs = Duration::from_secs(3);
    delay_for(three_secs).await;
    assert!(instant.elapsed() >= three_secs);
}

pub fn checked_add(&self, duration: Duration) -> Option<Instant>

Returns Some(t) where t is the time self + duration if t can be represented as Instant (which means it's inside the bounds of the underlying data structure), None otherwise.

pub fn checked_sub(&self, duration: Duration) -> Option<Instant>

Returns Some(t) where t is the time self - duration if t can be represented as Instant (which means it's inside the bounds of the underlying data structure), None otherwise.