tokio

Module util

Modules

as_ref ğŸ”’

atomic_cell ğŸ”’

bit ğŸ”’

blocking_check ğŸ”’

cacheline ğŸ”’

error ğŸ”’

idle_notified_set ğŸ”’

This module defines an IdleNotifiedSet, which is a collection of elements. Each element is intended to correspond to a task, and the collection will keep track of which tasks have had their waker notified, and which have not.

linked_list ğŸ”’

An intrusive double linked list of data.

markers ğŸ”’

memchr ğŸ”’

Search for a byte in a byte array using libc.

metric_atomics ğŸ”’

ptr_expose ğŸ”’

Utility for helping miri understand our exposed pointers.

rand ğŸ”’

rc_cell ğŸ”’

sharded_list ğŸ”’

sync_wrapper ğŸ”’

This module contains a type that can make Send + !Sync types Sync by disallowing all immutable access to the value.

trace ğŸ”’

try_lock ğŸ”’

typeid ğŸ”’

wake ğŸ”’

wake_list ğŸ”’

Functions

pin_as_deref_mut ğŸ”’

Copy of std::pin::Pin::as_deref_mut.