Module tokio::util

Modules

atomic_cell 🔒

bit 🔒

cacheline 🔒

error 🔒

idle_notified_set 🔒

This module defines an IdleNotifiedSet, which is a collection of elements. Each element is intended to correspond to a task, and the collection will keep track of which tasks have had their waker notified, and which have not.

linked_list 🔒

An intrusive double linked list of data.

markers 🔒

memchr 🔒

Search for a byte in a byte array using libc.

once_cell 🔒

rand 🔒

rc_cell 🔒

sharded_list 🔒

sync_wrapper 🔒

This module contains a type that can make Send + !Sync types Sync by disallowing all immutable access to the value.

trace 🔒

try_lock 🔒

wake 🔒

wake_list 🔒