tokio::util

Module idle_notified_set

This module defines an IdleNotifiedSet, which is a collection of elements. Each element is intended to correspond to a task, and the collection will keep track of which tasks have had their waker notified, and which have not.

Each entry in the set holds some user-specified value. The value’s type is specified using the T parameter. It will usually be a JoinHandle or similar.

Structs

EntryInOneOfTheLists 🔒

A handle to an entry that is guaranteed to be stored in the idle or notified list of its IdleNotifiedSet. This value borrows the IdleNotifiedSet mutably to prevent the entry from being moved to the Neither list, which only the IdleNotifiedSet may do.

IdleNotifiedSet 🔒

This is the main handle to the collection.

ListEntry 🔒

An entry in the list.

ListsInner 🔒

The linked lists hold strong references to the ListEntry items, and the ListEntry items also hold a strong reference back to the Lists object, but the destructor of the IdleNotifiedSet will clear the two lists, so once that object is destroyed, no ref-cycles will remain.

Enums

List 🔒

Which of the two lists in the shared Lists object is this entry stored in?

Functions

move_to_new_list 🔒 ⚠

Safety

Type Aliases

LinkedList 🔒

Lists 🔒