Module tokio::util::linked_list

An intrusive double linked list of data.

The data structure supports tracking pinned nodes. Most of the data structure’s APIs are unsafe as they require the caller to ensure the specified node is actually contained by the list.

Structs

DrainFilter 🔒

GuardedLinkedList 🔒

An intrusive linked list, but instead of keeping pointers to the head and tail nodes, it uses a special guard node linked with those nodes. It means that the list is circular and every pointer of a node from the list is not None, including pointers from the guard node.

LinkedList 🔒

An intrusive linked list.

Pointers 🔒

Previous / next pointers.

PointersInner 🔒

We do not want the compiler to put the noalias attribute on mutable references to this type, so the type has been made !Unpin with a PhantomPinned field.

Traits

Link 🔒

Defines how a type is tracked within a linked list.