tokio::util::linked_list

Trait Link

pub(crate) unsafe trait Link {
    type Handle;
    type Target;

    // Required methods
    fn as_raw(handle: &Self::Handle) -> NonNull<Self::Target>;
    unsafe fn from_raw(ptr: NonNull<Self::Target>) -> Self::Handle;
    unsafe fn pointers(
        target: NonNull<Self::Target>,
    ) -> NonNull<Pointers<Self::Target>>;
}

Defines how a type is tracked within a linked list.

In order to support storing a single type within multiple lists, accessing the list pointers is decoupled from the entry type.

Safety

Implementations must guarantee that Target types are pinned in memory. In other words, when a node is inserted, the value will not be moved as long as it is stored in the list.

Required Associated Types

type Handle

Handle to the list entry.

This is usually a pointer-ish type.

type Target

Node type.

Required Methods

fn as_raw(handle: &Self::Handle) -> NonNull<Self::Target>

Convert the handle to a raw pointer without consuming the handle.

unsafe fn from_raw(ptr: NonNull<Self::Target>) -> Self::Handle

Convert the raw pointer to a handle

unsafe fn pointers( target: NonNull<Self::Target>, ) -> NonNull<Pointers<Self::Target>>

Return the pointers for a node

Safety

The resulting pointer should have the same tag in the stacked-borrows stack as the argument. In particular, the method may not create an intermediate reference in the process of creating the resulting raw pointer.

The target pointer must be valid.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.