tokio::util::sharded_list

Trait ShardedListItem

pub(crate) unsafe trait ShardedListItem: Link {
    // Required method
    unsafe fn get_shard_id(target: NonNull<Self::Target>) -> usize;
}

Determines which linked list an item should be stored in.

Safety

Implementations must guarantee that the id of an item does not change from call to call.

Required Methods

unsafe fn get_shard_id(target: NonNull<Self::Target>) -> usize

Safety

The provided pointer must point at a valid list item.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.