tokio::util

Module wake

Structs

WakerRef 🔒

A Waker that is only valid for a given lifetime.

Traits

Wake 🔒

Simplified waking interface based on Arcs.

Functions

clone_arc_raw 🔒 ⚠

drop_arc_raw 🔒 ⚠

wake_arc_raw 🔒 ⚠

wake_by_ref_arc_raw 🔒 ⚠

waker_ref 🔒

Creates a reference to a Waker from a reference to Arc<impl Wake>.

waker_vtable 🔒